### 2023-07-27 #5695 @editorialbot
[PRE REVIEW]: Ambit - A FEniCS-based cardiovascular physics solver
**Submitting author:** @marchirschvogel (<a href="http://orcid.org/0000-0002-4575-9120">Marc Hirschvogel</a>)
**Repository:** https://github.com/marchirschvogel/ambit
**Branch with paper.md** (empty if default branch):
**Version:** development
