### 2023-07-27 1484876
Updates to partitioned solver
