+++ doc/joss/paper.md
@@ -44,7 +44,7 @@ Avenues for future functionality include cardiac electrophysiology, scalar trans
 
 # Statement of need
 
-Cardiovascular disease entities are the most prevalent ones in the industrialized world [@dimmeler2011,@luepker2011] and a leading cause of death worldwise. Therefore, models that promote a better understanding of cardiac diseases and their progression represent a valuable tool to guide or assist therapy planning, support device dimensioning and design [@hirschvogel2019], or help predict intervention planning [@bonini2022,@taylor2013].
+Cardiovascular disease entities are the most prevalent ones in the industrialized world [@dimmeler2011; @luepker2011] and a leading cause of death worldwise. Therefore, models that promote a better understanding of cardiac diseases and their progression represent a valuable tool to guide or assist therapy planning, support device dimensioning and design [@hirschvogel2019], or help predict intervention planning [@bonini2022; @taylor2013].
 
 Ambit represents a complete open-source code for simulating cardiac mechanics, encompassing advanced structural mechanics of the myocardium, ventricular fluid dynamics, reduced-dimensional blood flow and multi-physics coupling. Therefore, a wide range of mechanical problems can be simulated, and the code structure allows easy and straightforward extensibility (e.g. implementations of new constitutive models). Due to its simple design in terms of clearly organized input files, Ambit is easy to use and hence represents a valuable tool for novices or advanced researchers who want to address cardiovascular mechanics problems.
 
@@ -54,7 +54,7 @@ Figure \autoref{fig:codedesign} represents a basic sketch of the main building b
 
 The (coupled) problem object then is passed to a solver class, which calls the main routine to solve the nonlinear problem. This routine implements a time stepping scheme and a monolithic Newton solver which solves the (coupled multi-physics or single-field) problem and updates all variables simultaneously.
 
-![Basic sketch of Ambit code structure: Problem class, solver class, and main code execution flow. Single-physics problems that can be solved encompass solid mechanics (`solid`), fluid mechanics (`solid`), or 0D models (`flow0d`). Two-physics problems like 3D-0D coupling (`solid_flow0d`, `fluid_flow0d`), as well as fluid in ALE description (`fluid_ale`) are defined by instantiating the respective single-physics problems. Three-physics problems arise for coupling of ALE fluid to 0D models (`fluid_ale_flow0d`) or for fluid-solid interaction (`fsi`), whereas four-physics problems would encompass FSI linked to 0D models (`fsi_flow0d`). Note that the single-physics problem `ale` just mimics a dummy linear elastic solid and would be irrelevant as a standalone problem.\label{fig:codedesign}](code_design.pdf){ width=80% }
+![Basic sketch of Ambit code structure: Problem class, solver class, and main code execution flow. Single-physics problems that can be solved encompass solid mechanics (`solid`), fluid mechanics (`fluid`), or 0D models (`flow0d`). Two-physics problems like 3D-0D coupling (`solid_flow0d`, `fluid_flow0d`), as well as fluid in ALE description (`fluid_ale`) are defined by instantiating the respective single-physics problems. Three-physics problems arise for coupling of ALE fluid to 0D models (`fluid_ale_flow0d`) or for fluid-solid interaction (`fsi`), whereas four-physics problems would encompass FSI linked to 0D models (`fsi_flow0d`). Note that the single-physics problem `ale` just mimics a dummy linear elastic solid and would be irrelevant as a standalone problem.\label{fig:codedesign}](code_design.pdf){ width=80% }
 
 <!--# Citations
 
