### 2023-07-28 5dfdaf2
Fix leak in partitioned solve; add partitioned FrSI testcase
