### 2023-07-29 6558c42
Some code revamp: move ROM to solver base class
