+++ modules/coupling/fluid_ale.py
@@ -53,6 +53,8 @@ class FluidmechanicsAleProblem(problem_base):
         alevariables = {'Fale' : self.pba.ki.F(self.pba.d), 'Fale_old' : self.pba.ki.F(self.pba.d_old), 'w' : self.pba.wel, 'w_old' : self.pba.w_old, 'fluid_on_deformed' : self.fluid_on_deformed}
         self.pbf = FluidmechanicsProblem(io_params, time_params, fem_params_fluid, constitutive_models_fluid, bc_dict_fluid, time_curves, io, mor_params=mor_params, comm=self.comm, alevar=alevariables)
 
+        self.pbrom = self.pbf # ROM problem can only be fluid
+
         # modify results to write...
         self.pbf.results_to_write = io_params['results_to_write'][0]
         self.pba.results_to_write = io_params['results_to_write'][1]
@@ -63,7 +65,6 @@ class FluidmechanicsAleProblem(problem_base):
         self.noperiodicref = 1
 
         self.localsolve = False
-        self.have_rom = False
 
         # NOTE: Fluid and ALE function spaces should be of the same type, but are different objects.
         # For some reason, when applying a function from one funtion space as DBC to another function space,
@@ -77,9 +78,6 @@ class FluidmechanicsAleProblem(problem_base):
         self.set_variational_forms()
 
         self.numdof = self.pbf.numdof + self.pba.numdof
-
-        self.have_rom = self.pbf.have_rom
-        if self.have_rom: self.rom = self.pbf.rom
 
         self.sub_solve = False
 
@@ -323,23 +321,23 @@ class FluidmechanicsAleProblem(problem_base):
             w_vec.destroy()
 
 
-    def get_index_sets(self, isoptions={}):
+    def get_index_sets(self, isoptions={}, rom=None):
 
-        if self.have_rom: # currently, ROM can only be on (subset of) first variable
-            vred = PETSc.Vec().createMPI(size=(self.rom.V.getLocalSize()[1],self.rom.V.getSize()[1]), comm=self.comm)
-            self.rom.V.multTranspose(self.pbf.v.vector, vred)
-            vvec = vred
+        if rom is not None: # currently, ROM can only be on (subset of) first variable
+            vvec_or0 = rom.V.getOwnershipRangeColumn()[0]
+            vvec_ls = rom.V.getLocalSize()[1]
         else:
-            vvec = self.pbf.v.vector
+            vvec_or0 = self.pbf.v.vector.getOwnershipRange()[0]
+            vvec_ls = self.pbf.v.vector.getLocalSize()
 
-        offset_v = vvec.getOwnershipRange()[0] + self.pbf.p.vector.getOwnershipRange()[0] + self.pba.d.vector.getOwnershipRange()[0]
-        iset_v = PETSc.IS().createStride(vvec.getLocalSize(), first=offset_v, step=1, comm=self.comm)
+        offset_v = vvec_or0 + self.pbf.p.vector.getOwnershipRange()[0] + self.pba.d.vector.getOwnershipRange()[0]
+        iset_v = PETSc.IS().createStride(vvec_ls, first=offset_v, step=1, comm=self.comm)
 
         if isoptions['rom_to_new']:
-            iset_r = PETSc.IS().createStride(len(self.rom.im_rom_r), first=offset_v, step=1, comm=self.comm) # same offset, since contained in v
+            iset_r = PETSc.IS().createStride(len(rom.im_rom_r), first=offset_v, step=1, comm=self.comm) # same offset, since contained in v
             iset_v = iset_v.difference(iset_r) # subtract
 
-        offset_p = offset_v + vvec.getLocalSize()
+        offset_p = offset_v + vvec_ls
[…]
