+++ testing/runtests.py
@@ -128,7 +128,7 @@ if category=='solid_constraint' or category=='all':
 if category=='frsi' or category=='fluid_ale' or category=='all':
     errs['frsi_artseg_prefile 1'] = subprocess.call(['mpiexec', '-n', '1', 'python3', 'frsi_artseg_prefile.py'])
     errs['frsi_artseg_prefile 4'] = subprocess.call(['mpiexec', '-n', '4', 'python3', 'frsi_artseg_prefile.py'])
-    errs['frsi_artseg_prefile_partitioned 4'] = subprocess.call(['mpiexec', '-n', '4', 'python3', 'frsi_artseg_prefile_partitioned.py'])
+    errs['frsi_artseg_prefile_partitioned_iterative 4'] = subprocess.call(['mpiexec', '-n', '4', 'python3', 'frsi_artseg_prefile_partitioned_iterative.py'])
 
     errs['frsi_artseg_modepartitionunity 1'] = subprocess.call(['mpiexec', '-n', '1', 'python3', 'frsi_artseg_modepartitionunity.py'])
     errs['frsi_artseg_modepartitionunity 2'] = subprocess.call(['mpiexec', '-n', '2', 'python3', 'frsi_artseg_modepartitionunity.py'])

[…]
