### 2023-07-30 ad7b0dd
fix PETSc block extract by sorting constraint and reduced index sets; some change of solver print
