+++ modules/solver/solver_utils.py
@@ -48,7 +48,7 @@ class sol_utils():
         return err
 
 
-    def print_nonlinear_iter(self, it=0, resnorms=None, incnorms=None, header=False, ts=0, te=0, sub=False, ptype=None, prfxlen=1):
+    def print_nonlinear_iter(self, it=0, resnorms=None, incnorms=None, header=False, ts=0, te=0, sub=False, ptype=None):
 
         if ptype=='solid':
             if self.solver.pb[0].incompressible_2field:
@@ -138,31 +138,31 @@ class sol_utils():
             if self.solver.comm.rank == 0:
                 if numres==1:
                     if not sub:
-                        print(('{:<'+str(prfxlen)+'s}{:<6s}{:<25s}{:<3s}{:<7s}').format(' ','it |',eq1,'| ','timings'))
-                        print(('{:<'+str(prfxlen)+'s}{:<6s}{:<13s}{:<12s}{:<3s}{:<10s}{:<7s}').format(' ','#  |','||r_'+v1+'||_2','||Δ'+v1+'||_2','| ','te','ts'))
+                        print(('{:<'+str(self.solver.indlen)+'s}{:<6s}{:<25s}{:<3s}{:<7s}').format(' ','it |',eq1,'| ','timings'))
+                        print(('{:<'+str(self.solver.indlen)+'s}{:<6s}{:<13s}{:<12s}{:<3s}{:<10s}{:<7s}').format(' ','#  |','||r_'+v1+'||_2','||Δ'+v1+'||_2','| ','te','ts'))
                     else:
                         print(' ')
                         print('       ****************** 0D model solve ******************')
-                        print(('{:<'+str(prfxlen)+'s}{:<6s}{:<6s}{:<25s}{:<3s}{:<7s}').format(' ',' ','it |',eq1,'| ','timings'))
-                        print(('{:<'+str(prfxlen)+'s}{:<6s}{:<6s}{:<13s}{:<12s}{:<3s}{:<10s}{:<7s}').format(' ',' ','#  |','||r_'+v1+'||_2','||Δ'+v1+'||_2','| ','te','ts'))
+                        print(('{:<'+str(self.solver.indlen)+'s}{:<6s}{:<6s}{:<25s}{:<3s}{:<7s}').format(' ',' ','it |',eq1,'| ','timings'))
+                        print(('{:<'+str(self.solver.indlen)+'s}{:<6s}{:<6s}{:<13s}{:<12s}{:<3s}{:<10s}{:<7s}').format(' ',' ','#  |','||r_'+v1+'||_2','||Δ'+v1+'||_2','| ','te','ts'))
                 elif numres==2:
-                    print(('{:<'+str(prfxlen)+'s}{:<6s}{:<25s}{:<3s}{:<25s}{:<3s}{:<7s}').format(' ','it |',eq1,'| ',eq2,'| ','timings'))
[…]
