+++ modules/coupling/fluid_flow0d.py
@@ -369,11 +369,23 @@ class FluidmechanicsFlow0DProblem(problem_base):
         offset_v = vvec_or0 + self.pbf.p.vector.getOwnershipRange()[0] + self.lm.getOwnershipRange()[0]
         iset_v = PETSc.IS().createStride(vvec_ls, first=offset_v, step=1, comm=self.comm)
 
+        if isoptions['rom_to_new']:
+            iset_r = PETSc.IS().createStride(len(rom.im_rom_r), first=offset_v, step=1, comm=self.comm) # same offset, since contained in v
+            iset_v = iset_v.difference(iset_r) # subtract
+
         offset_p = offset_v + vvec_ls
         iset_p = PETSc.IS().createStride(self.pbf.p.vector.getLocalSize(), first=offset_p, step=1, comm=self.comm)
 
         offset_s = offset_p + self.pbf.p.vector.getLocalSize()
         iset_s = PETSc.IS().createStride(self.lm.getLocalSize(), first=offset_s, step=1, comm=self.comm)
+
+        if isoptions['rom_to_new']:
+            iset_s = iset_s.expand(iset_r) # add to 0D block
+            iset_s.sort() # should be sorted, otherwise PETSc may struggle to extract block
+            #iset_r = iset_r.expand(iset_s) # add to 0D block
+
+        print(iset_s.array)
+        # sys.exit()
 
         if isoptions['lms_to_p']:
             iset_p = iset_p.expand(iset_s) # add to pressure block
