### 2023-07-31 42b6b4d
Fix in ROM index set for block prec
