+++ modules/coupling/solid_constraint.py
@@ -322,14 +322,16 @@ class SolidmechanicsConstraintProblem(problem_base):
         if self.pbs.incompressible_2field:
             if isoptions['lms_to_p']:
                 iset_p = iset_p.expand(iset_s) # add to pressure block
-                return [iset_u, iset_p]
+                ilist = [iset_u, iset_p]
             elif isoptions['lms_to_v']:
                 iset_u = iset_u.expand(iset_s) # add to displacement block (could be bad...)
-                return [iset_u, iset_p]
+                ilist = [iset_u, iset_p]
             else:
-                return [iset_u, iset_p, iset_s]
+                ilist = [iset_u, iset_p, iset_s]
         else:
-            return [iset_u, iset_s]
+            ilist = [iset_u, iset_s]
+
+        return ilist
 
 
     ### now the base routines for this problem
