### 2023-08-01 #5695 @Kevin-Mattheus-Moerman
@marchirschvogel let me know if besides the above you have any other reviewers you'd like to recommend. You can mention their GitHub handles here but please leave out the `@` symbol so they are not tagged. Thanks.

Some other points that require you attention:
- [ ]  Given that you are the sole contributor, and that there is no history of issues or contributions added by others, can you confirm you have a long term commitment to this open source software project? We essentially want to check if this will be a maintained open source project with a growing community, or if this is simply a "quick GitHub upload of your PhD work", and that you will not maintain it in the future. JOSS only publishes works expected to be maintained/supported long term.
- [ ]  Can you add contributing guidelines to your project? You can add a simple CONTRIBUTING.md file which describes how you'd like others to contribute, and link to it in your README.md (some find this website useful for inspiration when creating such a file: https://contributing.md/example/).
