### 2023-08-01 #5695 @finsberg
> @axel-loewe, @finsberg @mbarzegary, @sdelandtsheer, @albansouche, @chennachaos, @JaroslavHron, @@IgorBaratta, @CZHU20 would you be interested in reviewing this JOSS submission entitled: "Ambit - A FEniCS-based cardiovascular physics solver"? As you know JOSS reviews focuses on the the evaluation of the [the software](https://github.com/marchirschvogel/ambit), and the review of a [short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05695/joss.05695/10.21105.joss.05695.pdf).
>
> Please let me know if you are interested by responding here. Thanks!

@Kevin-Mattheus-Moerman  I am happy to review this submission.
