### 2023-08-03 9b48904
revert to non-greek letter print, since not supported on every machine
