+++ modules/solver/solver_utils.py
@@ -82,7 +82,7 @@ class sol_utils():
                     v1, v2, v3 = 'u', 'p', 's'
                 if self.solver.pb[0].coupling_type == 'monolithic_lagrange':
                     eq1, eq2, eq3 = 'solid momentum', 'solid incompressibility', 'lm constraint'
-                    v1, v2, v3 = 'u', 'p', 'Λ'
+                    v1, v2, v3 = 'u', 'p', 'LM' # using greek symbol print (Λ) is not supported everywhere...
                 numres = 3
             else:
                 if self.solver.pb[0].coupling_type == 'monolithic_direct':
@@ -90,12 +90,12 @@ class sol_utils():
                     v1, v2 = 'u', 's'
                 if self.solver.pb[0].coupling_type == 'monolithic_lagrange':
                     eq1, eq2 = 'solid momentum', 'lm constraint'
-                    v1, v2 = 'u', 'Λ'
+                    v1, v2 = 'u', 'LM' # using greek symbol print (Λ) is not supported everywhere...
                 numres = 2
         elif ptype=='solid_constraint':
             if self.solver.pb[0].incompressible_2field:
                 eq1, eq2, eq3 = 'solid momentum', 'solid incompressibility', 'lm constraint'
-                v1, v2, v3 = 'u', 'p', 'Λ'
+                v1, v2, v3 = 'u', 'p', 'LM' # using greek symbol print (Λ) is not supported everywhere...
                 numres = 3
             else:
                 eq1, eq2 = 'solid momentum', 'lm constraint'
@@ -103,7 +103,7 @@ class sol_utils():
                 numres = 2
         elif ptype=='fluid_flow0d':
             eq1, eq2, eq3 = 'fluid momentum', 'fluid continuity', 'lm constraint'
-            v1, v2, v3 = 'v', 'p', 'Λ'
+            v1, v2, v3 = 'v', 'p', 'LM' # using greek symbol print (Λ) is not supported everywhere...
             numres = 3
         elif ptype=='fluid_ale':
             eq1, eq2, eq3 = 'fluid momentum', 'fluid continuity', 'ALE momentum'
@@ -111,58 +111,59 @@ class sol_utils():
             numres = 3
         elif ptype=='fluid_ale_flow0d':
             eq1, eq2, eq3, eq4 = 'fluid momentum', 'fluid continuity', 'lm constraint', 'ALE momentum'
-            v1, v2, v3, v4 = 'v', 'p', 'Λ', 'd'
+            v1, v2, v3, v4 = 'v', 'p', 'LM', 'd' # using greek symbol print (Λ) is not supported everywhere...
             numres = 4
         elif ptype=='fsi':
             if self.solver.pb[0].incompressible_2field:
[…]
