### 2023-08-07 #5695 @marchirschvogel
> @marchirschvogel let me know if besides the above you have any other reviewers you'd like to recommend. You can mention their GitHub handles here but please leave out the `@` symbol so they are not tagged. Thanks.
>
> Some other points that require you attention:
>
>     * [ ]   Given that you are the sole contributor, and that there is no history of issues or contributions added by others, can you confirm you have a long term commitment to this open source software project? We essentially want to check if this will be a maintained open source project with a growing community, or if this is simply a "quick GitHub upload of your old PhD work", and that you will not maintain it in the future. JOSS only publishes works expected to be maintained/supported long term.
>
>     * [ ]   Can you add contributing guidelines to your project? You can add a simple CONTRIBUTING.md file which describes how you'd like others to contribute, and link to it in your README.md (some find this website useful for inspiration when creating such a file: https://contributing.md/example/).
>
>     * [ ]  Please confirm you are the sole author, e.g. your PhD advisor or others need not be added here?
>
>     * [ ]  I have noticed the paper lacks an acknowledgements section. Although the section is not mandatory, it is usual for derivatives of PhD work to acknowledge funding for instance. So please check if adding this would be appropriate. You could also choose to acknowledge other contributors (not necessarily code) here if you like.

As I'm not sure my email reply went through, I'll respond here.

Thanks for your consideration. I think the suggested reviewers are just fine for this project, especially since they also have FEniCS experience.

Yes, I can confirm that I have a long term commitment to this project. Extensive work has been put into this code, far beyond my PhD thesis, and it contains many implementations of cardiac mechanics-related problems that are not easily available via open source and I am sure the code will find its users.

I've added a CONTRIBUTING.md file to the project.

Since the code actually has been developed after my PhD time and inbetween/outside of periods of funding/my postdoc time, I'm not sure how to design the acknowledgements section. But I'll think about it.

Best,
Marc
