### 2023-08-08 marchirschvogel/ambit#1 @finsberg
[JOSS REVIEW] Simplify installation instructions
Thanks for a great package!

I think the installation instruction could be simplified a lot which would make the library much more appealing to a greater audience.

- [x] #3

- [x] **Create an ambit docker image:** Instead to asking the user to first pull the dolfinx docker image, it would be better if you could build your own `ambit` image on top of dolfinx. That way, users could pull your image directly without the need to care about `dolfinx`. Also, this has another advantage that if `dolfinx` updates and breaks `ambit` you can still ship your own docker image that is based on an earlier version of `dolfinx` and then build new images when you have fixed all issues with the new version of `dolfinx`. You can see how this was can be set up here: https://github.com/finsberg/fenicsx-pulse/blob/main/.github/workflows/docker-image.yml

- [ ] **Conda package:** Since dolfinx is also availabel on [conda](https://anaconda.org/conda-forge/fenics-dolfinx) it would be good to also upload `ambit` to conda forge. You read how to do this [here](https://conda-forge.org/#contribute)

This is related to https://github.com/openjournals/joss-reviews/issues/5744
