### 2023-08-08 marchirschvogel/ambit#2 @finsberg
[JOSS REVIEW] BSD 4-Clause License is not OSI approved
According to https://opensource.org/licenses/?ls=BSD BSD 4-Clause is not OSI approved which is a [requirement for JOSS](https://joss.readthedocs.io/en/latest/review_criteria.html#software-license).

This is related to https://github.com/openjournals/joss-reviews/issues/5744
