### 2023-08-09 dc750a9
Monitor ROM projection timings if requested by user; add ROM input parameter check
