+++ testing/solid_flow0d_monodir_syspulcor_2Dheart_ROM.py
@@ -37,7 +37,6 @@ def main():
                             'restart_step'          : restart_step}
 
     ROM_PARAMS           = {'hdmfilenames'          : [basepath+'/input/checkpoint_romsnaps_2Dheart_u_*_1proc.dat'],
-                            'numhdms'               : 1,
                             'numsnapshots'          : 5,
                             'snapshotincr'          : 1,
                             'numredbasisvec'        : 5,
