+++ testing/frsi_artseg_modepartitionunity.py
@@ -41,8 +41,7 @@ def main():
                             'eigenvalue_cutoff'     : 1.0e-8,
                             'print_eigenproblem'    : True,
                             'surface_rom'           : [1,6],
-                            'filesource'            : 'rawtxt',
-                            'filereadin_tol'        : 1e-5}
+                            'filesource'            : 'rawtxt'}
 
     SOLVER_PARAMS        = {'solve_type'            : 'direct',
                             'tol_res'               : [1.0e-8,1.0e-8,1.0e-8],
