### 2023-08-09 1cd6b39
some debug prints
