+++ modules/coupling/fsi.py
@@ -64,6 +64,7 @@ class FSIProblem(problem_base):
 
         self.localsolve = False
         self.sub_solve = False
+        self.print_debug = self.io.print_debug
 
 
     def get_problem_var_list(self):
