### 2023-08-09 0f774f9
Some fixes in ROM framework
