### 2023-08-15 4e74381
Frist step of major code revamp: move data structure creation out of solver loops
