+++ modules/fluid/fluid.py
@@ -339,6 +339,13 @@ class FluidmechanicsProblem(problem_base):
         self.V_rom = self.V_v
         self.print_debug = self.io.print_debug
 
+        # number of fields involved
+        self.nfields = 2
+
+        # residual and matrix lists
+        self.r_list, self.r_list_rom = [None]*self.nfields, [None]*self.nfields
+        self.K_list, self.K_list_rom = [[None]*self.nfields for _ in range(self.nfields)],  [[None]*self.nfields for _ in range(self.nfields)]
+
 
     def get_problem_var_list(self):
 
@@ -700,6 +707,31 @@ class FluidmechanicsProblem(problem_base):
             sys.stdout.flush()
 
 
+    def set_problem_vector_matrix_structures(self, rom=None):
+
+        self.r_v = fem.petsc.create_vector(self.res_v)
+        if self.num_dupl > 1:
+            self.r_p = fem.petsc.create_vector_block(self.res_p)
+        else:
+            self.r_p = fem.petsc.create_vector(self.res_p)
+
+        self.K_vv = fem.petsc.create_matrix(self.jac_vv)
+        if self.num_dupl > 1:
+            self.K_vp = fem.petsc.create_matrix_block(self.jac_vp_)
+            self.K_pv = fem.petsc.create_matrix_block(self.jac_pv_)
+        else:
+            self.K_vp = fem.petsc.create_matrix(self.jac_vp)
+            self.K_pv = fem.petsc.create_matrix(self.jac_pv)
+
+        if self.stabilization is not None:
+            if self.num_dupl > 1:
+                self.K_pp = fem.petsc.create_matrix_block(self.jac_pp_)
+            else:
+                self.K_pp = fem.petsc.create_matrix(self.jac_pp)
+        else:
+            self.K_pp = None
+
+
     def assemble_residual(self, t, subsolver=None):
 
         # NOTE: we do not linearize integrated pressure-dependent valves w.r.t. p,
@@ -710,76 +742,82 @@ class FluidmechanicsProblem(problem_base):
         #     self.evaluate_robin_valve(t)
 
         # assemble velocity rhs vector
-        r_v = fem.petsc.assemble_vector(self.res_v)
-        fem.apply_lifting(r_v, [self.jac_vv], [self.bc.dbcs], x0=[self.v.vector], scale=-1.0)
-        r_v.ghostUpdate(addv=PETSc.InsertMode.ADD, mode=PETSc.ScatterMode.REVERSE)
-        fem.set_bc(r_v, self.bc.dbcs, x0=self.v.vector, scale=-1.0)
+        with self.r_v.localForm() as r_local: r_local.set(0.0)
+        fem.petsc.assemble_vector(self.r_v, self.res_v)
+        fem.apply_lifting(self.r_v, [self.jac_vv], [self.bc.dbcs], x0=[self.v.vector], scale=-1.0)
+        self.r_v.ghostUpdate(addv=PETSc.InsertMode.ADD, mode=PETSc.ScatterMode.REVERSE)
+        fem.set_bc(self.r_v, self.bc.dbcs, x0=self.v.vector, scale=-1.0)
 
         # assemble pressure rhs vector
+        with self.r_p.localForm() as r_local: r_local.set(0.0)
         if self.num_dupl > 1:
-            r_p = fem.petsc.assemble_vector_block(self.res_p, self.dummat, bcs=[]) # ghosts are updated inside assemble_vector_block
+            fem.petsc.assemble_vector_block(self.r_p, self.res_p, self.dummat, bcs=[]) # ghosts are updated inside assemble_vector_block
         else:
-            r_p = fem.petsc.assemble_vector(self.res_p)
[…]
