+++ modules/ioparams.py
@@ -22,7 +22,7 @@ def check_params_io(params):
                     'output_path',
                     'output_path_0D',
                     'output_path_pre',
-                    'print_debug',
+                    'print_enhanced_info',
                     'problem_type',
                     'restart_io_type',
                     'restart_step',
@@ -204,7 +204,6 @@ def check_params_rom(params):
                     'numsnapshots',
                     'partitions',
                     'print_eigenproblem',
-                    'print_projection_info',
                     'redbasisvec_penalties',
                     'snapshotincr',
                     'surface_rom',
