### 2023-08-16 f6acd84
Final step of data structure revamp: reuse matrices in all preconditioner setups; add own bgs2x2 testcase; rename some testcases
