+++ modules/solver/solver_nonlin.py
@@ -28,7 +28,7 @@ import ioparams
 # standard nonlinear solver for FEM problems
 class solver_nonlinear:
 
-    def __init__(self, pb, solver_params, cp=None):
+    def __init__(self, pb, solver_params, subsolver=None, cp=None):
 
         ioparams.check_params_solver(solver_params)
 
@@ -80,7 +80,9 @@ class solver_nonlinear:
             self.indp2 = self.lsp+2 # indent length for partitioned problem print
             self.lsp += 54
 
-        self.indlen = 1
+        self.indlen_ = [1]*self.nprob
+        # currently, we only have setup the output for two partitioned problems (will there ever be cases with more...?)
+        if self.nprob>1: self.indlen_[1] = self.indp2
 
         self.r_list_sol = [[]]*self.nprob
         self.K_list_sol = [[]]*self.nprob
@@ -93,11 +95,16 @@ class solver_nonlinear:
                 self.r_list_sol[npr] = self.pb[npr].r_list
                 self.K_list_sol[npr] = self.pb[npr].K_list
 
-        # sub-solver (for Lagrange-type constraints governed by a nonlinear system, e.g. 3D-0D coupling)
-        if self.pb[0].sub_solve:
-            self.subsol = solver_nonlinear_ode([self.pb[0].pb0], solver_params['subsolver_params'])
-        else:
-            self.subsol = None
+        # nested and merged (monolithic) residual and matrix objects
+        self.r_full_nest = [None]*self.nprob
+        self.K_full_nest = [None]*self.nprob
+        self.P_full_nest = [None]*self.nprob
+
+        self.r_full_merged = [None]*self.nprob
+        self.K_full_merged = [None]*self.nprob
+        self.P_full_merged = [None]*self.nprob
+
+        self.subsol = subsolver
 
         self.initialize_petsc_solver()
 
@@ -187,11 +194,6 @@ class solver_nonlinear:
         for k in is_option_keys:
             if k not in self.iset_options.keys(): self.iset_options[k] = False
 
-        if any(list(self.iset_options.values())):
-            self.merge_prec_mat = True
-        else:
-            self.merge_prec_mat = False
-
         self.iset = [[]]*self.nprob
 
         try: self.print_local_iter = solver_params['print_local_iter']
@@ -223,25 +225,36 @@ class solver_nonlinear:
             # create solver
             self.ksp[npr] = PETSc.KSP().create(self.comm)
 
+            # get reduced matrices to set ksp operators
+            if self.pb[npr].rom:
+                self.jacobian_problem_actions(0., npr, 0.)
+
             if self.solvetype[npr]=='direct':
 
                 self.ksp[npr].setType("preonly")
                 self.ksp[npr].getPC().setType("lu")
                 self.ksp[npr].getPC().setFactorSolverType(self.direct_solver)
 
+                # prepare merged matrix structure
+                if self.solvetype[npr]=='direct' and self.nfields[npr] > 1:
+                    self.K_full_nest[npr] = PETSc.Mat().createNest(self.K_list_sol[npr], isrows=None, iscols=None, comm=self.comm)
+                    self.K_full_merged[npr] = self.K_full_nest[npr].convert("aij")
+
             elif self.solvetype[npr]=='iterative':
 
                 self.ksp[npr].setInitialGuessNonzero(False)
                 self.ksp[npr].setNormType(self.linnormtype) # cf. https://www.mcs.anl.gov/petsc/petsc4py-current/docs/apiref/petsc4py.PETSc.KSP.NormType-class.html
 
-                # get matrices to set ksp operators
-                if self.pb[npr].rom:
-                    self.jacobian_problem_actions(0., npr, 0.)
[…]

[…]
