### 2023-08-17 5b36a9e
Some clean-up and fixes in solver and prec classes, increased doc of tests
