+++ testing/fluid_p1p1_stab_cylinder_schur2x2.py
@@ -1,14 +1,11 @@
 #!/usr/bin/env python3
 
-"""
-transient incompressible Navier-Stokes flow in a cylinder with axial Neumann and two outflows
-- stabilized P1P1 elements for velocity and pressure
-- Backward-Euler time stepping scheme
-- 2 material domains in fluid (w/ same parameters though)
-- internal valve, requiring duplicate pressure nodes at that internal surface
-- works currently only with mixed dolfinx branch (USE_MIXED_DOLFINX_BRANCH)
-- checkpoint writing of domain-wise discontinuous pressure field
-"""
+# tests:
+# - steady incompressible Navier-Stokes flow in a cylinder with axial Neumann and two outflows
+# - Stabilized P1P1 elements for velocity and pressure
+# - Backward-Euler time stepping scheme
+# - 2 material domains in fluid (w/ same parameters though)
+# - Schur2x2 block preconditioner
 
 import ambit
 
@@ -22,34 +19,33 @@ import resultcheck
 def main():
     
     basepath = str(Path(__file__).parent.absolute())
-
-    # reads in restart step from the command line
-    try: restart_step = int(sys.argv[1])
-    except: restart_step = 0
+    
 
     IO_PARAMS           = {'problem_type'          : 'fluid',
-                           'USE_MIXED_DOLFINX_BRANCH' : True,
-                           'duplicate_mesh_domains': [1,2],
                            'mesh_domain'           : basepath+'/input/cylinder_domain.xdmf',
                            'mesh_boundary'         : basepath+'/input/cylinder_boundary.xdmf',
                            'write_results_every'   : 1,
-                           'write_restart_every'   : 9,
-                           'restart_step'          : restart_step,
                            'output_path'           : basepath+'/tmp/',
                            'results_to_write'      : ['velocity','pressure'],
-                           'simname'               : 'fluid_p1p1_stab_cylinder_valve'}
+                           'simname'               : 'fluid_p1p1_stab_cylinder'}
 
-    SOLVER_PARAMS       = {'solve_type'            : 'direct',
-                           'direct_solver'         : 'mumps',
+    SOLVER_PARAMS       = {'solve_type'            : 'iterative',
+                           'iterative_solver'      : 'fgmres',
+                           'block_precond'         : 'schur2x2',
+                           'precond_fields'        : [{'prec':'amg'}, {'prec':'amg'}], # v, p
+                           'tol_lin_rel'           : 1.0e-5,
+                           'tol_lin_abs'           : 1.0e-30,
+                           'res_lin_monitor'       : 'rel',
+                           'print_liniter_every'   : 10,
                            'tol_res'               : 1.0e-8,
                            'tol_inc'               : 1.0e-8}
 
     TIME_PARAMS_FLUID   = {'maxtime'               : 1.0,
                            'numstep'               : 10,
-                           #'numstep_stop'          : 3,
+                           'numstep_stop'          : 2,
                            'timint'                : 'ost',
                            'theta_ost'             : 1.0,
-                           'fluid_governing_type'  : 'navierstokes_transient'}
+                           'fluid_governing_type'  : 'navierstokes_steady'}
     
     FEM_PARAMS          = {'order_vel'             : 1,
[…]
