### 2023-08-17 aedb9a0
Make prec reuse option adaptive, depending on user input; add single-field amg testcase for solid
