### 2023-08-17 db0fea5
Don't change reference when prec is reused
