### 2023-08-17 715c023
setup field-specific precs at the beginning
