### 2023-08-18 48c47a0
Only create merged mat objects when needed
