### 2023-08-18 fb0f4bd
fix in timing measurements for partitioned solves
