### 2023-08-18 099658c
move counter adapt out of problem loop
