### 2023-08-18 8c015e6
fix destroy for partitioned solve; some comments
