+++ src/ambit/ioroutines.py
@@ -91,6 +91,12 @@ class IO:
             try: self.mt_d = infile.read_meshtags(self.mesh, name=self.gridname_domain)
             except: self.mt_d = None
 
+        # mesh degree and scalar, vector, tensor output function spaces
+        self.mesh_degree = self.mesh._ufl_domain._ufl_coordinate_element.degree()
+        self.V_out_scalar = fem.FunctionSpace(self.mesh, ("CG", self.mesh_degree))
+        self.V_out_vector = fem.VectorFunctionSpace(self.mesh, ("CG", self.mesh_degree))
+        self.V_out_tensor = fem.TensorFunctionSpace(self.mesh, ("CG", self.mesh_degree))
+
         # master mesh object (need if fields are actually subdomains, e.g. in FSI)
         self.mesh_master = self.mesh
         self.mt_d_master = self.mt_d
@@ -171,7 +177,9 @@ class IO:
 
         outfile = io.XDMFFile(self.comm, self.output_path_pre+'/results_'+pb.simname+'_'+name+'.xdmf', 'w')
         outfile.write_mesh(self.mesh)
-        outfile.write_function(func, t)
+        func_out = fem.Function(self.V_out_vector)
+        func_out.interpolate(func)
+        outfile.write_function(func_out, t)
 
 
     def write_restart(self, pb, N):
@@ -354,118 +362,166 @@ class IO_solid(IO):
                 for res in pb.results_to_write:
 
                     if res=='displacement':
-                        self.resultsfiles[res].write_function(pb.u, indicator)
+                        u_out = fem.Function(self.V_out_vector)
+                        u_out.interpolate(pb.u)
+                        self.resultsfiles[res].write_function(u_out, indicator)
                     elif res=='velocity': # passed in v is not a function but form, so we have to project
                         self.v_proj = project(pb.vel, pb.V_u, pb.dx_, nm="Velocity") # class variable for testing
-                        self.resultsfiles[res].write_function(self.v_proj, indicator)
+                        v_out = fem.Function(self.V_out_vector)
+                        v_out.interpolate(self.v_proj)
+                        self.resultsfiles[res].write_function(v_out, indicator)
                     elif res=='acceleration': # passed in a is not a function but form, so we have to project
                         self.a_proj = project(pb.acc, pb.V_u, pb.dx_, nm="Acceleration") # class variable for testing
-                        self.resultsfiles[res].write_function(self.a_proj, indicator)
+                        a_out = fem.Function(self.V_out_vector)
+                        a_out.interpolate(self.a_proj)
+                        self.resultsfiles[res].write_function(a_out, indicator)
                     elif res=='pressure':
-                        self.resultsfiles[res].write_function(pb.p, indicator)
+                        p_out = fem.Function(self.V_out_scalar)
+                        p_out.interpolate(pb.p)
+                        self.resultsfiles[res].write_function(p_out, indicator)
                     elif res=='cauchystress':
                         stressfuncs=[]
                         for n in range(pb.num_domains):
                             stressfuncs.append(pb.ma[n].sigma(pb.u,pb.p,pb.vel,ivar=pb.internalvars))
                         cauchystress = project(stressfuncs, pb.Vd_tensor, pb.dx_, nm="CauchyStress")
-                        self.resultsfiles[res].write_function(cauchystress, indicator)
+                        cauchystress_out = fem.Function(self.V_out_tensor)
+                        cauchystress_out.interpolate(cauchystress)
+                        self.resultsfiles[res].write_function(cauchystress_out, indicator)
                     elif res=='cauchystress_nodal':
                         stressfuncs=[]
                         for n in range(pb.num_domains):
                             stressfuncs.append(pb.ma[n].sigma(pb.u,pb.p,pb.vel,ivar=pb.internalvars))
[…]
