### 2023-08-19 a182e17
fix split mesh IO; rename test folder
