### 2023-08-19 a3d84eb
update install instructions
