+++ README.md
@@ -57,7 +57,7 @@ Still experimental / to-do:
 
 * If 0D models should be used, it seems that we have to install sympy (not part of docker container anymore) - in the folder where you pulled ambit to, do:
 
-``cd ambit && mkdir modules/ext && pip3 install --target=ext mpmath --no-deps --no-cache-dir && pip3 install --target=ext sympy --no-deps --no-cache-dir && cd ..``
+``cd ambit && mkdir ext && pip3 install --target=ext mpmath --no-deps --no-cache-dir && pip3 install --target=ext sympy --no-deps --no-cache-dir && cd ..``
 
 * Launch the container in a konsole/terminal window by simply typing
 
