### 2023-08-21 314a00e
Options for different types of Schur scaling: diag, rowsum, none; minor other things
