+++ src/ambit/boundaryconditions.py
@@ -651,6 +651,14 @@ class boundary_cond_fluid(boundary_cond):
 
                         w += self.vf.deltaW_ext_robin_dashpot_normal_cur(v, r['visc'], db_, Fale=Fale)
 
+                elif r['dir'] == 'normal_cross': # cross normal
+
+                    for i in range(len(r['id'])):
+
+                        db_ = ufl.ds(domain=self.io.mesh_master, subdomain_data=mdata, subdomain_id=r['id'][i], metadata={'quadrature_degree': self.quad_degree})
+
+                        w += self.vf.deltaW_ext_robin_dashpot_normal_cross(v, r['visc'], db_, Fale=Fale)
+
                 else:
                     raise NameError("Unknown dir option for Robin BC!")
 
