+++ src/ambit/solid/solid_material.py
@@ -243,7 +243,6 @@ class materiallaw:
 
 
 
-
 # inversion of growth tensors can be performed with Sherman-Morrison formula:
 # TeX: \left(\boldsymbol{A}-\boldsymbol{u}\otimes \boldsymbol{v}\right)^{-1} = \boldsymbol{A}^{-1} + \frac{\boldsymbol{A}^{-1}\,\boldsymbol{u}\otimes \boldsymbol{v}\,\boldsymbol{A}^{-1}}{1-\boldsymbol{v}\cdot \boldsymbol{A}^{-1}\boldsymbol{u}}
 
