### 2023-08-22 89f33d5
More user control for Schur scaling
