### 2023-08-22 df35aaa
Add back function names to output
