+++ src/ambit/ioroutines.py
@@ -177,7 +177,7 @@ class IO:
 
         outfile = io.XDMFFile(self.comm, self.output_path_pre+'/results_'+pb.simname+'_'+name+'.xdmf', 'w')
         outfile.write_mesh(self.mesh)
-        func_out = fem.Function(self.V_out_vector)
+        func_out = fem.Function(self.V_out_vector, name=func.name)
         func_out.interpolate(func)
         outfile.write_function(func_out, t)
 
@@ -362,21 +362,21 @@ class IO_solid(IO):
                 for res in pb.results_to_write:
 
                     if res=='displacement':
-                        u_out = fem.Function(self.V_out_vector)
+                        u_out = fem.Function(self.V_out_vector, name=pb.u.name)
                         u_out.interpolate(pb.u)
                         self.resultsfiles[res].write_function(u_out, indicator)
                     elif res=='velocity': # passed in v is not a function but form, so we have to project
                         self.v_proj = project(pb.vel, pb.V_u, pb.dx_, nm="Velocity") # class variable for testing
-                        v_out = fem.Function(self.V_out_vector)
+                        v_out = fem.Function(self.V_out_vector, name=self.v_proj.name)
                         v_out.interpolate(self.v_proj)
                         self.resultsfiles[res].write_function(v_out, indicator)
                     elif res=='acceleration': # passed in a is not a function but form, so we have to project
                         self.a_proj = project(pb.acc, pb.V_u, pb.dx_, nm="Acceleration") # class variable for testing
-                        a_out = fem.Function(self.V_out_vector)
+                        a_out = fem.Function(self.V_out_vector, name=self.a_proj.name)
                         a_out.interpolate(self.a_proj)
                         self.resultsfiles[res].write_function(a_out, indicator)
                     elif res=='pressure':
-                        p_out = fem.Function(self.V_out_scalar)
+                        p_out = fem.Function(self.V_out_scalar, name=pb.p.name)
                         p_out.interpolate(pb.p)
                         self.resultsfiles[res].write_function(p_out, indicator)
                     elif res=='cauchystress':
@@ -384,7 +384,7 @@ class IO_solid(IO):
                         for n in range(pb.num_domains):
                             stressfuncs.append(pb.ma[n].sigma(pb.u,pb.p,pb.vel,ivar=pb.internalvars))
                         cauchystress = project(stressfuncs, pb.Vd_tensor, pb.dx_, nm="CauchyStress")
-                        cauchystress_out = fem.Function(self.V_out_tensor)
+                        cauchystress_out = fem.Function(self.V_out_tensor, name=cauchystress.name)
                         cauchystress_out.interpolate(cauchystress)
                         self.resultsfiles[res].write_function(cauchystress_out, indicator)
                     elif res=='cauchystress_nodal':
@@ -392,7 +392,7 @@ class IO_solid(IO):
                         for n in range(pb.num_domains):
                             stressfuncs.append(pb.ma[n].sigma(pb.u,pb.p,pb.vel,ivar=pb.internalvars))
                         cauchystress_nodal = project(stressfuncs, pb.V_tensor, pb.dx_, nm="CauchyStress_nodal")
-                        cauchystress_nodal_out = fem.Function(self.V_out_tensor)
+                        cauchystress_nodal_out = fem.Function(self.V_out_tensor, name=cauchystress_nodal.name)
                         cauchystress_nodal_out.interpolate(cauchystress_nodal)
                         self.resultsfiles[res].write_function(cauchystress_nodal_out, indicator)
                     elif res=='cauchystress_principal':
@@ -401,7 +401,7 @@ class IO_solid(IO):
[…]
