### 2023-08-23 448f924
Add option for reduced stabilization scheme
