+++ src/ambit/fluid/fluid.py
@@ -486,7 +486,7 @@ class FluidmechanicsProblem(problem_base):
             assert(self.order_vel==self.order_pres)
 
             vscale = self.stabilization['vscale']
-            # these cell metrics don't seem to work for linear hexes... :(
+
             h = self.io.hd0 # cell diameter (could also use max edge length self.io.emax0, but seems to yield similar/same results)
 
             # full scheme
@@ -531,6 +531,9 @@ class FluidmechanicsProblem(problem_base):
             # reduced scheme: missing transient NS term as well as divergence stress term of strong residual
             if self.stabilization['scheme']=='supg_pspg2':
 
+                try: symm = self.stabilization['symmetric']
+                except: symm = False
+
                 # optimized for first-order
                 assert(self.order_vel==1)
                 assert(self.order_pres==1)
@@ -547,8 +550,8 @@ class FluidmechanicsProblem(problem_base):
                     delta2 = dscales[1] * self.rho[n] * h * vscale
                     delta3 = dscales[2] * h / vscale
 
-                    self.deltaW_int     += self.vf.stab_v(delta1, delta2, delta3, self.v, self.p_[j], self.dx_[n], w=self.alevar['w'], Fale=self.alevar['Fale'])
-                    self.deltaW_int_old += self.vf.stab_v(delta1, delta2, delta3, self.v_old, self.p_old_[j], self.dx_[n], w=self.alevar['w_old'], Fale=self.alevar['Fale_old'])
+                    self.deltaW_int     += self.vf.stab_v(delta1, delta2, delta3, self.v, self.p_[j], self.dx_[n], w=self.alevar['w'], Fale=self.alevar['Fale'], symmetric=symm)
+                    self.deltaW_int_old += self.vf.stab_v(delta1, delta2, delta3, self.v_old, self.p_old_[j], self.dx_[n], w=self.alevar['w_old'], Fale=self.alevar['Fale_old'], symmetric=symm)
 
                     self.deltaW_p[n]     += self.vf.stab_p(delta1, delta3, self.v, self.p_[j], self.var_p_[j], self.rho[n], self.dx_[n], w=self.alevar['w'], Fale=self.alevar['Fale'])
                     self.deltaW_p_old[n] += self.vf.stab_p(delta1, delta3, self.v_old, self.p_old_[j], self.var_p_[j], self.rho[n], self.dx_[n], w=self.alevar['w_old'], Fale=self.alevar['Fale_old'])
