### 2023-08-28 dc6147c
Add new CRLinoutlink 0D model: two 2-el WKs in series linking a 3D in- to a 3D outflow
