+++ tests/solid_flow0d_monolagr_CRLinoutlink_chambers.py
@@ -1,0 +1,123 @@
+#!/usr/bin/env python3
+
+"""
+solid 3D-0D coupling: hollow solid chambers linked via CRLinoutlink 0D model (two CRL models in series that link in- to outflux 3D model surface)
+one active block contracts and fills the other
+"""
+
+import ambit
+
+import sys, traceback
+import numpy as np
+from pathlib import Path
+
+import resultcheck
+
+
+def main():
+
+    basepath = str(Path(__file__).parent.absolute())
+
+    IO_PARAMS            = {'problem_type'          : 'solid_flow0d',
+                            'mesh_domain'           : basepath+'/input/chambers2hex_domain.xdmf',
+                            'mesh_boundary'         : basepath+'/input/chambers2hex_boundary.xdmf',
+                            'fiber_data'            : [np.array([1.0,0.0,0.0])],
+                            'write_results_every'   : 1,
+                            'output_path'           : basepath+'/tmp/',
+                            'results_to_write'      : ['displacement'],
+                            'simname'               : 'CRLinoutlink'}
+
+    SOLVER_PARAMS        = {'solve_type'            : 'direct',
+                            'tol_res'               : 1.0e-8,
+                            'tol_inc'               : 1.0e-8,
+                            'subsolver_params'      : {'tol_res' : 1.0e-8, 'tol_inc' : 1.0e-8}}
+
+    TIME_PARAMS_SOLID    = {'maxtime'               : 0.3,
+                            'numstep'               : 10,
+                            #'numstep_stop'          : 10,
+                            'timint'                : 'ost',
+                            'theta_ost'             : 1.0}
+
+    TIME_PARAMS_FLOW0D   = {'timint'                : 'ost',
+                            'theta_ost'             : 1.0,
+                            'initial_conditions'    : {'q_in_0' : 0.0, 'q_d_0' : 0.0, 'p_d_0' : 0.0, 'q_out_0' : 0.0}}
+
+    MODEL_PARAMS_FLOW0D  = {'modeltype'             : 'CRLinoutlink',
+                            'parameters'            : {'C_in' : 0.001, 'R_in' : 1e-6, 'L_in' : 0., 'C_out' :  0.001, 'R_out' : 1e-6, 'L_out' : 0.}}
+
+    FEM_PARAMS           = {'order_disp'            : 1,
+                            'order_pres'            : 1,
+                            'quad_degree'           : 5,
+                            'incompressible_2field' : False}
+
+    COUPLING_PARAMS      = {'surface_ids'           : [[2],[4]],
+                            'coupling_quantity'     : ['pressure']*2,
+                            'variable_quantity'     : ['flux']*2,
+                            'cq_factor'             : [1.,-1.], # out-flow positive, in-flow negative
+                            'coupling_type'         : 'monolithic_lagrange',
+                            'print_subiter'         : True}
+
+    MATERIALS            = {'MAT1' : {'neohooke_dev' : {'mu' : 100.}, 'ogden_vol' : {'kappa' : 100./(1.-2.*0.49)}, 'inertia' : {'rho0' : 1.0e-6}, 'active_fiber' : {'sigma0' : 50.0, 'alpha_max' : 15.0, 'alpha_min' : -20.0, 'activation_curve' : 1}},
[…]
