### 2023-08-29 882c60f
avoid vector creations in assemble for 3D0D
