### 2023-08-29 5213df8
some restructuring in solver
