+++ src/ambit/solver/solver_nonlin.py
@@ -110,6 +110,56 @@ class solver_nonlinear:
 
         self.initialize_petsc_solver()
 
+        # offset array for multi-field systems
+        self.offsetarr = [[]]*self.nprob
+        for npr in range(self.nprob):
+
+            self.offsetarr[npr] = [0]
+            off=0
+            for n in range(self.nfields[npr]):
+                if n==0:
+                    if self.pb[npr].rom: # currently, ROM is only implemented for the first variable in the system!
+                        off += self.pb[npr].rom.V.getLocalSize()[1]
+                    else:
+                        off += self.x[npr][0].getLocalSize()
+                else:
+                    off += self.x[npr][n].getLocalSize()
+
+                self.offsetarr[npr].append(off)
+
+        self.del_x, self.del_x_rom, self.x_start = [], [], []
+        self.del_x_sol = [[]]*self.nprob
+
+        for npr in range(self.nprob):
+
+            self.del_x.append([[]]*self.nfields[npr])
+            self.x_start.append([[]]*self.nfields[npr])
+            self.del_x_rom.append([[]]*self.nfields[npr])
+
+        for npr in range(self.nprob):
+
+            for n in range(self.nfields[npr]):
+                # solution increments for Newton
+                self.del_x[npr][n] = self.x[npr][n].duplicate()
+                self.del_x[npr][n].set(0.0)
+                if self.pb[npr].rom and npr==0:
+                    if n==0:
+                        self.del_x_rom[npr][n] = self.pb[npr].rom.V.createVecRight()
+                        self.del_x_rom[npr][n].set(0.0)
+                    else:
+                        self.del_x_rom[npr][n] = self.del_x[npr][n]
+                # start vector (needed for reset of Newton in case of divergence)
+                self.x_start[npr][n] = self.x[npr][n].duplicate()
+                self.x[npr][n].assemble()
+                if self.pb[npr].sub_solve: # can only be a 0D model so far...
+                    self.s_start = self.pb[npr].pb0.s.duplicate()
+                    self.pb[npr].pb0.s.assemble()
+
+            if self.pb[npr].rom:
+                self.del_x_sol[npr] = self.del_x_rom[npr]
+            else:
+                self.del_x_sol[npr] = self.del_x[npr]
+
         self.li_s = [] # linear iterations over all solves
         self.ni_all = 0 # all nonlinear iterations over all solves (to determine prec updates for instance)
 
@@ -446,57 +496,16 @@ class solver_nonlinear:
 
     def newton(self, t, localdata={}):
 
-        # offset array for multi-field systems
-        self.offsetarr = [[]]*self.nprob
-        for npr in range(self.nprob):
 
-            self.offsetarr[npr] = [0]
-            off=0
+
+        # set start vectors
+        for npr in range(self.nprob):
             for n in range(self.nfields[npr]):
-                if n==0:
-                    if self.pb[npr].rom: # currently, ROM is only implemented for the first variable in the system!
-                        off += self.pb[npr].rom.V.getLocalSize()[1]
-                    else:
-                        off += self.x[npr][0].getLocalSize()
-                else:
-                    off += self.x[npr][n].getLocalSize()
-
-                self.offsetarr[npr].append(off)
-
-        del_x, del_x_rom, x_start = [], [], []
-        del_x_sol = [[]]*self.nprob
-
-        for npr in range(self.nprob):
-
-            del_x.append([[]]*self.nfields[npr])
-            x_start.append([[]]*self.nfields[npr])
-            del_x_rom.append([[]]*self.nfields[npr])
-
-        for npr in range(self.nprob):
-
-            for n in range(self.nfields[npr]):
[…]
