+++ src/ambit/coupling/solid_flow0d.py
@@ -777,7 +777,7 @@ class SolidmechanicsFlow0DSolver(solver_base):
         if (self.pb.pbs.prestress_initial or self.pb.pbs.prestress_initial_only) and self.pb.pbs.restart_step == 0:
             # solve solid prestress problem
             self.solverprestr.solve_initial_prestress()
-            self.solverprestr.solnln.ksp[0].destroy()
+            self.solverprestr.solnln.destroy()
 
         # consider consistent initial acceleration
         if self.pb.pbs.timint != 'static' and self.pb.pbs.restart_step == 0 and not self.pb.restart_multiscale:
