+++ src/ambit/solid/solid.py
@@ -1059,7 +1059,6 @@ class SolidmechanicsSolver(solver_base):
             if self.pb.comm.rank == 0:
                 print("Prestress only done. To resume, set file path(s) in 'prestress_from_file' and read in u_pre.")
                 sys.stdout.flush()
-            self.solnln.destroy()
             os._exit(0)
 
         # reset PTC flag to what it was
@@ -1070,7 +1069,6 @@ class SolidmechanicsSolver(solver_base):
         # set flag to false again
         self.pb.prestress_initial = False
         self.pb.set_problem_residual_jacobian_forms()
-        self.solnln.destroy()
 
 
 # prestress solver, to be called from other (coupled) problems
