### 2023-08-29 f23f9ae
destroy 0d ksp
