### 2023-08-29 171e750
further solver revamp
