+++ src/ambit/solver/solver_nonlin.py
@@ -108,8 +108,6 @@ class solver_nonlinear:
 
         self.subsol = subsolver
 
-        self.initialize_petsc_solver()
-
         # offset array for multi-field systems
         self.offsetarr = [[]]*self.nprob
         for npr in range(self.nprob):
@@ -159,6 +157,8 @@ class solver_nonlinear:
                 self.del_x_sol[npr] = self.del_x_rom[npr]
             else:
                 self.del_x_sol[npr] = self.del_x[npr]
+
+        self.initialize_petsc_solver()
 
         self.li_s = [] # linear iterations over all solves
         self.ni_all = 0 # all nonlinear iterations over all solves (to determine prec updates for instance)
@@ -287,9 +287,14 @@ class solver_nonlinear:
             # create solver
             self.ksp[npr] = PETSc.KSP().create(self.comm)
 
-            # get reduced matrices to set ksp operators
+            # perform initial matrix and residual reduction to set the correct arrays
             if self.pb[npr].rom:
-                self.jacobian_problem_actions(0., npr, 0.)
+                self.pb[npr].rom.reduce_stiffness(self.pb[npr].K_list, self.pb[npr].K_list_rom, self.pb[npr].K_list_tmp)
+                self.pb[npr].rom.reduce_residual(self.pb[npr].r_list, self.pb[npr].r_list_rom, x=self.x[npr][0])
+
+            # create nested matrix and residual structures
+            self.K_full_nest[npr] = PETSc.Mat().createNest(self.K_list_sol[npr], isrows=None, iscols=None, comm=self.comm)
+            self.r_full_nest[npr] = PETSc.Vec().createNest(self.r_list_sol[npr])
 
             if self.solvetype[npr]=='direct':
 
@@ -299,17 +304,25 @@ class solver_nonlinear:
 
                 # prepare merged matrix structure
                 if self.solvetype[npr]=='direct' and self.nfields[npr] > 1:
-                    self.K_full_nest[npr] = PETSc.Mat().createNest(self.K_list_sol[npr], isrows=None, iscols=None, comm=self.comm)
+
                     self.K_full_merged[npr] = PETSc.Mat()
                     self.K_full_nest[npr].convert("aij", out=self.K_full_merged[npr])
+                    # solution increment
+                    self.del_full = self.K_full_merged[npr].createVecLeft()
 
             elif self.solvetype[npr]=='iterative':
 
                 self.ksp[npr].setInitialGuessNonzero(False)
                 self.ksp[npr].setNormType(self.linnormtype)
 
-                self.K_full_nest[npr] = PETSc.Mat().createNest(self.K_list_sol[npr], isrows=None, iscols=None, comm=self.comm)
                 self.P_full_nest[npr] = self.K_full_nest[npr]
+
+                # solution increment
+                if self.nfields[npr] > 1:
+                    self.del_full = PETSc.Vec().createNest(self.del_x_sol[npr])
+                    # need to merge for non-fieldsplit-type preconditioners
+                    if not self.block_precond[npr] == 'fieldsplit':
+                        self.del_full = PETSc.Vec().createWithArray(self.del_full.getArray())
 
                 # prepare merged preconditioner matrix structure
                 if self.merge_prec_mat:
@@ -496,16 +509,12 @@ class solver_nonlinear:
 
     def newton(self, t, localdata={}):
 
-
-
         # set start vectors
         for npr in range(self.nprob):
             for n in range(self.nfields[npr]):
                 self.x_start[npr][n].axpby(1.0, 0.0, self.x[npr][n])
                 if self.pb[npr].sub_solve: # can only be a 0D model so far...
[…]
