+++ src/ambit/solver/solver_nonlin.py
@@ -296,11 +296,14 @@ class solver_nonlinear:
             self.K_full_nest[npr] = PETSc.Mat().createNest(self.K_list_sol[npr], isrows=None, iscols=None, comm=self.comm)
             self.r_full_nest[npr] = PETSc.Vec().createNest(self.r_list_sol[npr])
 
+            # preconditioner object (also used in direct solver context)
+            self.pc = self.ksp[npr].getPC()
+
             if self.solvetype[npr]=='direct':
 
                 self.ksp[npr].setType("preonly")
-                self.ksp[npr].getPC().setType("lu")
-                self.ksp[npr].getPC().setFactorSolverType(self.direct_solver)
+                self.pc.setType("lu")
+                self.pc.setFactorSolverType(self.direct_solver)
 
                 # prepare merged matrix structure
                 if self.solvetype[npr]=='direct' and self.nfields[npr] > 1:
@@ -344,7 +347,7 @@ class solver_nonlinear:
                     if self.block_precond[npr] == 'fieldsplit':
 
                         # see e.g. https://petsc.org/main/manual/ksp/#sec-block-matrices
-                        self.ksp[npr].getPC().setType("fieldsplit")
+                        self.pc.setType("fieldsplit")
                         # cf. https://petsc.org/main/manualpages/PC/PCCompositeType
 
                         if self.fieldsplit_type=='jacobi':
@@ -359,77 +362,78 @@ class solver_nonlinear:
                         else:
                             raise ValueError("Unknown fieldsplit_type option.")
 
-                        self.ksp[npr].getPC().setFieldSplitType(splittype)
+                        self.pc.setFieldSplitType(splittype)
 
                         nsets = len(self.iset[npr])
 
                         # normally, nsets = self.nfields, but for a surface-projected ROM (FrSI) problem, we have one more index set than fields
-                        if nsets==2:   self.ksp[npr].getPC().setFieldSplitIS(("f1", self.iset[npr][0]),("f2", self.iset[npr][1]))
-                        elif nsets==3: self.ksp[npr].getPC().setFieldSplitIS(("f1", self.iset[npr][0]),("f2", self.iset[npr][1]),("f3", self.iset[npr][2]))
-                        elif nsets==4: self.ksp[npr].getPC().setFieldSplitIS(("f1", self.iset[npr][0]),("f2", self.iset[npr][1]),("f3", self.iset[npr][2]),("f4", self.iset[npr][3]))
-                        elif nsets==5: self.ksp[npr].getPC().setFieldSplitIS(("f1", self.iset[npr][0]),("f2", self.iset[npr][1]),("f3", self.iset[npr][2]),("f4", self.iset[npr][3]),("f5", self.iset[npr][4]))
+                        if nsets==2:   self.pc.setFieldSplitIS(("f1", self.iset[npr][0]),("f2", self.iset[npr][1]))
+                        elif nsets==3: self.pc.setFieldSplitIS(("f1", self.iset[npr][0]),("f2", self.iset[npr][1]),("f3", self.iset[npr][2]))
+                        elif nsets==4: self.pc.setFieldSplitIS(("f1", self.iset[npr][0]),("f2", self.iset[npr][1]),("f3", self.iset[npr][2]),("f4", self.iset[npr][3]))
+                        elif nsets==5: self.pc.setFieldSplitIS(("f1", self.iset[npr][0]),("f2", self.iset[npr][1]),("f3", self.iset[npr][2]),("f4", self.iset[npr][3]),("f5", self.iset[npr][4]))
                         else: raise RuntimeError("Currently, no more than 5 fields/index sets are supported.")
 
                         # get the preconditioners for each block
-                        self.ksp[npr].getPC().setUp()
[…]
