### 2023-08-29 8667f21
make arrays class variable
