+++ src/ambit/solver/solver_nonlin.py
@@ -313,6 +313,8 @@ class solver_nonlinear:
                     # solution increment
                     self.del_full = self.K_full_merged[npr].createVecLeft()
 
+                    self.r_arr = np.zeros(self.r_full_nest[npr].getLocalSize())
+
             elif self.solvetype[npr]=='iterative':
 
                 self.ksp[npr].setInitialGuessNonzero(False)
@@ -324,6 +326,7 @@ class solver_nonlinear:
                 if self.nfields[npr] > 1:
                     if not self.block_precond[npr] == 'fieldsplit':
                         self.del_full = self.K_full_nest[npr].createVecLeft()
+                        self.r_arr = np.zeros(self.r_full_nest[npr].getLocalSize())
                     else:
                         self.del_full = PETSc.Vec().createNest(self.del_x_sol[npr])
 
@@ -585,10 +588,10 @@ class solver_nonlinear:
                             if self.comm.rank == 0:
                                 print(' '*self.indlen_[npr] + '      === MAT merge, te = %.4f s' % (tme))
                                 sys.stdout.flush()
+                        tmp=self.r_full_nest[npr].getArray(readonly=True)
 
-                        r_arr = self.r_full_nest[npr].getArray(readonly=True)
-                        self.r_full_merged[npr] = PETSc.Vec().createWithArray(r_arr)
-                        del r_arr
+                        self.r_arr[:] = self.r_full_nest[npr].getArray(readonly=True)
+                        self.r_full_merged[npr] = PETSc.Vec().createWithArray(self.r_arr)
 
                         self.ksp[npr].setOperators(self.K_full_merged[npr])
                         te += time.time() - tes
@@ -633,9 +636,8 @@ class solver_nonlinear:
 
                         # need to merge for non-fieldsplit-type preconditioners
                         if not self.block_precond[npr] == 'fieldsplit':
-                            r_arr = self.r_full_nest[npr].getArray(readonly=True)
-                            self.r_full_merged[npr] = PETSc.Vec().createWithArray(r_arr)
-                            del r_arr
+                            self.r_arr[:] = self.r_full_nest[npr].getArray(readonly=True)
+                            self.r_full_merged[npr] = PETSc.Vec().createWithArray(self.r_arr)
                             r = self.r_full_merged[npr]
                         else:
                             r = self.r_full_nest[npr]
