### 2023-08-29 329d372
move create vec out of solver loop
