### 2023-08-30 a84cc51
make sure that on-the-fly created numpy arrays are deleted once not needed anymore
