+++ src/ambit/timeintegration.py
@@ -274,48 +274,38 @@ class timeintegration_solid(timeintegration):
 
 
     def update_fields_newmark(self, u, u_old, v_old, a_old):
-        # update fields at the end of each time step
-        # get vectors (references)
-        u_vec, u0_vec  = u.vector, u_old.vector
-        v0_vec, a0_vec = v_old.vector, a_old.vector
-        u_vec.assemble(), u0_vec.assemble(), v0_vec.assemble(), a0_vec.assemble()
 
         # use update functions using vector arguments
-        a_vec = self.update_a_newmark(u_vec, u0_vec, v0_vec, a0_vec, ufl=False)
-        v_vec = self.update_v_newmark(u_vec, u0_vec, v0_vec, a0_vec, ufl=False)
+        a_vec = self.update_a_newmark(u.vector, u_old.vector, v_old.vector, a_old.vector, ufl=False)
+        v_vec = self.update_v_newmark(u.vector, u_old.vector, v_old.vector, a_old.vector, ufl=False)
 
-        self.update_a_v_u(a_old, v_old, u_old, a_vec, v_vec, u_vec)
+        self.update_a_v_u(a_old, v_old, u_old, a_vec, v_vec, u)
 
 
     def update_fields_ost(self, u, u_old, v_old, a_old):
-        # update fields at the end of each time step
-        # get vectors (references)
-        u_vec, u0_vec  = u.vector, u_old.vector
-        v0_vec, a0_vec = v_old.vector, a_old.vector
-        u_vec.assemble(), u0_vec.assemble(), v0_vec.assemble(), a0_vec.assemble()
 
         # use update functions using vector arguments
-        a_vec = self.update_a_ost(u_vec, u0_vec, v0_vec, a0_vec, ufl=False)
-        v_vec = self.update_v_ost(u_vec, u0_vec, v0_vec, a0_vec, ufl=False)
+        a_vec = self.update_a_ost(u.vector, u_old.vector, v_old.vector, a_old.vector, ufl=False)
+        v_vec = self.update_v_ost(u.vector, u_old.vector, v_old.vector, a_old.vector, ufl=False)
 
-        self.update_a_v_u(a_old, v_old, u_old, a_vec, v_vec, u_vec)
+        self.update_a_v_u(a_old, v_old, u_old, a_vec, v_vec, u)
 
 
-    def update_a_v_u(self, a_old, v_old, u_old, a, v, u):
+    def update_a_v_u(self, a_old, v_old, u_old, a_vec, v_vec, u):
 
         # update acceleration: a_old <- a
-        a_old.vector.axpby(1.0, 0.0, a)
+        a_old.vector.axpby(1.0, 0.0, a_vec)
         a_old.vector.ghostUpdate(addv=PETSc.InsertMode.INSERT, mode=PETSc.ScatterMode.FORWARD)
 
         # update velocity: v_old <- v
-        v_old.vector.axpby(1.0, 0.0, v)
+        v_old.vector.axpby(1.0, 0.0, v_vec)
         v_old.vector.ghostUpdate(addv=PETSc.InsertMode.INSERT, mode=PETSc.ScatterMode.FORWARD)
 
         # update displacement: u_old <- u
-        u_old.vector.axpby(1.0, 0.0, u)
+        u_old.vector.axpby(1.0, 0.0, u.vector)
         u_old.vector.ghostUpdate(addv=PETSc.InsertMode.INSERT, mode=PETSc.ScatterMode.FORWARD)
 
-        a.destroy(), v.destroy() # auxiliary vectors that have been created, so destroy
+        a_vec.destroy(), v_vec.destroy() # auxiliary vectors that have been created, so destroy
 
 
     def compute_genalpha_params(self, rho_inf): # cf. Chung and Hulbert (1993)
[…]
