+++ src/ambit/timeintegration.py
@@ -170,13 +170,13 @@ class timeintegration_solid(timeintegration):
         return timefac_m, timefac
 
 
-    def update_timestep(self, u, u_old, v_old, a_old, p, p_old, internalvars, internalvars_old):
+    def update_timestep(self, u, u_old, v, v_old, a, a_old, p, p_old, internalvars, internalvars_old):
 
         # now update old kinematic fields with new quantities
         if self.timint == 'genalpha':
-            self.update_fields_newmark(u, u_old, v_old, a_old)
+            self.update_fields_newmark(u, u_old, v, v_old, a, a_old)
         if self.timint == 'ost':
-            self.update_fields_ost(u, u_old, v_old, a_old)
+            self.update_fields_ost(u, u_old, v, v_old, a, a_old)
 
         # update pressure variable
         if self.incompressible_2field:
@@ -192,7 +192,7 @@ class timeintegration_solid(timeintegration):
         self.update_time_funcs()
 
 
-    def update_a_newmark(self, u, u_old, v_old, a_old, ufl=True):
+    def update_a_newmark(self, u, u_old, v_old, a_old, aout=None, ufl=True):
         # update formula for acceleration
         if ufl: # ufl form
             dt_ = self.dt
@@ -202,17 +202,13 @@ class timeintegration_solid(timeintegration):
             dt_ = float(self.dt)
             beta_ = float(self.beta)
 
-            aout = a_old.duplicate()
-
             aout.axpby(-(1.-2.*beta_)/(2.*beta_), 0.0, a_old)
             aout.axpy(-1./(beta_*dt_), v_old)
             aout.axpy(1./(beta_*dt_*dt_), u)
             aout.axpy(-1./(beta_*dt_*dt_), u_old)
 
-            return aout
 
-
-    def update_a_ost(self, u, u_old, v_old, a_old, ufl=True):
+    def update_a_ost(self, u, u_old, v_old, a_old, aout=None, ufl=True):
         # update formula for acceleration
         if ufl: # ufl form
             dt_ = self.dt
@@ -222,17 +218,13 @@ class timeintegration_solid(timeintegration):
             dt_ = float(self.dt)
             theta_ = float(self.theta_ost)
 
-            aout = a_old.duplicate()
-
             aout.axpby(-(1.-theta_)/theta_, 0.0, a_old)
             aout.axpy(-1./(theta_*theta_*dt_), v_old)
             aout.axpy(1./(theta_*theta_*dt_*dt_), u)
             aout.axpy(-1./(theta_*theta_*dt_*dt_), u_old)
 
-            return aout
 
-
-    def update_v_newmark(self, u, u_old, v_old, a_old, ufl=True):
+    def update_v_newmark(self, u, u_old, v_old, a_old, vout=None, ufl=True):
         # update formula for velocity
         if ufl: # ufl form
             dt_ = self.dt
@@ -244,17 +236,13 @@ class timeintegration_solid(timeintegration):
             gamma_ = float(self.gamma)
             beta_ = float(self.beta)
 
-            vout = v_old.duplicate()
-
             vout.axpby(-(gamma_-beta_)/beta_, 0.0, v_old)
             vout.axpy(-(gamma_-2.*beta_)/(2.*beta_) * dt_, a_old)
             vout.axpy(gamma_/(beta_*dt_), u)
             vout.axpy(-gamma_/(beta_*dt_), u_old)
 
-            return vout
 
-
-    def update_v_ost(self, u, u_old, v_old, a_old, ufl=True):
+    def update_v_ost(self, u, u_old, v_old, a_old, vout=None, ufl=True):
[…]
