+++ src/ambit/solid/solid.py
@@ -159,6 +159,9 @@ class SolidmechanicsProblem(problem_base):
         self.u     = fem.Function(self.V_u, name="Displacement")
         self.p     = fem.Function(self.V_p, name="Pressure")
 
+        # auxiliary velocity and acceleration vectors
+        self.v     = fem.Function(self.V_u, name="Velocity")
+        self.a     = fem.Function(self.V_u, name="Acceleration")
         # values of previous time step
         self.u_old = fem.Function(self.V_u)
         self.v_old = fem.Function(self.V_u)
@@ -949,7 +952,7 @@ class SolidmechanicsProblem(problem_base):
     def update(self):
 
         # update - displacement, velocity, acceleration, pressure, all internal variables, all time functions
-        self.ti.update_timestep(self.u, self.u_old, self.v_old, self.a_old, self.p, self.p_old, self.internalvars, self.internalvars_old)
+        self.ti.update_timestep(self.u, self.u_old, self.v, self.v_old, self.a, self.a_old, self.p, self.p_old, self.internalvars, self.internalvars_old)
 
 
     def print_to_screen(self):
