### 2023-08-30 marchirschvogel/ambit#2 @marchirschvogel
Thanks for pointing to this. I now moved the code to the MIT license.
