+++ src/ambit/flow0d/flow0d.py
@@ -1,485 +1,0 @@
-#!/usr/bin/env python3
-
-# Copyright (c) 2019-2023, Dr.-Ing. Marc Hirschvogel
-# All rights reserved.
-
-# This source code is licensed under the BSD-style license found in the
-# LICENSE file in the root directory of this source tree.
-
-import sys
-import numpy as np
-
-from petsc4py import PETSc
-import timeintegration
-import solver_nonlin
-import ioparams
-
-from base import problem_base, solver_base
-
-# framework of 0D flow models, relating pressure p (and its derivative) to fluxes q
-
-class Flow0DProblem(problem_base):
-
-    def __init__(self, io_params, time_params, model_params, time_curves, coupling_params={}, comm=None):
-        super().__init__(io_params, time_params, comm)
-
-        ioparams.check_params_time_flow0d(time_params)
-
-        self.problem_physics = 'flow0d'
-
-        self.time_params = time_params
-
-        # only relevant to syspul* models
-        try:
-            self.chamber_models = model_params['chamber_models']
-            if 'ao' not in self.chamber_models.keys(): self.chamber_models['ao'] = {'type' : '0D_rigid'} # add aortic root model
-        except: self.chamber_models = {}
-
-        try: self.coronary_model = model_params['coronary_model']
-        except: self.coronary_model = None
-
-        try: self.vad_model = model_params['vad_model']
-        except: self.vad_model = None
-
-        try: self.excitation_curve = model_params['excitation_curve']
-        except: self.excitation_curve = None
-
-        try: initial_file = time_params['initial_file']
-        except: initial_file = ''
-
-        # could use extra write frequency setting for 0D model (i.e. for coupled problem)
-        try: self.write_results_every_0D = io_params['write_results_every_0D']
-        except: self.write_results_every_0D = io_params['write_results_every']
-
-        # for restart
-        try: self.write_restart_every = io_params['write_restart_every']
-        except: self.write_restart_every = -1
-
-        # could use extra output path setting for 0D model (i.e. for coupled problem)
-        try: self.output_path_0D = io_params['output_path_0D']
-        except: self.output_path_0D = io_params['output_path']
-
-        # whether to output midpoint (t_{n+theta}) of state variables or endpoint (t_{n+1}) - for post-processing
-        try: self.output_midpoint = io_params['output_midpoint_0D']
-        except: self.output_midpoint = False
-
-        try: valvelaws = model_params['valvelaws']
-        except: valvelaws = {'av' : ['pwlin_pres',0], 'mv' : ['pwlin_pres',0], 'pv' : ['pwlin_pres',0], 'tv' : ['pwlin_pres',0]}
-
-        try: self.cq = coupling_params['coupling_quantity']
-        except: self.cq = ['volume']*5
-
-        try: self.vq = coupling_params['variable_quantity']
-        except: self.vq = ['pressure']*5
-
-        try: self.eps_periodic = time_params['eps_periodic']
-        except: self.eps_periodic = 1.0e-20
-
-        try: self.periodic_checktype = time_params['periodic_checktype']
-        except: self.periodic_checktype = ['allvar']
-
-        try: self.prescribed_variables = model_params['prescribed_variables']
-        except: self.prescribed_variables = {}
-
-        try: self.perturb_type = model_params['perturb_type'][0]
-        except: self.perturb_type = None
-
-        try: self.perturb_factor = model_params['perturb_type'][1]
[…]
