+++ src/ambit/coupling/fluid_flow0d.py
@@ -1,618 +1,0 @@
-#!/usr/bin/env python3
-
-# Copyright (c) 2019-2023, Dr.-Ing. Marc Hirschvogel
-# All rights reserved.
-
-# This source code is licensed under the BSD-style license found in the
-# LICENSE file in the root directory of this source tree.
-
-import time, sys, copy
-import numpy as np
-from dolfinx import fem
-import dolfinx.fem.petsc
-import ufl
-from petsc4py import PETSc
-
-import solver_nonlin
-import expression, ioparams
-from mpiroutines import allgather_vec
-
-from fluid import FluidmechanicsProblem
-from flow0d import Flow0DProblem
-from base import problem_base, solver_base
-
-
-class FluidmechanicsFlow0DProblem(problem_base):
-
-    def __init__(self, io_params, time_params_fluid, time_params_flow0d, fem_params, constitutive_models, model_params_flow0d, bc_dict, time_curves, coupling_params, io, mor_params={}, comm=None, alevar={}):
-        super().__init__(io_params, time_params_fluid, comm)
-
-        self.problem_physics = 'fluid_flow0d'
-
-        self.coupling_params = coupling_params
-
-        self.surface_vq_ids = self.coupling_params['surface_ids']
-        try: self.surface_p_ids = self.coupling_params['surface_p_ids']
-        except: self.surface_p_ids = self.surface_vq_ids
-
-        self.num_coupling_surf = len(self.surface_vq_ids)
-
-        try: self.cq_factor = self.coupling_params['cq_factor']
-        except: self.cq_factor = [1.]*self.num_coupling_surf
-
-        try: self.eps_fd = self.coupling_params['eps_fd']
-        except: self.eps_fd = 1.0e-5
-
-        try: self.print_subiter = self.coupling_params['print_subiter']
-        except: self.print_subiter = False
-
-        try: self.restart_periodicref = self.coupling_params['restart_periodicref']
-        except: self.restart_periodicref = 0
-
-        try: self.Nmax_periodicref = self.coupling_params['Nmax_periodicref']
-        except: self.Nmax_periodicref = 10
-
-        # only option in fluid mechanics!
-        self.coupling_type = 'monolithic_lagrange'
-
-        # assert that we do not have conflicting timings
-        time_params_flow0d['maxtime'] = time_params_fluid['maxtime']
-        time_params_flow0d['numstep'] = time_params_fluid['numstep']
-
-        # initialize problem instances (also sets the variational forms for the fluid problem)
-        self.pbf = FluidmechanicsProblem(io_params, time_params_fluid, fem_params, constitutive_models, bc_dict, time_curves, io, mor_params=mor_params, comm=self.comm, alevar=alevar)
-        self.pb0 = Flow0DProblem(io_params, time_params_flow0d, model_params_flow0d, time_curves, coupling_params, comm=self.comm)
-
-        self.pbrom = self.pbf # ROM problem can only be fluid
-
-        # indicator for no periodic reference state estimation
-        self.noperiodicref = 1
-
-        self.set_variational_forms()
-
-        self.numdof = self.pbf.numdof + self.lm.getSize()
-
-        self.localsolve = self.pbf.localsolve
-
-        self.sub_solve = True
-        self.print_enhanced_info = self.pbf.io.print_enhanced_info
-
-        # 3D fluxes
[…]

[…]
