### 2023-08-30 58b9068
update install instructions
