+++ README.md
@@ -39,38 +39,48 @@ Still experimental / to-do:
 - ... whatever might be wanted in some future ...
 
 
-### How do I get set up? ###
+### Install ###
 
-* Clone the repo:
+* First, you need to [install FEniCSx](https://github.com/FEniCS/dolfinx#installation), either from source or using latest tested Ambit-compatible digest (19 Aug 2023) of dolfinx Docker image:
+```
+docker pull dolfinx/dolfinx@sha256:1f374e90d5e918a71a4bdba994bf434cdaf84fedc47aa11ac33295864180fb76
+```
 
-``git clone https://github.com/marchirschvogel/ambit.git``
+* Ambit can then be installed using pip:
+```
+python3 -m pip install ambit-fe
+```
 
-* Either, FEniCS-x from a Docker container can be used, or it can be installed from source (see https://github.com/FEniCS/dolfinx if needed)
+* Alternatively, clone the repository
+```
+git clone https://github.com/marchirschvogel/ambit.git
+```
 
-* Assuming Docker is installed (if not, see e.g. https://docs.docker.com/engine/security/rootless), get latest tested Ambit-compatible digest (19 Aug 2023) of dolfinx Docker image:
-
-``docker pull dolfinx/dolfinx@sha256:1f374e90d5e918a71a4bdba994bf434cdaf84fedc47aa11ac33295864180fb76``
-
-* Put the following shortcut in .bashrc (replacing <PATH_TO_AMBIT_FOLDER> with the path to the ambit folder):
-
-``alias fenicsdocker='docker run -ti -v $HOME:/home/shared -v <PATH_TO_AMBIT_FOLDER>:/home/shared/ambit -w /home/shared/ --env-file <PATH_TO_AMBIT_FOLDER>/.env.list --rm dolfinx/dolfinx@sha256:1f374e90d5e918a71a4bdba994bf434cdaf84fedc47aa11ac33295864180fb76'``
+* and add the following shortcut to your .bashrc (replacing <PATH_TO_AMBIT_FOLDER> with the path to the ambit folder):
+```
+alias fenicsdocker='docker run -ti -v $HOME:/home/shared -v <PATH_TO_AMBIT_FOLDER>:/home/shared/ambit -w /home/shared/ --env-file <PATH_TO_AMBIT_FOLDER>/.env.list --rm dolfinx/dolfinx@sha256:1f374e90d5e918a71a4bdba994bf434cdaf84fedc47aa11ac33295864180fb76'
+```
 
 * If 0D models should be used, it seems that we have to install sympy (not part of docker container anymore) - in the folder where you pulled ambit to, do:
-
-``cd ambit && mkdir ext && pip3 install --target=ext mpmath --no-deps --no-cache-dir && pip3 install --target=ext sympy --no-deps --no-cache-dir && cd ..``
+```
+cd ambit && mkdir ext && pip3 install --target=ext mpmath --no-deps --no-cache-dir && pip3 install --target=ext sympy --no-deps --no-cache-dir && cd ..
+```
 
 (if this fails, add ``--system`` to each of the pip3 install commands)
 
 * Launch the container in a konsole/terminal window by simply typing
-
-``fenicsdocker``
+```
+fenicsdocker
+```
 
 * Have a look at example input files in ambit/tests and the file ambit_template.py in the main folder as example of all available input options
 
 * Best, check if all testcases run and pass, by navigating to ambit/tests and executing
-
-``./runtests.py``
+```
+./runtests.py
+```
 
 * Build your input file and run it with the command
-
-``mpiexec -n <NUMBER_OF_CORES> python3 your_file.py``
+```
+mpiexec -n <NUMBER_OF_CORES> python3 your_file.py
+```
