+++ src/ambit_fe/__init__.py
@@ -5,3 +5,5 @@
 
 # This source code is licensed under the MIT-style license found in the
 # LICENSE file in the root directory of this source tree.
+
+from . import ambit
