+++ docker/Dockerfile
@@ -2,6 +2,7 @@
 FROM ghcr.io/fenics/dolfinx/dolfinx@sha256:1f374e90d5e918a71a4bdba994bf434cdaf84fedc47aa11ac33295864180fb76
 
 ENV DEB_PYTHON_INSTALL_LAYOUT=deb_system
+ENV PYTHONPATH "${PYTHONPATH}"
 
 # install pip and git with apt
 RUN apt-get update && \
@@ -15,7 +16,7 @@ WORKDIR /tmp
 
 # install Python module
 RUN git clone https://github.com/marchirschvogel/ambit.git
-RUN python3 -m pip install --no-cache-dir ./ambit-fe[all]
+RUN python3 -m pip install --no-cache-dir ambit-fe
 
 # remove contents of the temporary directory to minimize the size of the image
 RUN rm -rf /tmp
