+++ pyproject.toml
@@ -2,8 +2,8 @@
 requires = ["setuptools>=61.0.0", "wheel"]
 
 [project]
-name = "ambit-fe"
-version = "1.0.0"
+name = "ambit_fe"
+version = "1.0.2"
 description = "A FEniCS-based cardiovascular physics solver"
 authors = [{name = "Marc Hirschvogel", email = "marc.hirschvogel@deepambit.com"}]
 license = {file = "LICENSE"}
