### 2023-08-30 630d53e
Updates such that Ambit can be installed via Python package index
