+++ tests/solid_plasticity_vonmises.py
@@ -6,13 +6,12 @@ tests:
 TODO:  NOT YET FULLY IMPLEMENTED!
 """
 
-import ambit
+import ambit_fe
 
 import sys, traceback
 import numpy as np
 from pathlib import Path
 
-import resultcheck
 
 
 def main():
@@ -59,7 +58,7 @@ def main():
 
 
     # problem setup
-    problem = ambit.Ambit(IO_PARAMS, TIME_PARAMS, SOLVER_PARAMS, FEM_PARAMS, MATERIALS, BC_DICT, time_curves=time_curves())
+    problem = ambit_fe.ambit_main.Ambit(IO_PARAMS, TIME_PARAMS, SOLVER_PARAMS, FEM_PARAMS, MATERIALS, BC_DICT, time_curves=time_curves())
 
     # solve time-dependent problem
     problem.solve_problem()
@@ -78,8 +77,8 @@ def main():
     #u_corr[1] = -1.0862313365225019e-07 # y
     #u_corr[2] = -0.000897803340365617 # z
 
-    #check1 = resultcheck.results_check_node(problem.mp.u, check_node, u_corr, problem.mp.V_u, problem.mp.comm, tol=tol, nm='u')
-    #success = resultcheck.success_check([check1], problem.mp.comm)
+    #check1 = ambit_fe.resultcheck.results_check_node(problem.mp.u, check_node, u_corr, problem.mp.V_u, problem.mp.comm, tol=tol, nm='u')
+    #success = ambit_fe.resultcheck.success_check([check1], problem.mp.comm)
 
     #return success
 

[…]
