### 2023-08-30 70630a4
Add further install info to README
