+++ tests/fluid_p1p1_stab_cylinder_valve.py
@@ -17,7 +17,6 @@ import numpy as np
 from pathlib import Path
 
 
-
 def main():
     
     basepath = str(Path(__file__).parent.absolute())
