+++ tests/solid_mat_uniax_hex_2field.py
@@ -18,7 +18,6 @@ import numpy as np
 from pathlib import Path
 
 
-
 def main():
 
     basepath = str(Path(__file__).parent.absolute())
