+++ tests/flow0d_0Dheart_syspulcorvad.py
@@ -11,8 +11,6 @@ import numpy as np
 from pathlib import Path
 
 
-
-
 def main():
 
     basepath = str(Path(__file__).parent.absolute())
