+++ tests/solid_flow0d_monolagr_CRLinoutlink_chambers.py
@@ -12,7 +12,6 @@ import numpy as np
 from pathlib import Path
 
 
-
 def main():
 
     basepath = str(Path(__file__).parent.absolute())
