+++ tests/solid_membrane.py
@@ -11,7 +11,6 @@ import numpy as np
 from pathlib import Path
 
 
-
 def main():
 
     basepath = str(Path(__file__).parent.absolute())
