+++ tests/flow0d_0Dheart_syspulcor.py
@@ -10,7 +10,6 @@ import numpy as np
 from pathlib import Path
 
 
-
 def main():
 
     basepath = str(Path(__file__).parent.absolute())
