+++ tests/solid_flow0d_monodir2field_4elwindkesselLpZ_chamber.py
@@ -13,7 +13,6 @@ import numpy as np
 from pathlib import Path
 
 
-
 def main():
 
     basepath = str(Path(__file__).parent.absolute())
