+++ tests/solid_robin_genalpha_amg.py
@@ -16,7 +16,6 @@ import numpy as np
 from pathlib import Path
 
 
-
 def main():
 
     basepath = str(Path(__file__).parent.absolute())
