+++ tests/fluid_ale_flow0d_lalv_syspul_prescribed.py
@@ -18,7 +18,6 @@ import numpy as np
 from pathlib import Path
 
 
-
 def main():
     
     basepath = str(Path(__file__).parent.absolute())
