### 2023-08-31 marchirschvogel/ambit#4 @finsberg
Add automated tests
There are a lot of tests in the repository and there should be a process for running these tests whenever someone pushes to the `main` (or `master` in this case) branch.

Please add a GitHub action (or using another provider) for running these tests.

This is related to https://github.com/openjournals/joss-reviews/issues/5744
