### 2023-08-31 d5930be
Add pre-built Docker container and update install instructions
