### 2023-08-31 25606dc
Add mixed Docker image
