+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=61.0.0", "wheel"]
 
 [project]
 name = "ambit-fe"
-version = "1.0.4"
+version = "1.0.5"
 description = "A FEniCS-based cardiovascular physics solver"
 authors = [{name = "Marc Hirschvogel", email = "marc.hirschvogel@deepambit.com"}]
 license = {file = "LICENSE"}
