### 2023-09-08 963c1fc
some changes and optimizations in solver
