+++ src/ambit_fe/solver/preconditioner.py
@@ -49,31 +49,30 @@ class block_precond():
             self.ksp_fields.append( PETSc.KSP().create(self.comm) )
         # set the options
         for n in range(self.nfields):
-            pc_field = self.ksp_fields[n].getPC()
             if self.precond_fields[n]['prec'] == 'amg':
                 try: solvetype = self.precond_fields[n]['solve']
                 except: solvetype = "preonly"
                 self.ksp_fields[n].setType(solvetype)
                 try: amgtype = self.precond_fields[n]['amgtype']
                 except: amgtype = "hypre"
-                pc_field.setType(amgtype)
+                self.ksp_fields[n].getPC().setType(amgtype)
                 if amgtype=="hypre":
-                    pc_field.setHYPREType("boomeramg")
+                    self.ksp_fields[n].getPC().setHYPREType("boomeramg")
                 # set operators and setup field prec
-                pc_field.setOperators(operator_mats[n])
-                pc_field.setUp()
+                self.ksp_fields[n].getPC().setOperators(operator_mats[n])
+                self.ksp_fields[n].getPC().setUp()
                 # TODO: Some additional hypre options we might wanna set... which are optimal here???
                 # opts = PETSc.Options()
                 # opts.setValue('pc_hypre_parasails_reuse', True) # - does this exist???
                 # opts.setValue('pc_hypre_boomeramg_cycle_type', 'v') # v, w
                 # opts.setValue('pc_hypre_boomeramg_max_iter', 1)
                 # opts.setValue('pc_hypre_boomeramg_relax_type_all',  'symmetric-SOR/Jacobi')
-                # pc_field.setFromOptions()
-                # print(pc_fieldself.ksp_fields[n].getPC().view())
+                # self.ksp_fields[n].getPC().setFromOptions()
+                # print(self.ksp_fields[n].getPC().view())
             elif self.precond_fields[n]['prec'] == 'direct':
                 self.ksp_fields[n].setType("preonly")
-                pc_field.setType("lu")
-                pc_field.setFactorSolverType("mumps")
+                self.ksp_fields[n].getPC().setType("lu")
+                self.ksp_fields[n].getPC().setFactorSolverType("mumps")
             else:
                 raise ValueError("Currently, only either 'amg' or 'direct' are supported as field-specific preconditioner.")
 
@@ -104,7 +103,12 @@ class schur_2x2(block_precond):
         self.B  = self.P.createSubMatrix(self.iset[1],self.iset[0])
         self.C  = self.P.createSubMatrix(self.iset[1],self.iset[1])
 
-        self.Adinv = self.A.duplicate(copy=False)
+        # the matrix to later insert the diagonal
+        self.Adinv = PETSc.Mat().createAIJ(self.A.getSizes(), bsize=None, nnz=(1,1), csr=None, comm=self.comm)
+        self.Adinv.setUp()
+        self.Adinv.assemble()
+        # set 1's to get correct allocation pattern
+        self.Adinv.shift(1.)
 
[…]
