+++ src/ambit_fe/__init__.py
@@ -6,7 +6,5 @@
 # This source code is licensed under the MIT-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-# from __future__ import absolute_import
-
 from . import ambit_main
 from . import resultcheck
