+++ src/ambit_fe/ioroutines.py
@@ -312,7 +312,7 @@ class IO:
 
             # project to fiber function space
             if self.order_fib_input != order_disp:
-                fib_func.append( project(fib_func_input[si], V_fib, dx_, nm='Fiber'+str(si+1)) )
+                fib_func.append( project(fib_func_input[si], V_fib, dx_, nm='Fiber'+str(si+1), comm=self.comm) )
             else:
                 fib_func.append( fib_func_input[si] )
 
@@ -368,12 +368,12 @@ class IO_solid(IO):
                         u_out.interpolate(pb.u)
                         self.resultsfiles[res].write_function(u_out, indicator)
                     elif res=='velocity': # passed in v is not a function but form, so we have to project
-                        self.v_proj = project(pb.vel, pb.V_u, pb.dx_, nm="Velocity") # class variable for testing
+                        self.v_proj = project(pb.vel, pb.V_u, pb.dx_, nm="Velocity", comm=self.comm) # class variable for testing
                         v_out = fem.Function(self.V_out_vector, name=self.v_proj.name)
                         v_out.interpolate(self.v_proj)
                         self.resultsfiles[res].write_function(v_out, indicator)
                     elif res=='acceleration': # passed in a is not a function but form, so we have to project
-                        self.a_proj = project(pb.acc, pb.V_u, pb.dx_, nm="Acceleration") # class variable for testing
+                        self.a_proj = project(pb.acc, pb.V_u, pb.dx_, nm="Acceleration", comm=self.comm) # class variable for testing
                         a_out = fem.Function(self.V_out_vector, name=self.a_proj.name)
                         a_out.interpolate(self.a_proj)
                         self.resultsfiles[res].write_function(a_out, indicator)
@@ -385,7 +385,7 @@ class IO_solid(IO):
                         stressfuncs=[]
                         for n in range(pb.num_domains):
                             stressfuncs.append(pb.ma[n].sigma(pb.u,pb.p,pb.vel,ivar=pb.internalvars))
-                        cauchystress = project(stressfuncs, pb.Vd_tensor, pb.dx_, nm="CauchyStress")
+                        cauchystress = project(stressfuncs, pb.Vd_tensor, pb.dx_, nm="CauchyStress", comm=self.comm)
                         cauchystress_out = fem.Function(self.V_out_tensor, name=cauchystress.name)
                         cauchystress_out.interpolate(cauchystress)
                         self.resultsfiles[res].write_function(cauchystress_out, indicator)
@@ -393,7 +393,7 @@ class IO_solid(IO):
                         stressfuncs=[]
                         for n in range(pb.num_domains):
                             stressfuncs.append(pb.ma[n].sigma(pb.u,pb.p,pb.vel,ivar=pb.internalvars))
-                        cauchystress_nodal = project(stressfuncs, pb.V_tensor, pb.dx_, nm="CauchyStress_nodal")
+                        cauchystress_nodal = project(stressfuncs, pb.V_tensor, pb.dx_, nm="CauchyStress_nodal", comm=self.comm)
                         cauchystress_nodal_out = fem.Function(self.V_out_tensor, name=cauchystress_nodal.name)
                         cauchystress_nodal_out.interpolate(cauchystress_nodal)
                         self.resultsfiles[res].write_function(cauchystress_nodal_out, indicator)
@@ -402,7 +402,7 @@ class IO_solid(IO):
                         for n in range(pb.num_domains):
                             evals, _, _ = spectral_decomposition_3x3(pb.ma[n].sigma(pb.u,pb.p,pb.vel,ivar=pb.internalvars))
[…]
