+++ src/ambit_fe/multiscale/solid_flow0d_growthremodel_main.py
@@ -12,12 +12,13 @@ from dolfinx import fem
 import ufl
 from petsc4py import PETSc
 
-from . import utilities
-from ..solid import SolidmechanicsProblem, SolidmechanicsSolver
-from ..solid_flow0d import SolidmechanicsFlow0DProblem, SolidmechanicsFlow0DSolver
+from .. import utilities
+from ..solid.solid_main import SolidmechanicsProblem, SolidmechanicsSolver
+from ..coupling.solid_flow0d_main import SolidmechanicsFlow0DProblem, SolidmechanicsFlow0DSolver
+from ..base import problem_base, solver_base
 
 
-class SolidmechanicsFlow0DMultiscaleGrowthRemodelingProblem():
+class SolidmechanicsFlow0DMultiscaleGrowthRemodelingProblem(problem_base):
 
     def __init__(self, io_params, time_params_solid_small, time_params_solid_large, time_params_flow0d, fem_params, constitutive_models, model_params_flow0d, bc_dict, time_curves, coupling_params, multiscale_params, io, comm=None):
 
@@ -80,7 +81,7 @@ class SolidmechanicsFlow0DMultiscaleGrowthRemodelingProblem():
         self.tol_outer = multiscale_params['tol_outer']
 
         # override by tol_small
-        self.pbsmall.pbf.eps_periodic = self.tol_small
+        self.pbsmall.pb0.eps_periodic = self.tol_small
         self.pblarge.tol_stop_large = self.tol_large
 
         # store to ensure prestressed state is kept throughout the whole cycle (small scale prestress_initial gets set to False after initial prestress)
@@ -92,9 +93,9 @@ class SolidmechanicsFlow0DMultiscaleGrowthRemodelingProblem():
         self.simname_large = self.pblarge.simname + '_large'
 
         if gandr_trigger_phase == 'end_diastole':
-            self.pbsmall.t_gandr_setpoint = self.pbsmall.pbf.cardvasc0D.t_ed
+            self.pbsmall.t_gandr_setpoint = self.pbsmall.pb0.cardvasc0D.t_ed
         elif gandr_trigger_phase == 'end_systole':
-            self.pbsmall.t_gandr_setpoint = self.pbsmall.pbf.cardvasc0D.t_es
+            self.pbsmall.t_gandr_setpoint = self.pbsmall.pb0.cardvasc0D.t_es
         else:
             raise NameError("Unknown growth multiscale_trigger_phase")
 
@@ -125,15 +126,15 @@ class SolidmechanicsFlow0DMultiscaleGrowthRemodelingProblem():
 
 
 
-class SolidmechanicsFlow0DMultiscaleGrowthRemodelingSolver():
+class SolidmechanicsFlow0DMultiscaleGrowthRemodelingSolver(solver_base):
 
-    def __init__(self, problem, solver_params_solid, solver_params_flow0d):
+    def __init__(self, problem, solver_params):
 
         self.pb = problem
 
         # initialize solver instances
-        self.solversmall = SolidmechanicsFlow0DSolver(self.pb.pbsmall, solver_params_solid, solver_params_flow0d)
-        self.solverlarge = SolidmechanicsSolver(self.pb.pblarge, solver_params_solid)
+        self.solversmall = SolidmechanicsFlow0DSolver(self.pb.pbsmall, solver_params)
+        self.solverlarge = SolidmechanicsSolver(self.pb.pblarge, solver_params)
 
         # read restart information
         if self.pb.restart_cycle > 0:
[…]
