+++ tests/solid_flow0d_multiscalegrowthremodeling_concentric.py
@@ -20,7 +20,7 @@ def main():
     IO_PARAMS            = {'problem_type'          : 'solid_flow0d_multiscale_gandr',
                             'mesh_domain'           : basepath+'/input/heart3Dcoarse_domain.xdmf',
                             'mesh_boundary'         : basepath+'/input/heart3Dcoarse_boundary.xdmf',
-                            'fiber_data'            : [basepath+'/input/fib_fiber_coords_nodal_3Dcoarse.txt',basepath+'/input/fib_sheet_coords_nodal_3Dcoarse.txt'],
+                            'fiber_data'            : [basepath+'/input/fib_fiber_nodal_3Dcoarse.txt',basepath+'/input/fib_sheet_nodal_3Dcoarse.txt'],
                             'write_results_every'   : 1,
                             'output_path'           : basepath+'/tmp',
                             'results_to_write'      : ['displacement','theta','phi_remod','trmandelstress_e'],
