### 2023-09-18 0703753
ROM sanity check
