+++ src/ambit_fe/solid/solid_main.py
@@ -220,6 +220,9 @@ class SolidmechanicsProblem(problem_base):
 
         self.mor_params = mor_params
 
+        # sanity check
+        if bool(self.mor_params): assert(self.have_rom)
+
         # initialize solid time-integration class
         self.ti = timeintegration.timeintegration_solid(time_params, fem_params, time_curves, self.t_init, self.comm)
 
