+++ tests/flow0d_0Dheart_syspul.py
@@ -38,8 +38,14 @@ def main():
 
     MODEL_PARAMS      = {'modeltype'             : 'syspul',
                          'parameters'            : param(),
-                         'chamber_models'        : {'lv' : {'type' : '0D_elast', 'activation_curve' : 2}, 'rv' : {'type' : '0D_elast', 'activation_curve' : 2}, 'la' : {'type' : '0D_elast', 'activation_curve' : 1}, 'ra' : {'type' : '0D_elast', 'activation_curve' : 1}},
-                         'valvelaws'             : {'av' : ['smooth_pres_momentum',0], 'mv' : ['pwlin_pres'], 'pv' : ['pwlin_pres'], 'tv' : ['pwlin_pres']}}
+                         'chamber_models'        : {'lv' : {'type' : '0D_elast', 'activation_curve' : 2},
+                                                    'rv' : {'type' : '0D_elast', 'activation_curve' : 2},
+                                                    'la' : {'type' : '0D_elast', 'activation_curve' : 1},
+                                                    'ra' : {'type' : '0D_elast', 'activation_curve' : 1}},
+                         'valvelaws'             : {'av' : ['smooth_pres_momentum',0],
+                                                    'mv' : ['pwlin_pres'],
+                                                    'pv' : ['pwlin_pres'],
+                                                    'tv' : ['pwlin_pres']}}
 
 
     # define your time curves here (syntax: tcX refers to curve X)
