+++ tests/flow0d_0Dheart_syspulcorvad.py
@@ -41,8 +41,14 @@ def main():
                          'coronary_model'        : 'ZCRp_CRd_lr',
                          'vad_model'             : 'lvad',
                          'parameters'            : param(),
-                         'chamber_models'        : {'lv' : {'type' : '0D_elast', 'activation_curve' : 2}, 'rv' : {'type' : '0D_elast', 'activation_curve' : 2}, 'la' : {'type' : '0D_elast', 'activation_curve' : 1}, 'ra' : {'type' : '0D_elast', 'activation_curve' : 1}},
-                         'valvelaws'             : {'av' : ['smooth_pres_momentum',0], 'mv' : ['pwlin_pres'], 'pv' : ['pwlin_pres'], 'tv' : ['pwlin_pres']}}
+                         'chamber_models'        : {'lv' : {'type' : '0D_elast', 'activation_curve' : 2},
+                                                    'rv' : {'type' : '0D_elast', 'activation_curve' : 2},
+                                                    'la' : {'type' : '0D_elast', 'activation_curve' : 1},
+                                                    'ra' : {'type' : '0D_elast', 'activation_curve' : 1}},
+                         'valvelaws'             : {'av' : ['smooth_pres_momentum',0],
+                                                    'mv' : ['pwlin_pres'],
+                                                    'pv' : ['pwlin_pres'],
+                                                    'tv' : ['pwlin_pres']}}
 
 
     # define your time curves here (syntax: tcX refers to curve X)
