+++ src/ambit_fe/solid/solid_main.py
@@ -120,9 +120,10 @@ class SolidmechanicsProblem(problem_base):
 
         self.Vex = self.io.mesh.ufl_domain().ufl_coordinate_element()
 
-        # check if we want to use model order reduction and if yes, initialize MOR class
-        try: self.have_rom = io_params['use_model_order_red']
-        except: self.have_rom = False
+        # model order reduction
+        self.mor_params = mor_params
+        if bool(self.mor_params): self.have_rom = True
+        else: self.have_rom = False
 
         # create finite element objects for u and p
         P_u = ufl.VectorElement("CG", self.io.mesh.ufl_cell(), self.order_disp)
@@ -219,9 +220,6 @@ class SolidmechanicsProblem(problem_base):
             self.numdof = self.u.vector.getSize()
 
         self.mor_params = mor_params
-
-        # sanity check
-        if bool(self.mor_params): assert(self.have_rom)
 
         # initialize solid time-integration class
         self.ti = timeintegration.timeintegration_solid(time_params, fem_params, time_curves, self.t_init, self.comm)
