+++ src/ambit_fe/flow0d/cardiovascular0D_syspulcaprespir.py
@@ -794,49 +794,3 @@ class cardiovascular0Dsyspulcaprespir(cardiovascular0Dsyspulcap):
             sys.stdout.flush()
 
         if not isinstance(var, np.ndarray): del var_sq
-
-
-
-
-def postprocess_groups_syspulcaprespir(groups, coronarymodel=None, indpertaftercyl=0,multiscalegandr=False):
-
-    import cardiovascular0D_syspulcap
-
-    cardiovascular0D_syspulcap.postprocess_groups_syspulcap(groups,indpertaftercyl,multiscalegandr)
-
-    # index 14
-    groups.append({'ppO2_time_sys_l'  : ['ppO2_at_l', 'ppO2_v_l', 'ppO2_ar_sys', 'ppO2_arspl_sys', 'ppO2_arespl_sys', 'ppO2_armsc_sys', 'ppO2_arcer_sys', 'ppO2_arcor_sys', 'ppO2_venspl_sys', 'ppO2_venespl_sys', 'ppO2_venmsc_sys', 'ppO2_vencer_sys', 'ppO2_vencor_sys', 'ppO2_ven_sys'],
-                   'tex'              : ['$p_{\\\mathrm{O}_2,\\\mathrm{at}}^{\\\ell}$', '$p_{\\\mathrm{O}_2,\\\mathrm{v}}^{\\\ell}$', '$p_{\\\mathrm{O}_2,\\\mathrm{ar}}^{\\\mathrm{sys}}$', '$p_{\\\mathrm{O}_2,\\\mathrm{ar,spl}}^{\\\mathrm{sys}}$', '$p_{\\\mathrm{O}_2,\\\mathrm{ar,espl}}^{\\\mathrm{sys}}$', '$p_{\\\mathrm{O}_2,\\\mathrm{ar,msc}}^{\\\mathrm{sys}}$', '$p_{\\\mathrm{O}_2,\\\mathrm{ar,cer}}^{\\\mathrm{sys}}$', '$p_{\\\mathrm{O}_2,\\\mathrm{ar,cor}}^{\\\mathrm{sys}}$', '$p_{\\\mathrm{O}_2,\\\mathrm{ven,spl}}^{\\\mathrm{sys}}$', '$p_{\\\mathrm{O}_2,\\\mathrm{ven,espl}}^{\\\mathrm{sys}}$', '$p_{\\\mathrm{O}_2,\\\mathrm{ven,msc}}^{\\\mathrm{sys}}$', '$p_{\\\mathrm{O}_2,\\\mathrm{ven,cer}}^{\\\mathrm{sys}}$', '$p_{\\\mathrm{O}_2,\\\mathrm{ven,cor}}^{\\\mathrm{sys}}$', '$p_{\\\mathrm{O}_2,\\\mathrm{ven}}^{\\\mathrm{sys}}$'],
-                   'lines'            : [1, 2, 3, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15]})
-    # index 15
-    groups.append({'ppCO2_time_sys_l' : ['ppCO2_at_l', 'ppCO2_v_l', 'ppCO2_ar_sys', 'ppCO2_arspl_sys', 'ppCO2_arespl_sys', 'ppCO2_armsc_sys', 'ppCO2_arcer_sys', 'ppCO2_arcor_sys', 'ppCO2_venspl_sys', 'ppCO2_venespl_sys', 'ppCO2_venmsc_sys', 'ppCO2_vencer_sys', 'ppCO2_vencor_sys', 'ppCO2_ven_sys'],
[…]
