+++ src/ambit_fe/postprocess/flow0d_plot.py
@@ -51,39 +51,33 @@ def postprocess0D(path, sname, nstep_cycl, T_cycl, t_ed, t_es, model, coronarymo
 
     if model == 'syspul':
 
-        spec = importlib.util.spec_from_file_location('cardiovascular0D_syspul', str(fpath)+'/../flow0d/cardiovascular0D_syspul.py')
-        module_name = importlib.util.module_from_spec(spec)
-        spec.loader.exec_module(module_name)
-        module_name.postprocess_groups_syspul(groups,coronarymodel,indpertaftercyl,multiscalegandr)
+        postprocess_groups_syspul(groups,coronarymodel,indpertaftercyl,multiscalegandr)
         iscirculation = True
         calculate_function_params = calc_func_params
 
     elif model == 'syspulcap':
 
-        spec = importlib.util.spec_from_file_location('cardiovascular0D_syspulcap', str(fpath)+'/../flow0d/cardiovascular0D_syspulcap.py')
-        module_name = importlib.util.module_from_spec(spec)
-        spec.loader.exec_module(module_name)
-        module_name.postprocess_groups_syspulcap(groups,coronarymodel,indpertaftercyl,multiscalegandr)
+        postprocess_groups_syspulcap(groups,coronarymodel,indpertaftercyl,multiscalegandr)
         iscirculation = True
         calculate_function_params = calc_func_params
 
     elif model == 'syspulcapcor':
 
-        spec = importlib.util.spec_from_file_location('cardiovascular0D_syspulcap', str(fpath)+'/../flow0d/cardiovascular0D_syspulcap.py')
-        module_name = importlib.util.module_from_spec(spec)
-        spec.loader.exec_module(module_name)
-        module_name.postprocess_groups_syspulcapcor(groups,coronarymodel,indpertaftercyl,multiscalegandr)
+        postprocess_groups_syspulcapcor(groups,coronarymodel,indpertaftercyl,multiscalegandr)
         iscirculation = True
         calculate_function_params = calc_func_params
 
     elif model == 'syspulcaprespir':
 
-        spec = importlib.util.spec_from_file_location('cardiovascular0D_syspulcaprespir', str(fpath)+'/../flow0d/cardiovascular0D_syspulcaprespir.py')
-        module_name = importlib.util.module_from_spec(spec)
-        spec.loader.exec_module(module_name)
-        module_name.postprocess_groups_syspulcaprespir(groups,coronarymodel,indpertaftercyl,multiscalegandr)
+        postprocess_groups_syspulcaprespir(groups,coronarymodel,indpertaftercyl,multiscalegandr)
         iscirculation = True
         calculate_function_params = calc_func_params
+
+    elif model == '2elwindkessel':
+
+        # TODO: Should we implement this?
+        iscirculation = False
+        pass
 
     elif model == '4elwindkesselLsZ':
 
[…]

[…]
