### 2023-09-21 e10d1fa
Improve printing
