+++ src/ambit_fe/mor/mor_main.py
@@ -12,6 +12,7 @@ from dolfinx import fem, io
 from petsc4py import PETSc
 import numpy as np
 
+from .. import utilities
 from .. import ioparams
 from ..meshutils import gather_surface_dof_indices
 
@@ -160,18 +161,14 @@ class ModelOrderReduction():
 
         ts = time.time()
 
-        if self.pb.comm.rank==0:
-            print("Performing Proper Orthogonal Decomposition (POD) ...")
-            sys.stdout.flush()
+        utilities.print_status("Performing Proper Orthogonal Decomposition (POD)...", self.pb.comm)
 
         # gather snapshots (mostly displacements or velocities)
         for h in range(self.numhdms):
 
             for i in range(self.numsnapshots):
 
-                if self.pb.comm.rank==0:
-                    print("Reading snapshot %i ..." % (i+1))
-                    sys.stdout.flush()
+                utilities.print_status("Reading snapshot %i ..." % (i+1), self.pb.comm)
 
                 step = self.snapshotoffset + (i+1)*self.snapshotincr
 
@@ -221,9 +218,7 @@ class ModelOrderReduction():
         nconv = eigsolver.getConverged()
 
         if self.print_eigenproblem:
-            if self.pb.comm.rank==0:
-                print("Number of converged eigenpairs: %d" % nconv)
-                sys.stdout.flush()
+            utilities.print_status("Number of converged eigenpairs: %d" % (nconv), self.pb.comm)
 
         evecs, evals = [], []
         self.numredbasisvec_true = 0
@@ -234,23 +229,17 @@ class ModelOrderReduction():
             vi, _ = C_d.getVecs()
 
             if self.print_eigenproblem:
-                if self.pb.comm.rank==0:
-                    print("   k                        ||Ax-kx||/||kx||")
-                    print("   ----------------------   ----------------")
-                    sys.stdout.flush()
+                utilities.print_status("   k                        ||Ax-kx||/||kx||", self.pb.comm)
+                utilities.print_status("   ----------------------   ----------------", self.pb.comm)
 
             for i in range(len(self.redbasisvec_indices)):
                 k = eigsolver.getEigenpair(self.redbasisvec_indices[i], vr, vi)
                 error = eigsolver.computeError(self.redbasisvec_indices[i])
                 if self.print_eigenproblem:
                     if k.imag != 0.0:
-                        if self.pb.comm.rank==0:
-                            print('{:<3s}{:<4.4e}{:<1s}{:<4.4e}{:<1s}{:<3s}{:<4.4e}'.format(' ',k.real,'+',k.imag,'j',' ',error))
-                            sys.stdout.flush()
+                        utilities.print_status("{:<3s}{:<4.4e}{:<1s}{:<4.4e}{:<1s}{:<3s}{:<4.4e}".format(" ",k.real,"+",k.imag,"j"," ",error), self.pb.comm)
                     else:
-                        if self.pb.comm.rank==0:
-                            print('{:<3s}{:<4.4e}{:<15s}{:<4.4e}'.format(' ',k.real,' ',error))
-                            sys.stdout.flush()
+                        utilities.print_status("{:<3s}{:<4.4e}{:<15s}{:<4.4e}".format(" ",k.real," ",error), self.pb.comm)
 
                 # store
                 evecs.append(copy.deepcopy(vr)) # need copy here, otherwise reference changes
@@ -259,9 +248,7 @@ class ModelOrderReduction():
                 if k.real > self.eigenvalue_cutoff: self.numredbasisvec_true += 1
 
         if len(self.redbasisvec_indices) != self.numredbasisvec_true:
[…]
