### 2023-09-22 0216527
revert to old nnz
