### 2023-09-24 a97118b
Optimizations for 3D-0D offdiagonal block assembly; clean-up printouts
