### 2023-09-25 ee5162f
Slight improvements for offdiag 3D0D block assembly
