### 2023-09-27 260cb24
delete unneeded lines in solver loops
