### 2023-10-03 #5744 @Kevin-Mattheus-Moerman
@finsberg, @IgorBaratta, @CZHU20 thanks again for your help in reviewing this submission for JOSS! Would you be able to provide an update on how you are getting on? Perhaps you could summerise the key points/issues that require the author's attention? Thanks!
