### 2023-10-03 marchirschvogel/ambit#5 @finsberg
Documentation
I think the documentation that you have written in https://github.com/marchirschvogel/ambit/blob/master/doc/ambit_doc.pdf is really good, and I think it deserves to be added as a part on an online documentation hosted on e.g GitHub pages or [Read The Docs](https://about.readthedocs.com/?ref=readthedocs.com). This is fairly low effort and makes the user experience drastically better. For example you can convert the latex file to markdown using [pandoc](https://pandoc.org) and use some tool to turn this into an online documentation (you can read about some of the tools [here](https://realpython.com/documenting-python-code/#documentation-tools-and-resources))

As a part of this I also think you should also add API documentation and make sure to also add docstrings to the APIs that a user would need to use.

This is related to https://github.com/openjournals/joss-reviews/issues/5744
