### 2023-10-03 #5744 @finsberg
@Kevin-Mattheus-Moerman I am still awaiting feedback on https://github.com/marchirschvogel/ambit/issues/4. I also just submitted a new issue [here](https://github.com/marchirschvogel/ambit/issues/5). I would say https://github.com/marchirschvogel/ambit/issues/1 has been satisfactory addressed.

In my opinion there are still a few things that needs to be addressed before this is ready for JOSS.
