### 2023-10-08 b8a629e
enhanced options for Schur block scaling
