### 2023-10-10 4a62d43
Updates to boundary condition code: reduce code redundancies
