+++ src/ambit_fe/fluid/fluid_main.py
@@ -281,7 +281,7 @@ class FluidmechanicsProblem(problem_base):
         if not bool(self.alevar):
             # standard Eulerian fluid
             self.alevar = {'Fale' : None, 'Fale_old' : None, 'w' : None, 'w_old' : None, 'fluid_on_deformed' : 'no'}
-            self.vf = fluid_variationalform.variationalform(self.var_v, self.var_p_, self.io.n0, formulation=self.fluid_formulation)
+            self.vf = fluid_variationalform.variationalform(self.var_v, var_p=self.var_p_, n0=self.io.n0, formulation=self.fluid_formulation)
 
         else:
             # mid-point representation of ALE velocity
@@ -291,18 +291,18 @@ class FluidmechanicsProblem(problem_base):
 
             if self.alevar['fluid_on_deformed'] == 'consistent':
                 # fully consistent ALE formulation of Navier-Stokes
-                self.vf = fluid_variationalform.variationalform_ale(self.var_v, self.var_p_, self.io.n0, formulation=self.fluid_formulation)
+                self.vf = fluid_variationalform.variationalform_ale(self.var_v, var_p=self.var_p_, n0=self.io.n0, formulation=self.fluid_formulation)
 
             elif self.alevar['fluid_on_deformed'] == 'from_last_step':
                 # ALE formulation of Navier-Stokes using metrics (Fale, w) from the last converged step... more efficient but not fully consistent
                 self.alevar['Fale'], self.alevar['w'] = self.alevar['Fale_old'], self.alevar['w_old']
-                self.vf = fluid_variationalform.variationalform_ale(self.var_v, self.var_p_, self.io.n0, formulation=self.fluid_formulation)
+                self.vf = fluid_variationalform.variationalform_ale(self.var_v, var_p=self.var_p_, n0=self.io.n0, formulation=self.fluid_formulation)
 
             elif self.alevar['fluid_on_deformed'] == 'mesh_move':
                 # Navier-Stokes formulated w.r.t. the current, moved frame... more efficient than 'consistent' approach but not fully consistent
                 # WARNING: This is unsuitable for FrSI, as we need gradients w.r.t. the reference frame on the reduced boundary!
                 self.alevar = {'Fale' : None, 'Fale_old' : None, 'w' : None, 'w_old' : None, 'fluid_on_deformed' : 'mesh_move'}
-                self.vf = fluid_variationalform.variationalform(self.var_v, self.var_p_, self.io.n0, formulation=self.fluid_formulation)
+                self.vf = fluid_variationalform.variationalform(self.var_v, var_p=self.var_p_, n0=self.io.n0, formulation=self.fluid_formulation)
 
             else:
                 raise ValueError("Unknown fluid_on_deformed option!")
@@ -381,59 +381,59 @@ class FluidmechanicsProblem(problem_base):
 
             # kinetic virtual power
             if self.fluid_governing_type=='navierstokes_transient':
-                self.deltaW_kin     += self.vf.deltaW_kin_navierstokes_transient(self.acc, self.v, self.rho[n], self.dx_[n], w=self.alevar['w'], Fale=self.alevar['Fale'])
[…]
