+++ src/ambit_fe/fluid/fluid_variationalform.py
@@ -15,18 +15,9 @@ from ..variationalform import variationalform_base
 # TeX: \delta \mathcal{P} = \delta \mathcal{P}_{\mathrm{kin}} + \delta \mathcal{P}_{\mathrm{int}} - \delta \mathcal{P}_{\mathrm{ext}} = 0, \quad \forall \; \delta\boldsymbol{v}
 class variationalform(variationalform_base):
 
-    def __init__(self, var_v, var_p, n, formulation='nonconservative'):
-        self.var_v = var_v
-        # self.var_p = var_p
-
-        self.n = n
-        self.formulation = formulation
-        # for convenience, use naming n0 for ALE (it's a normal determined by the initial mesh!)
-        self.n0 = self.n
-
     ### Kinetic virtual power \delta \mathcal{P}_{\mathrm{kin}}
 
-    def deltaW_kin_navierstokes_transient(self, a, v, rho, ddomain, w=None, Fale=None):
+    def deltaW_kin_navierstokes_transient(self, a, v, rho, ddomain, w=None, F=None):
         # standard Eulerian fluid
         if self.formulation=='nonconservative':
             # non-conservative form for Navier-Stokes:
@@ -43,7 +34,7 @@ class variationalform(variationalform_base):
         else:
             raise ValueError("Unknown fluid formulation! Choose either 'nonconservative' or 'conservative'.")
 
-    def deltaW_kin_navierstokes_steady(self, v, rho, ddomain, w=None, Fale=None):
+    def deltaW_kin_navierstokes_steady(self, v, rho, ddomain, w=None, F=None):
         if self.formulation=='nonconservative':
             return rho*ufl.dot(ufl.grad(v) * v, self.var_v)*ddomain
         elif self.formulation=='conservative':
@@ -51,23 +42,23 @@ class variationalform(variationalform_base):
         else:
             raise ValueError("Unknown fluid formulation!")
 
-    def deltaW_kin_stokes_transient(self, a, v, rho, ddomain, w=None, Fale=None):
+    def deltaW_kin_stokes_transient(self, a, v, rho, ddomain, w=None, F=None):
 
         return rho*ufl.dot(a, self.var_v)*ddomain
 
     ### Internal virtual power \delta \mathcal{P}_{\mathrm{int}}
 
     # TeX: \int\limits_{\Omega}\boldsymbol{\sigma} : \boldsymbol{\nabla}(\delta\boldsymbol{v})\,\mathrm{d}v
-    def deltaW_int(self, sig, ddomain, Fale=None):
+    def deltaW_int(self, sig, ddomain, F=None):
 
         return ufl.inner(sig, ufl.grad(self.var_v))*ddomain
 
     # TeX: \int\limits_{\Omega}\boldsymbol{\nabla}\cdot\boldsymbol{v}\,\delta p\,\mathrm{d}v
-    def deltaW_int_pres(self, v, var_p, ddomain, w=None, Fale=None):
+    def deltaW_int_pres(self, v, var_p, ddomain, w=None, F=None):
 
         return ufl.div(v)*var_p*ddomain
 
-    def res_v_strong_navierstokes_transient(self, a, v, rho, sig, w=None, Fale=None):
+    def res_v_strong_navierstokes_transient(self, a, v, rho, sig, w=None, F=None):
         if self.formulation=='nonconservative':
[…]
