### 2023-10-11 70dddde
Move active stress ufl forms out of time loop; add option for spatially varying active stress
