### 2023-10-14 23c13f3
Move ufl integration measure to central place; some steps towards monolithic FSI (not yet operational)
