+++ src/ambit_fe/solid/solid_main.py
@@ -52,8 +52,8 @@ class SolidmechanicsProblem(problem_base):
         # number of distinct domains (each one has to be assigned a own material model)
         self.num_domains = len(constitutive_models)
         # for FSI, we want to specify the subdomains
-        try: domain_ids = self.io.io_params['domain_ids_solid']
-        except: domain_ids = np.arange(1,self.num_domains+1)
+        try: self.domain_ids = self.io.io_params['domain_ids_solid']
+        except: self.domain_ids = np.arange(1,self.num_domains+1)
 
         self.constitutive_models = utilities.mat_params_to_dolfinx_constant(constitutive_models, self.io.mesh)
 
@@ -70,15 +70,20 @@ class SolidmechanicsProblem(problem_base):
 
         self.fem_params = fem_params
 
+        # create domain and boundaty integration measures
+        if self.io.mt_d_master is not None:
+            self.dx = ufl.Measure("dx", domain=self.io.mesh_master, subdomain_data=self.io.mt_d_master, metadata={'quadrature_degree': self.quad_degree})
+        else:
+            self.dx = ufl.Measure("dx", domain=self.io.mesh_master, metadata={'quadrature_degree': self.quad_degree})
+
+        self.ds = ufl.Measure("ds", domain=self.io.mesh_master, subdomain_data=self.io.mt_b1_master, metadata={'quadrature_degree': self.quad_degree})
+
         # collect domain data
-        self.dx_, self.rho0 = [], []
-        for i, n in enumerate(domain_ids):
-            # integration domains
-            if self.io.mt_d_master is not None: self.dx_.append(ufl.dx(domain=self.io.mesh_master, subdomain_data=self.io.mt_d_master, subdomain_id=n, metadata={'quadrature_degree': self.quad_degree}))
-            else:                               self.dx_.append(ufl.dx(domain=self.io.mesh_master, metadata={'quadrature_degree': self.quad_degree}))
+        self.rho0 = []
+        for n, N in enumerate(self.domain_ids):
             # data for inertial forces: density
             if self.timint != 'static':
-                self.rho0.append(self.constitutive_models['MAT'+str(i+1)]['inertia']['rho0'])
+                self.rho0.append(self.constitutive_models['MAT'+str(n+1)]['inertia']['rho0'])
 
         try: self.prestress_initial = fem_params['prestress_initial']
         except: self.prestress_initial = False
@@ -319,7 +324,7 @@ class SolidmechanicsProblem(problem_base):
 
         # growth threshold (as function, since in multiscale approach, it can vary element-wise)
         if self.have_growth and self.localsolve:
-            growth_thres_proj = project(self.mat_growth_thres, self.Vd_scalar, self.dx_, comm=self.comm)
+            growth_thres_proj = project(self.mat_growth_thres, self.Vd_scalar, self.dx, domids=self.domain_ids, comm=self.comm)
             self.growth_thres.vector.ghostUpdate(addv=PETSc.InsertMode.ADD, mode=PETSc.ScatterMode.REVERSE)
             self.growth_thres.interpolate(growth_thres_proj)
 
@@ -330,11 +335,11 @@ class SolidmechanicsProblem(problem_base):
             for nf in range(len(self.io.fiber_data)):
                 fibarray.append('f'+str(nf+1))
 
[…]
