+++ src/ambit_fe/ioparams.py
@@ -11,6 +11,8 @@ def check_params_io(params):
 
     valid_params = ['duplicate_mesh_domains',
                     'fiber_data',
+                    'domain_ids_solid',
+                    'domain_ids_fluid',
                     'gridname_domain',
                     'gridname_boundary',
                     'indicate_results_by',
@@ -29,6 +31,7 @@ def check_params_io(params):
                     'restart_step',
                     'results_to_write',
                     'simname',
+                    'surface_ids_interface',
                     'USE_MIXED_DOLFINX_BRANCH',
                     'volume_laplace',
                     'write_results_every',
@@ -190,7 +193,8 @@ def check_params_coupling_fluid_ale(params):
     valid_params = ['coupling_ale_fluid',
                     'coupling_fluid_ale',
                     'coupling_strategy',
-                    'fluid_on_deformed']
+                    'fluid_on_deformed',
+                    'fsi_governing_type']
 
     for k in params.keys():
         if k not in valid_params:
