### 2023-10-16 a25f2b2
Revamp tests to be called via pytest; add github workflow for automatic test execution
