+++ tests/test_flow0d_0Dheart_syspulcorvad.py
@@ -1,0 +1,262 @@
+#!/usr/bin/env python3
+
+"""
+ATTENTION: 0D VAD model not yet fully functional!
+"""
+
+import ambit_fe
+
+import sys
+import numpy as np
+from pathlib import Path
+import pytest
+
+
+@pytest.mark.flow0d
+@pytest.mark.skip(reason="Not yet ready for testing.")
+def test_main():
+
+    basepath = str(Path(__file__).parent.absolute())
+
+    # reads in restart step from the command line
+    try: restart_step = int(sys.argv[1])
+    except: restart_step = 0
+
+    IO_PARAMS         = {'problem_type'          : 'flow0d',
+                         'write_results_every'   : -999,
+                         'output_path'           : basepath+'/tmp',
+                         'simname'               : 'test',
+                         'write_restart_every'   : 50,
+                         'restart_step'          : restart_step}
+
+    SOLVER_PARAMS     = {'tol_res'               : 1.0e-8,
+                         'tol_inc'               : 1.0e-8}
+
+    TIME_PARAMS       = {'maxtime'               : 10*1.0,
+                         'numstep'               : 10*100,
+                         'timint'                : 'ost',
+                         'theta_ost'             : 0.5,
+                         'initial_conditions'    : init(),
+                         'eps_periodic'          : 0.03,
+                         'periodic_checktype'    : ['pQvar']}
+
+    MODEL_PARAMS      = {'modeltype'             : 'syspul',
+                         'coronary_model'        : 'ZCRp_CRd_lr',
+                         'vad_model'             : 'lvad',
+                         'parameters'            : param(),
+                         'chamber_models'        : {'lv' : {'type' : '0D_elast', 'activation_curve' : 2},
+                                                    'rv' : {'type' : '0D_elast', 'activation_curve' : 2},
+                                                    'la' : {'type' : '0D_elast', 'activation_curve' : 1},
+                                                    'ra' : {'type' : '0D_elast', 'activation_curve' : 1}},
+                         'valvelaws'             : {'av' : ['smooth_pres_momentum',0],
+                                                    'mv' : ['pwlin_pres'],
+                                                    'pv' : ['pwlin_pres'],
+                                                    'tv' : ['pwlin_pres']}}
+
+
+    # define your time curves here (syntax: tcX refers to curve X)
+    class time_curves():
+
+        def tc1(self, t): # atrial activation
+
+            act_dur = 2.*param()['t_ed']
+            t0 = 0.
+
+            if t >= t0 and t <= t0 + act_dur:
+                return 0.5*(1.-np.cos(2.*np.pi*(t-t0)/act_dur))
+            else:
+                return 0.0
+
+        def tc2(self, t): # ventricular activation
+
+            act_dur = 1.8*(param()['t_es'] - param()['t_ed'])
[…]
