+++ tests/flow0d_0Dheart_syspulcap.py
@@ -1,384 +1,0 @@
-#!/usr/bin/env python3
-
-"""
-"""
-
-import ambit_fe
-
-import sys, traceback
-import numpy as np
-from pathlib import Path
-
-
-def main():
-
-    basepath = str(Path(__file__).parent.absolute())
-
-    IO_PARAMS         = {'problem_type'          : 'flow0d',
-                         'write_results_every'   : -999,
-                         'output_path'           : basepath+'/tmp',
-                         'simname'               : 'test'}
-
-    SOLVER_PARAMS     = {'tol_res'               : 1.0e-8,
-                         'tol_inc'               : 1.0e-8}
-
-    TIME_PARAMS       = {'maxtime'               : 10*1.0,
-                         'numstep'               : 10*100,
-                         'timint'                : 'ost',
-                         'theta_ost'             : 0.5,
-                         'initial_conditions'    : init(),
-                         'eps_periodic'          : 0.03,
-                         'periodic_checktype'    : ['pQvar']}
-
-    MODEL_PARAMS      = {'modeltype'             : 'syspulcap',
-                         'parameters'            : param(),
-                         'chamber_models'        : {'lv' : {'type' : '0D_elast', 'activation_curve' : 2},
-                                                    'rv' : {'type' : '0D_elast', 'activation_curve' : 2},
-                                                    'la' : {'type' : '0D_elast', 'activation_curve' : 1},
-                                                    'ra' : {'type' : '0D_elast', 'activation_curve' : 1}}}
-
-
-    # define your time curves here (syntax: tcX refers to curve X)
-    class time_curves():
-
-        def tc1(self, t): # atrial activation
-
-            act_dur = 2.*param()['t_ed']
-            t0 = 0.
-
-            if t >= t0 and t <= t0 + act_dur:
-                return 0.5*(1.-np.cos(2.*np.pi*(t-t0)/act_dur))
-            else:
-                return 0.0
-
-        def tc2(self, t): # ventricular activation
-
-            act_dur = 1.8*(param()['t_es'] - param()['t_ed'])
-            t0 = param()['t_ed']
-
-            if t >= t0 and t <= t0 + act_dur:
-                return 0.5*(1.-np.cos(2.*np.pi*(t-t0)/act_dur))
-            else:
-                return 0.0
-
-
-    # problem setup
-    problem = ambit_fe.ambit_main.Ambit(IO_PARAMS, TIME_PARAMS, SOLVER_PARAMS, constitutive_params=MODEL_PARAMS, time_curves=time_curves())
-
-    # solve time-dependent problem
-    problem.solve_problem()
-
-
-    # --- results check
-    tol = 1.0e-6
-
-    s_corr = np.zeros(problem.mp.cardvasc0D.numdof)
-
-    # correct results
-    s_corr[0] = 2.9433925704892139E+04
-    s_corr[1] = 7.0386145039016035E-01
[…]
