+++ tests/test_solid_flow0d_monodir2field_flux_syspulcap_3Dheart_schur3x3.py
@@ -1,0 +1,414 @@
+#!/usr/bin/env python3
+
+"""
+3D biventricular generic heart, testing of:
+- incompressible Neo-Hookean material (p2p1 interpolation), viscous Green material
+- 3D-0D monolithic solution of 3D heart w/ syspulcap circulation (flux coupling)
+- Robin BCs in xyz and normal direction (spring and dashpot)
+- OST time-integration for solid
+- 3x3 block iterative method for incompressible solid coupled to 0D model
+"""
+
+import ambit_fe
+
+import sys
+import numpy as np
+from pathlib import Path
+import pytest
+
+
+@pytest.mark.solid_flow0d
+def test_main():
+
+    basepath = str(Path(__file__).parent.absolute())
+
+
+    IO_PARAMS            = {'problem_type'          : 'solid_flow0d',
+                            'mesh_domain'           : basepath+'/input/heart3Dcoarse_domain.xdmf',
+                            'mesh_boundary'         : basepath+'/input/heart3Dcoarse_boundary.xdmf',
+                            'write_results_every'   : -999,
+                            'output_path'           : basepath+'/tmp/',
+                            'results_to_write'      : ['displacement','pressure'],
+                            'simname'               : 'test',
+                            'ode_parallel'          : True}
+
+    SOLVER_PARAMS        = {'solve_type'            : 'iterative',
+                            'iterative_solver'      : 'gmres',
+                            'block_precond'         : 'schur3x3',
+                            'precond_fields'        : [{'prec':'amg'}, {'prec':'amg'}, {'prec':'direct'}],
+                            'tol_res'               : [1.0e-8,1.0e-8,1.0e-6], # u,p,0d
+                            'tol_inc'               : [1.0e-8,1.0e-8,1.0e-6], # u,p,0d
+                            'tol_lin_rel'           : 1.0e-9,
+                            'lin_norm_type'         : 'preconditioned',
+                            'print_liniter_every'   : 50,
+                            'divergence_continue'   : None,
+                            'ptc'                   : False,
+                            'k_ptc_initial'         : 0.1}
+
+    TIME_PARAMS_SOLID    = {'maxtime'               : 1.0,
+                            'numstep'               : 100,
+                            'numstep_stop'          : 1,
+                            'timint'                : 'ost',
+                            'theta_ost'             : 0.5}
+
+    TIME_PARAMS_FLOW0D   = {'timint'                : 'ost',
+                            'theta_ost'             : 0.5,
+                            'initial_conditions'    : init(),
+                            'eps_periodic'          : 1.0e-3,
+                            'periodic_checktype'    : None}
+
+    MODEL_PARAMS_FLOW0D  = {'modeltype'             : 'syspulcap',
[…]

[…]
