### 2023-10-16 #5744 @marchirschvogel
> @marchirschvogel can you address the issues raised by the reviewers ☝️

Yes, I am currently addressing the issues, just having updated the tests such that they can be run with pytest. Currently figuring out how they can be run using Github workflows when pushing, since the code needs to run inside a container for this.
I will also work on the documentation and tutorials for the respective problem types the next days.
