### 2023-10-16 #5744 @marchirschvogel
> Hi @Kevin-Mattheus-Moerman, sorry for the delay. I have just gone through the entire checklist and submitted a GitHub issue with my suggestions. I agree with @finsberg that it's not ready for publication yet.
>
> Regarding the installation process, I have verified that the Docker image works fine. However, I am having trouble installing Fenicsx locally on my machine, and therefore, I cannot verify if `pip install` works. Hopefully, other reviewers can confirm this.

@CZHU20 Thanks for checking this. I've also added a devenv Docker container (`docker pull ghcr.io/marchirschvogel/ambit:devenv`) where all the dependencies are installed. Hence, you could verify the pip install inside this container.
