+++ .github/workflows/docker-image.yml
@@ -3,13 +3,12 @@ name: Create and publish a Docker image
 
 # Configures this workflow to run every time a change is pushed to the branch called `release`.
 on:
-  push:
-    branches: ['release']
+  push
 
 # Defines two custom environment variables for the workflow. These are used for the Container registry domain, and a name for the Docker image that this workflow builds.
 env:
   REGISTRY: ghcr.io
-  IMAGE_NAME: ${{ github.repository }}
+  IMAGE_NAME: ${{ github.repository }}:latest
 
 # There is a single job in this workflow. It's configured to run on the latest available version of Ubuntu.
 jobs:
