+++ .github/workflows/docker-image.yml
@@ -3,7 +3,8 @@ name: Create and publish a Docker image
 
 # Configures this workflow to run every time a change is pushed to the branch called `release`.
 on:
-  push
+  push:
+    branches: ['release'] # TODO: Should be push to master, but action currently fails when trying to push Docker image (ERROR: unauthorized: access token has insufficient scopes)
 
 # Defines two custom environment variables for the workflow. These are used for the Container registry domain, and a name for the Docker image that this workflow builds.
 env:
