### 2023-10-17 marchirschvogel/ambit#4 @marchirschvogel
Thanks for pointing to this. I've added an automated install and testing workflow (.github/workflows/test.yml) that is executed at each push event to the master branch.
