### 2023-10-18 f9d41d5
fix for outflow stab; some doc updates; code coverage increase in tests
