### 2023-11-08 5951571
Add examples and comments, prepare code for edge integration in 3D
