+++ doc/examples/solid_flow0d_heart_cycle.py
@@ -1,0 +1,351 @@
+#!/usr/bin/env python3
+
+"""
+This example demonstrates how to set up and simulate a two-chamber (left and right ventricular) solid mechanics heart model coupled to a closed-loop
+0D circulatory system. A full dynamic heart cycle of duration 1 s is simulated, where the active contraction is modeled by a prescribed active stress approach.
+Passive material behavior of the heart muscle is governed by the Holzapfel-Ogden anisotropic strain energy function and a strain rate-dependent viscous
+model.
+We start the simulation with "prestressing" using the MULF method (Gee et al. 2010, Schein and Gee 2021), which allows to imprint loads without changing the geometry,
+where the solid is loaded to the initial left and right ventricular pressures.
+Thereafter, we kickstart the dynamic simulation with passive ventricular filling by the systole of the atria (0D chamber models). Ventricular systole
+happens in t \in [0.2 s, 0.53 s], hence lasting a third of the whole cycle time. After systole, the heart relaxes and eventually fills to about the same pressure
+as it has been initialized to.
+
+NOTE: For demonstrative purposes, a fairly coarse finite element discretization is chosen here, which by no means yields a spatially converged solution and which
+may be prone to locking phenomena. The user may increse the parameter 'order_disp' in the FEM_PARAMS section from 1 to 2 (and maybe increase 'quad_degree' to 6)
+such that quadratic finite element ansatz functions (instead of linear ones) are used. While this will increase accuracy and mitigate locking, computation time will
+increase.
+
+INSTRUCTIONS:
+Run the simulation, either in one of the provided Docker containers or using your own FEniCSx/Ambit installation, using the command
+mpiexec -n 1 python3 solid_flow0d_heart_cycle.py
+It is fully sufficient to use one core (mpiexec -n 1) for the presented setup, while you might want to use more (e.g., mpiexec -n 4) is you increase 'order_disp' to 2.
+
+Open the results file results_solid_flow0d_heart_cycle_displacement.xdmf in Paraview, and visualize the deformation over the heart cycle.
+
+For postprocessing of the time courses of pressures, volumes, and fluxes of the 0D model, make sure to have Gnuplot (and TeX) installed.
+Navigate to the output folder (tmp/) and execute the script flow0d_plot.py (which lies in ambit/src/ambit_fe/postprocess/):
+flow0d_plot.py -s solid_flow0d_heart_cycle
+A folder 'plot_solid_flow0d_heart_cycle' is created inside tmp/. Look at the results of pressures (p), volumes (V), and fluxes (q,Q) over time.
+Subscripts v, at, ar, ven refer to 'ventricular', 'atrial', 'arterial', and 'venous', respectively. Superscripts l, r, sys, pul refer to 'left', 'right', 'systemic', and
+'pulmonary', respectively.
+Try to understand the time courses of the respective pressures, as well as the plots of ventricular pressure over volume.
+Check that the overall system volume is constant and around 4-5 liters.
+"""
+
+import ambit_fe
+
+import sys
+import numpy as np
+from pathlib import Path
+
+
+def main():
+
+    basepath = str(Path(__file__).parent.absolute())
+
+
+    """
+    Parameters for input/output
+    """
+    IO_PARAMS            = {# problem type 'solid_flow0d' indicates a coupling of the individual problems 'solid' and 'flow0d'
+                            'problem_type'          : 'solid_flow0d',
+                            # the meshes for the domain and boundary topology are specified separately
+                            'mesh_domain'           : basepath+'/input/heart3D_domain.xdmf',
+                            'mesh_boundary'         : basepath+'/input/heart3D_boundary.xdmf',
[…]
