### 2023-11-09 2527e2c
Update to examples, minor fixes
