### 2023-11-09 97a98ac
Add solid mechanics example
