### 2023-11-10 18c3070
Add option for truly 2D meshes; improve documentation; add further and revamp examples
