+++ doc/examples/fluid_flow0d/fluid_flow0d_pipe.py
@@ -1,0 +1,166 @@
+#!/usr/bin/env python3
+
+"""
+A fluid mechanics problem of incompressible Navier-Stokes flow through a pipe, which is separated by an internal valve (always closed). A 0D model of two 2-element Windkessels in series
+bypasses flow from the one region to the other. The setup is rather academic but should illustrate how 0D models can be used to connect independent fluid regions.
+"""
+
+import ambit_fe
+import numpy as np
+from pathlib import Path
+
+
+def main():
+
+    basepath = str(Path(__file__).parent.absolute())
+
+
+    """
+    Parameters for input/output
+    """
+    IO_PARAMS            = {# problem type 'fluid_flow0d' indicates a coupling of the individual problems 'fluid' and 'flow0d'
+                            'problem_type'          : 'fluid_flow0d',
+                            # For this setup in order to model an internal valve/Dirichlet boundary, we need to duplicate the pressure space at this boundary; otherwise,
+                            # any Dirichlet condition on the velocity inside the domain will not prevent fluid to experience a pressure gradient across that plane, leading
+                            # to unphysical de- and acceleration of fluid infront of and behind the valve.
+                            # This duplicate pressure space can only be achieved using the mixed Dolfinx branch, which is installed in the Ambit devenv Docker container.
+                            # In the future, this functionality is expected to be merged into the main branch of Dolfinx (as it has been announced).
+                            'USE_MIXED_DOLFINX_BRANCH' : True,
+                            # indicate which domain IDs (specified in 'mesh_domain' file) should be split (creating a submesh)
+                            'duplicate_mesh_domains': [1,2],
+                            # the meshes for the domain and boundary topology are specified separately
+                            'mesh_domain'           : basepath+'/input/pipe_domain.xdmf',
+                            'mesh_boundary'         : basepath+'/input/pipe_boundary.xdmf',
+                            'order_fib_input'       : 1,
+                            # at which step frequency to write results (set to 0 in order to not write any output)
+                            'write_results_every'   : 1,
+                            # where to write the output to
+                            'output_path'           : basepath+'/tmp/',
+                            # which results to write: here, all 3D fields need to be specified, while the 0D model results are output nevertheless
+                            'results_to_write'      : ['velocity','pressure'],
+                            # the 'midfix' for all simulation result file names: will be results_<simname>_<field>.xdmf/.h5/.txt
+                            'simname'               : 'fluid_flow0d_pipe'}
+                      
+    """
+    Parameters for the linear and nonlinear solution schemes
+    """
+    SOLVER_PARAMS        = {# this specifies which linear solution strategy to use; since this problem has less than 10'000 degrees of freedom, we comfortably can use a direct solver
+                            'solve_type'            : 'direct',
+                            # residual and increment tolerances: first value for the fluid mechanics problem (momentum balance and continuity), second value for the 3D-0D constraint problem
+                            'tol_res'               : [1e-8, 1e-8, 1e-6],
+                            'tol_inc'               : [1e-8, 1e-8, 1e-6],
+                            # subsolver tolerances for the 0D model solver: tolerances for residual and increment
+                            'subsolver_params'      : {'tol_res' : 1e-6,
+                                                       'tol_inc' : 1e-6}}
+
+    """
+    Parameters for the solid mechanics time integration scheme, plus the global time parameters
+    """
+    TIME_PARAMS_FLUID    = {# the maximum simulation time - here 0.2 seconds
+                            'maxtime'               : 0.2,
[…]
