+++ doc/examples/fluid/README.md
@@ -2,6 +2,8 @@
 
 This example shows how to set up 2D fluid flow in a channel around a rigid obstacle, using Taylor-Hood elements.
 
+![Simulation setup.](https://github.com/marchirschvogel/ambit/assets/52761273/494703fd-72cb-4f83-9984-8f3e9b3b63c8)
+**Simulation setup.**
 
 ### Instructions ###
 
@@ -16,3 +18,6 @@ Open the results file results_channel_velocity.xdmf and results_channel_pressure
 ### Solution
 
 The figure shows the velocity magnitude (top) as well as the pressure (bottom part) at the end of the simulation.
+
+![Velocity streamlines at end of simulation, color indicates magnitude of the velocity.](https://github.com/marchirschvogel/ambit/assets/52761273/fb3445f8-b928-4379-9c91-7c001fb6671b)
+**Velocity streamlines at end of simulation, color indicates magnitude of the velocity.**
