### 2023-11-14 eda9a07
Revamp docs structure
