+++ docs/source/index.rst
@@ -1,0 +1,2 @@
+Welcome to Ambit's documentation!
+=================================
