+++ docs/TeX/ambit_doc.tex
@@ -158,7 +158,7 @@ Avenues for future functionality include cardiac electrophysiology, scalar trans
 
 In the following, a brief description of the supported problem types is given, including the strong and weak form of the underlying equations as well as the discrete assembled systems that are solved.\\
 
-Examples of input files for the respective problem types can be found in the folder \verb\doc/examples\ (with detailed setup descriptions) or amogst the test cases in the folder \verb\tests\.
+Examples of input files for the respective problem types can be found in the folder \verb.demos. (with detailed setup descriptions) or amogst the test cases in the folder \verb.tests..
 
 %\begin{figure}[!htp]
 %\centering
@@ -168,9 +168,9 @@ Examples of input files for the respective problem types can be found in the fol
 
 \section{Solid mechanics}
 
--- Example: \verb\doc/examples/solid\\\
+-- Example: \verb.demos/solid.\\
 
--- Problem type: \verb\solid\\\
+-- Problem type: \verb.solid.\\
 
 -- Solid mechanics are formulated in a Total Lagrangian frame
 
@@ -283,9 +283,9 @@ r_p(\bs{u};\delta p) &:= \delta \mathcal{W}_{\mathrm{pres}}(\bs{u};\delta p) = 0
 
 \subsection{Eulerian reference frame}
 
--- Example: \verb\doc/examples/fluid\\\
+-- Example: \verb.demos/fluid.\\
 
--- Problem type: \verb\fluid\\\
+-- Problem type: \verb.fluid.\\
 
 -- Incompressible Navier-Stokes equations in Eulerian reference frame\\
 
@@ -348,7 +348,7 @@ r_p(\bs{v};\delta p) &:= \delta \mathcal{P}_{\mathrm{pres}}(\bs{v};\delta p), \q
 \subsubsection{Stabilization}\label{subsubsec:stab}
 Streamline-upwind Petrov-Galerkin/pressure-stabilizing Petrov-Galerkin (SUPG/PSPG) methods are implemented, either using the full or a reduced scheme\\
 
-Full scheme according to \cite{tezduyar2000}: \verb\supg_pspg\:\\
+Full scheme according to \cite{tezduyar2000}: \verb.supg_pspg.:\\
 -- Velocity residual operator (\ref{eq:res_v_fluid}) is augmented with the following terms:
 \begin{align}
 r_v \leftarrow r_v &+ \frac{1}{\rho}\int\limits_{\Om_t} \tau_{\mathrm{SUPG}}\,(\bs{\nabla}\delta\bs{v})\,\bs{v} \cdot \left[\rho\left(\frac{\partial \bs{v}}{\partial t} + (\bs{\nabla}\bs{v})\,\bs{v}\right) - \bs{\nabla} \cdot \bs{\sigma}(\bs{v},p)\right]\,\mathrm{d}v \\
@@ -360,7 +360,7 @@ r_p \leftarrow r_p &+ \frac{1}{\rho}\int\limits_{\Om_t} \tau_{\mathrm{PSPG}}\,(\
 \end{align}\\
 
 
[…]
