+++ docs/source/index.rst
@@ -1,2 +1,9 @@
 Welcome to Ambit's documentation!
 =================================
+
+Contents
+--------
+
+.. toctree::
+
+   documentation
