+++ docs/source/documentation.md
@@ -1,622 +1,0 @@
----
-author:
-- 'Dr.-Ing. Marc Hirschvogel'
-bibliography:
-- 'ref.bib'
-title: 'Ambit -- A FEniCS-based cardiovascular physics solver'
----
-
-Preface
-=======
-
-Ambit is an open-source software tool written in Python for parallel
-multi-physics simulations focusing on -- but not limited to -- cardiac
-mechanics. Amongst others, it contains re-implementations and
-generalizations of methods developed by the author for his PhD thesis
-[@hirschvogel2018]. Ambit makes use of the open-source finite element
-library FEniCS/dolfinx (https://fenicsproject.org) [@logg2012] along
-with the linear algebra package PETSc (https://petsc.org)
-[@petsc-user-ref]. It is constantly updated to ensure compatibility with
-a recent dolfinx development version, hence guaranteeing a
-state-of-the-art finite element and linear algebra backend.\
-Ambit is designed such that the user only needs to provide input files
-that define parameters through Python dictionaries, hence no programming
-or in-depth knowledge of any library-specific syntax is required.\
-Ambit provides general nonlinear (compressible or incompressible) finite
-strain solid dynamics [@holzapfel2000], implementing a range of
-hyperelastic, viscous, and active material models. Specifically, the
-well-known anisotropic Holzapfel-Ogden [@holzapfel2009] and Guccione
-models [@guccione1995] for structural description of the myocardium are
-provided, along with a bunch of other models. It further implements
-strain- and stress-mediated volumetric growth models [@goektepe2010]
-that allow to model (maladaptive) ventricular shape and size changes.
-Inverse mechanics approaches to imprint loads into a reference state are
-implemented using the so-called prestressing method [@gee2010] in
-displacement formulation [@schein2021].\
-Furthermore, fluid dynamics in terms of incompressible
-Navier-Stokes/Stokes equations -- either in Eulerian or Arbitrary
-Lagrangian-Eulerian (ALE) reference frames -- are implemented.
-Taylor-Hood elements or equal-order approximations with SUPG/PSPG
-stabilization [@tezduyar2000] can be used.\
-A variety of reduced 0D lumped models targeted at blood circulation
-modeling are implemented, including 3- and 4-element Windkessel models
-[@westerhof2009] as well as closed-loop full circulation
-[@hirschvogel2017] and coronary flow models [@arthurs2016].\
-Monolithic multi-physics coupling of solid, fluid, and ALE-fluid with 0D
-lumped models is implemented such that cardiovascular simulations with
[…]
