+++ docs/source/documentation.rst
@@ -4,10 +4,6 @@ Ambit – A FEniCS-based cardiovascular physics solver
 
 :Author: Dr.-Ing. Marc Hirschvogel
 
-.. role:: raw-latex(raw)
-   :format: latex
-..
-
 Preface
 =======
 
@@ -15,12 +11,11 @@ Preface
   multi-physics simulations focusing on – but not limited to – cardiac
   mechanics. Amongst others, it contains re-implementations and
   generalizations of methods developed by the author for his PhD thesis
-  :raw-latex:`\cite{hirschvogel2018}`. Ambit makes use of the
-  open-source finite element library FEniCS/dolfinx
-  (https://fenicsproject.org) :raw-latex:`\cite{logg2012}` along with
-  the linear algebra package PETSc (https://petsc.org)
-  :raw-latex:`\cite{petsc-user-ref}`. It is constantly updated to ensure
-  compatibility with a recent dolfinx development version, hence
+  (Hirschvogel 2018). Ambit makes use of the open-source finite element
+  library FEniCS/dolfinx (https://fenicsproject.org) (Logg, Mardal, and
+  Wells 2012) along with the linear algebra package PETSc
+  (https://petsc.org) (Balay et al. 2022). It is constantly updated to
+  ensure compatibility with a recent dolfinx development version, hence
   guaranteeing a state-of-the-art finite element and linear algebra
   backend.
 | Ambit is designed such that the user only needs to provide input files
@@ -28,28 +23,27 @@ Preface
   programming or in-depth knowledge of any library-specific syntax is
   required.
 | Ambit provides general nonlinear (compressible or incompressible)
-  finite strain solid dynamics :raw-latex:`\cite{holzapfel2000}`,
-  implementing a range of hyperelastic, viscous, and active material
-  models. Specifically, the well-known anisotropic Holzapfel-Ogden
-  :raw-latex:`\cite{holzapfel2009}` and Guccione models
-  :raw-latex:`\cite{guccione1995}` for structural description of the
-  myocardium are provided, along with a bunch of other models. It
-  further implements strain- and stress-mediated volumetric growth
-  models :raw-latex:`\cite{goektepe2010}` that allow to model
+  finite strain solid dynamics (Holzapfel 2000), implementing a range of
+  hyperelastic, viscous, and active material models. Specifically, the
+  well-known anisotropic Holzapfel-Ogden (Holzapfel and Ogden 2009) and
+  Guccione models (Guccione, Costa, and McCulloch 1995) for structural
+  description of the myocardium are provided, along with a bunch of
+  other models. It further implements strain- and stress-mediated
+  volumetric growth models (Göktepe et al. 2010) that allow to model
   (maladaptive) ventricular shape and size changes. Inverse mechanics
   approaches to imprint loads into a reference state are implemented
-  using the so-called prestressing method :raw-latex:`\cite{gee2010}` in
-  displacement formulation :raw-latex:`\cite{schein2021}`.
+  using the so-called prestressing method (Gee, Förster, and Wall 2010)
+  in displacement formulation (Schein and Gee 2021).
 | Furthermore, fluid dynamics in terms of incompressible
   Navier-Stokes/Stokes equations – either in Eulerian or Arbitrary
   Lagrangian-Eulerian (ALE) reference frames – are implemented.
[…]
