+++ docs/requirements.txt
@@ -1,2 +1,3 @@
 myst_parser
 sphinx-rtd-theme
+sphinxcontrib-bibtex
