### 2023-11-14 cf8a6e0
Fix citations in doc
