### 2023-11-14 365b5d8
Fix citations in doc
