### 2023-11-14 2fb6cbc
More fixes to doc
