### 2023-11-14 marchirschvogel/ambit#5 @marchirschvogel
Thanks for pointing to this, I was not aware of these tools for documentation. I've now put the whole documentation on https://ambit.readthedocs.io/ and indeed think that this will enhance user experience.

I am not quite sure yet how to best add API documentation and if docstrings need to be added to each function/class. The user actually only interacts with the main function by passing lists of dictionaries and would not need to call lower level functions.
But I will keep that on my agenda and am happy to receive any advice.

Let me know if the issue can be closed at that stage or whether further aspects need to be considered.
