### 2023-11-15 marchirschvogel/ambit#5 @finsberg
Hi @marchirschvogel! Great job, I think this looks really good!
Regarding doc strings I think you should add docstrings to the functions that a user interacts with, i.e at least need to have this on the main function. To generate documentation directly from docstrings you can e.g use `sphinx-apidoc` see https://www.sphinx-doc.org/en/master/man/sphinx-apidoc.html and https://www.sphinx-doc.org/en/master/tutorial/automatic-doc-generation.html

Also would it be possible to add the demos as a part of the online documentation?
