### 2023-11-17 252e842
Add some docstrings; update doc
