+++ README.md
@@ -2,11 +2,12 @@
 
 * Ambit - A FEniCS-based cardiovascular physics solver
 
-3D nonlinear solid and fluid mechanics finite element Python code using FEniCS and PETSc libraries, supporting
+3D nonlinear solid and fluid mechanics finite element Python code using FEniCS and PETSc libraries
 
-[**Heart cycle simulation of a generic bi-ventricular heart model coupled to a closed-loop circulation model.**](https://github.com/marchirschvogel/ambit/assets/52761273/61336888-82e1-4484-97ea-b77d3fb4ce2f)
-
+[**Heart cycle simulation of a generic bi-ventricular heart model coupled to a closed-loop circulation model.**](https://github.com/marchirschvogel/ambit/assets/52761273/5541e44a-e086-4833-aff1-f18fff274891)
 **Heart cycle simulation of a generic bi-ventricular heart model coupled to a closed-loop circulation model.**
+
+The following is supported:
 
 * Solid mechanics
   - Finite strain elastodynamics, implementing a range of hyperelastic isotropic and anisotropic as well as viscous constitutive laws
