+++ README.md
@@ -4,7 +4,11 @@
 
 3D nonlinear solid and fluid mechanics finite element Python code using FEniCS and PETSc libraries
 
-[**Heart cycle simulation of a generic bi-ventricular heart model coupled to a closed-loop circulation model.**](https://github.com/marchirschvogel/ambit/assets/52761273/5541e44a-e086-4833-aff1-f18fff274891)
+
+
+https://github.com/marchirschvogel/ambit/assets/52761273/90e483d2-d517-4bcd-900c-e8107026fa14
+
+
 **Heart cycle simulation of a generic bi-ventricular heart model coupled to a closed-loop circulation model.**
 
 The following is supported:
