+++ README.md
@@ -4,9 +4,7 @@
 
 3D nonlinear solid and fluid mechanics finite element Python code using FEniCS and PETSc libraries
 
-
-
-https://github.com/marchirschvogel/ambit/assets/52761273/bf555110-8b41-47dc-b5bc-2b6d018e11ae
+https://github.com/marchirschvogel/ambit/assets/52761273/2a9605c0-db75-4bbe-8191-6fe0c4cee59b
 
 **Heart cycle simulation of a generic bi-ventricular heart model coupled to a closed-loop circulation model.**
 
