+++ docs/source/documentation.rst
@@ -15,11 +15,11 @@ Preface
   multi-physics simulations focusing on – but not limited to – cardiac
   mechanics. Amongst others, it contains re-implementations and
   generalizations of methods developed by the author for his PhD thesis
-  :cite:t:`hirschvogel2018`. Ambit makes use of the
+  :cite:p:`hirschvogel2018`. Ambit makes use of the
   open-source finite element library FEniCS/dolfinx
-  (https://fenicsproject.org) :cite:t:`logg2012` along with
+  (https://fenicsproject.org) :cite:p:`logg2012` along with
   the linear algebra package PETSc (https://petsc.org)
-  :cite:t:`petsc-user-ref`. It is constantly updated to ensure
+  :cite:p:`petsc-user-ref`. It is constantly updated to ensure
   compatibility with a recent dolfinx development version, hence
   guaranteeing a state-of-the-art finite element and linear algebra
   backend.
@@ -28,28 +28,28 @@ Preface
   programming or in-depth knowledge of any library-specific syntax is
   required.
 | Ambit provides general nonlinear (compressible or incompressible)
-  finite strain solid dynamics :cite:t:`holzapfel2000`,
+  finite strain solid dynamics :cite:p:`holzapfel2000`,
   implementing a range of hyperelastic, viscous, and active material
   models. Specifically, the well-known anisotropic Holzapfel-Ogden
-  :cite:t:`holzapfel2009` and Guccione models
-  :cite:t:`guccione1995` for structural description of the
+  :cite:p:`holzapfel2009` and Guccione models
+  :cite:p:`guccione1995` for structural description of the
   myocardium are provided, along with a bunch of other models. It
   further implements strain- and stress-mediated volumetric growth
-  models :cite:t:`goektepe2010` that allow to model
+  models :cite:p:`goektepe2010` that allow to model
   (maladaptive) ventricular shape and size changes. Inverse mechanics
   approaches to imprint loads into a reference state are implemented
-  using the so-called prestressing method :cite:t:`gee2010` in
-  displacement formulation :cite:t:`schein2021`.
+  using the so-called prestressing method :cite:p:`gee2010` in
+  displacement formulation :cite:p:`schein2021`.
 | Furthermore, fluid dynamics in terms of incompressible
   Navier-Stokes/Stokes equations – either in Eulerian or Arbitrary
   Lagrangian-Eulerian (ALE) reference frames – are implemented.
   Taylor-Hood elements or equal-order approximations with SUPG/PSPG
[…]
