### 2023-11-19 796c412
Move demo description into documentation
