+++ docs/source/documentation.rst
@@ -77,6 +77,8 @@ Preface
   in the folder ``demos`` (with detailed setup descriptions) or amogst
   the test cases in the folder ``tests``.
 
+.. _sec:solid:
+
 Solid mechanics
 ===============
 
@@ -196,7 +198,14 @@ Displacement-based
       \begin{aligned}
       \delta \mathcal{W}_{\mathrm{ext}}(\boldsymbol{u};\delta\boldsymbol{u}) &= -\int\limits_{\mathit{\Gamma}_{0}^{\mathrm{N}}} (\boldsymbol{n}_0 \otimes \boldsymbol{n}_0)\left[k\,\boldsymbol{u} + c\,\boldsymbol{v}(\boldsymbol{u})\right] \cdot \delta\boldsymbol{u}\,\mathrm{d}A \label{eq:deltaWext_robn}\end{aligned}
 
-– Discrete linear system
+– Discrete nonlinear system to solve in each time step :math:`n`:
+
+.. math::
+
+   \begin{aligned}
+   \left.\boldsymbol{\mathsf{r}}_{u}(\boldsymbol{\mathsf{u}})\right|_{n+1} = \boldsymbol{\mathsf{0}}\label{eq:nonlin_sys_solid}\end{aligned}
+
+– Discrete linear system to solve in each Newton iteration :math:`k`:
 
 .. math::
 
@@ -230,15 +239,26 @@ Incompressible mechanics: 2-field displacement and pressure variables
      \begin{aligned}
      \delta \mathcal{W}_{\mathrm{pres}}(\boldsymbol{u};\delta p) &= \int\limits_{\mathit{\Omega}_{0}} (J(\boldsymbol{u}) - 1) \,\delta p \,\mathrm{d}V \label{eq:deltaWpres}\end{aligned}
 
-– Discrete linear system
+– Discrete nonlinear system to solve in each time step :math:`n`:
+
+.. math::
+
+   \begin{aligned}
+   \boldsymbol{\mathsf{r}}_{n+1} = \begin{bmatrix} \boldsymbol{\mathsf{r}}_{u}(\boldsymbol{\mathsf{u}},\boldsymbol{\mathsf{p}}) \\ \boldsymbol{\mathsf{r}}_{p}(\boldsymbol{\mathsf{u}}) \end{bmatrix}_{n+1} = \boldsymbol{\mathsf{0}}\label{eq:nonlin_sys_solid_inc}\end{aligned}
+
+– Discrete linear system to solve in each Newton iteration :math:`k`:
 
 .. math::
 
    \begin{aligned}
[…]
