+++ docs/TeX/ref.bib
@@ -1,170 +1,0 @@
-@article{arthurs2016,
-author={C. J. Arthurs and K. D. Lau and K. N. Asrress and S. R. Redwood and C. A. Figueroa},
-title={A mathematical model of coronary blood flow control: simulation of patient-specific three-dimensional hemodynamics during exercise},
-journal={Am J Physiol Heart Circ Physiol},
-year={2016},
-pages={H1242--H1258},
-volume={310},
-number={9},
-doi={10.1152/ajpheart.00517.2015}
-}
-
-@techreport{petsc-user-ref,
-author={Satish Balay and Shrirang Abhyankar and Mark~F. Adams and Steven Benson and Jed Brown
-    and Peter Brune and Kris Buschelman and Emil Constantinescu and Lisandro Dalcin and Alp Dener
-    and Victor Eijkhout and William~D. Gropp and V\'{a}clav Hapla and Tobin Isaac and Pierre Jolivet
-    and Dmitry Karpeev and Dinesh Kaushik and Matthew~G. Knepley and Fande Kong and Scott Kruger
-    and Dave~A. May and Lois Curfman McInnes and Richard Tran Mills and Lawrence Mitchell and Todd Munson
-    and Jose~E. Roman and Karl Rupp and Patrick Sanan and Jason Sarich and Barry~F. Smith
-    and Stefano Zampini and Hong Zhang and Hong Zhang and Junchao Zhang},
-title={{PETSc/TAO} Users Manual},
-institution={Argonne National Laboratory},
-number={ANL-21/39 - Revision 3.17},
-year={2022},
-}
-
-@article{farhat2014,
-author={C. Farhat and P. Avery and T. Chapman and J. Cortial},
-title={Dimensional reduction of nonlinear finite element dynamic models with finite rotations and energy-based mesh sampling and weighting for computational efficiency},
-journal={International Journal for Numerical Methods in Engineering},
-year={2014},
-pages={625--662},
-volume={98},
-number={9},
-doi={10.1002/nme.4668}
-}
-
-@article{gee2009,
-author={M. W. Gee and C. T. Kelley and R. B. Lehoucq},
-title={Pseudo-transient continuation for nonlinear transient elasticity},
-journal={International Journal for Numerical Methods in Engineering},
-year={2009},
-pages={1209--1219},
-volume={78},
-number={10},
-doi={10.1002/nme.2527}
-}
-
-@article{gee2010,
-author={M. W. Gee and Ch. F{\"o}rster and W. A. Wall},
-title={A computational strategy for prestressing patient-specific biomechanical problems under finite deformation},
-journal={International Journal for Numerical Methods in Biomedical Engineering},
-year={2010},
-pages={52--72},
-volume={26},
-number={1},
-doi={10.1002/cnm.1236}
-}
-
-@article{goektepe2010,
-author={S. G{\"o}ktepe and O. J. Abilez and K. K. Parker and E. Kuhl},
-title={A multiscale model for eccentric and concentric cardiac growth through sarcomerogenesis},
-journal={J Theor Biol},
-year={2010},
-pages={433--442},
-volume={265},
-number={3},
-doi={10.1016/j.jtbi.2010.04.023}
-}
-
-@article{guccione1995,
-author={J. M. Guccione and K. D. Costa and A. D. McCulloch},
-title={Finite element stress analysis of left ventricular mechanics in the beating dog heart},
-journal={J Biomech},
-year={1995},
-pages={1167--1177},
-volume={28},
-number={10},
-doi={10.1016/0021-9290(94)00174-3}
-}
-
-@article{hirschvogel2017,
[…]
