+++ docs/source/ambit_doc.tex
@@ -163,10 +163,10 @@ Examples of input files for the respective problem types can be found in the fol
 %\begin{figure}[!htp]
 %\centering
 %\includegraphics[width=0.8\textwidth]{/home/mh/work/publications/paper_ambit/code_design.pdf}
-%\caption{Code design}\label{fig:code_design}
+%\caption{Code design}\label{fig_code_design}
 %\end{figure}
 
-\section{Solid mechanics}\label{sec:solid}
+\section{Solid mechanics}\label{sec_solid}
 
 -- Example: \verb.demos/solid.\\
 
@@ -179,22 +179,22 @@ Examples of input files for the respective problem types can be found in the fol
 \subsubsection{Displacement-based}
 -- Primary variable: displacement $\bs{u}$
 \begin{align}
-\bs{\nabla}_{0} \cdot \bs{P}(\bs{u},\bs{v}(\bs{u})) + \hat{\bs{b}}_{0} &= \rho_{0} \bs{a}(\bs{u}) &&\text{in} \; \mathit{\Omega}_{0} \times [0, T], \label{eq:divP} \\
-\bs{u} &= \hat{\bs{u}} &&\text{on} \; \mathit{\Gamma}_{0}^{\mathrm{D}} \times [0, T], \label{eq:bc_u}\\
-\bs{t}_{0} = \bs{P}\bs{n}_{0} &= \hat{\bs{t}}_{0} &&\text{on} \; \mathit{\Gamma}_{0}^{\mathrm{N}} \times [0, T], \label{eq:bc_N}\\
-\bs{u}(\bs{x}_{0},0) &= \hat{\bs{u}}_{0}(\bs{x}_{0}) &&\text{in} \; \mathit{\Omega}_{0}, \label{eq:ini_u}\\
-\bs{v}(\bs{x}_{0},0) &= \hat{\bs{v}}_{0}(\bs{x}_{0}) &&\text{in} \; \mathit{\Omega}_{0}, \label{eq:ini_v}
+\bs{\nabla}_{0} \cdot \bs{P}(\bs{u},\bs{v}(\bs{u})) + \hat{\bs{b}}_{0} &= \rho_{0} \bs{a}(\bs{u}) &&\text{in} \; \mathit{\Omega}_{0} \times [0, T], \label{eq_divP} \\
+\bs{u} &= \hat{\bs{u}} &&\text{on} \; \mathit{\Gamma}_{0}^{\mathrm{D}} \times [0, T], \label{eq_bc_u}\\
+\bs{t}_{0} = \bs{P}\bs{n}_{0} &= \hat{\bs{t}}_{0} &&\text{on} \; \mathit{\Gamma}_{0}^{\mathrm{N}} \times [0, T], \label{eq_bc_N}\\
+\bs{u}(\bs{x}_{0},0) &= \hat{\bs{u}}_{0}(\bs{x}_{0}) &&\text{in} \; \mathit{\Omega}_{0}, \label{eq_ini_u}\\
+\bs{v}(\bs{x}_{0},0) &= \hat{\bs{v}}_{0}(\bs{x}_{0}) &&\text{in} \; \mathit{\Omega}_{0}, \label{eq_ini_v}
 \end{align}
 
 \subsubsection{Incompressible mechanics}
 -- Primary variables: displacement $\bs{u}$ and pressure $p$
 \begin{align}
[…]
