+++ demos/fluid/README.md
@@ -1,3 +1,3 @@
 # README #
 
-Refer to example description at ...
+Refer to example description at https://ambit.readthedocs.io/en/latest/documentation.html#fluid
