+++ docs/source/ambit_doc.tex
@@ -1039,7 +1039,7 @@ This example demonstrates how to set up a quasi-static solid mechanics elasticit
 
 \begin{figure}
 \centering
-\includegraphics[width=0.7\columnwidth]{fig/cantilever_setup.png}
+\includegraphics[width=0.75\textwidth]{fig/cantilever_setup.png}
 \caption{Cantilever, problem setup.}
 \label{fig:cantilever_setup}
 \end{figure}
@@ -1058,7 +1058,7 @@ Figure \ref{fig:cantilever_results} shows the displacement magnitude at the end 
 
 \begin{figure}
 \centering
-\includegraphics[width=0.7\columnwidth]{fig/cantilever_results}
+\includegraphics[width=0.75\textwidth]{fig/cantilever_results.png}
 \caption{Cantilever, tip deformation. Color shows displacement magnitude.}
 \label{fig:cantilever_results}
 \end{figure}
@@ -1075,7 +1075,7 @@ This example shows how to set up 2D fluid flow in a channel around a rigid obsta
 
 \begin{figure}
 \centering
-\includegraphics[width=1.0\columnwidth]{fig/channel_setup.png}
+\includegraphics[width=0.75\textwidth]{fig/channel_setup.png}
 \caption{Channel flow, problem setup.}
 \label{fig:channel_setup}
 \end{figure}
@@ -1094,7 +1094,7 @@ Fig. \ref{fig:channel_results} shows the velocity magnitude (top) as well as the
 
 \begin{figure}[ht]
 \centering
-\includegraphics[width=1.0\columnwidth]{fig/channel_results.png}
+\includegraphics[width=0.75\textwidth]{fig/channel_results.png}
 \caption{Velocity magnitude (top part) and pressure (bottom part) at end of simulation.}
 \label{fig:channel_results}
 \end{figure}
@@ -1110,7 +1110,7 @@ a specified value, here \verb"`eps_periodic'" in the \verb"TIME_PARAMS" dictiona
 
 \begin{figure}
 \centering
-\includegraphics[width=0.5\columnwidth]{fig/syspul_setup.png}
+\includegraphics[width=0.6\textwidth]{fig/syspul_setup.png}
 \caption{0D heart, systemic and pulmonary circulation, problem setup.}
 \label{fig:syspul_setup}
 \end{figure}
@@ -1134,7 +1134,7 @@ The solution is depicted in fig. \ref{fig:syspul_results}, showing the time cour
 
 \begin{figure}
 \centering
-\includegraphics[width=1.0\columnwidth]{fig/syspul_results.png}
+\includegraphics[width=0.8\textwidth]{fig/syspul_results.png}
 \caption{A. Left heart and systemic pressures over time. B. Right heart and pulmonary pressures over time. C. Left and right ventricular and atrial volumes over time. D. Left and right ventricular pressure-volume relationships of periodic (5th) cycle.}
 \label{fig:syspul_results}
 \end{figure}
@@ -1159,7 +1159,7 @@ increase.
 
 \begin{figure}
 \centering
-\includegraphics[width=0.6\columnwidth]{fig/heart_syspul_setup.png}
+\includegraphics[width=0.6\textwidth]{fig/heart_syspul_setup.png}
 \caption{Generic 3D ventricular heart model coupled to a closed-loop systemic and pulmonary circulation model.}
 \label{fig:heart_syspul_setup}
 \end{figure}
@@ -1192,7 +1192,7 @@ in ventricular volume?
 
 \begin{figure}
 \centering
-\includegraphics[width=1.0\columnwidth]{fig/heart_syspul_results.png}
+\includegraphics[width=0.8\textwidth]{fig/heart_syspul_results.png}
[…]
