+++ docs/source/ref.bib
@@ -86,12 +86,21 @@ number={8},
 doi={10.1002/cnm.2842}
 }
 
-@book{hirschvogel2018,
+@book{hirschvogel2018diss,
 author={M. Hirschvogel},
 title={Computational modeling of patient-specific cardiac mechanics with model reduction-based parameter estimation and applications to novel heart assist technologies},
 edition={},
 publisher={PhD thesis, Technische Universit{\"a}t M{\"u}nchen},
 year={2018}
+}
+
+@book{hirschvogel2019disspub,
+author={M. Hirschvogel},
+title={Computational modeling of patient-specific cardiac mechanics with model reduction-based parameter estimation and applications to novel heart assist technologies},
+edition={1},
+publisher={Verlag Dr. Hut, MediaTUM},
+year={2019},
+url={https://mediatum.ub.tum.de/1445317}
 }
 
 @article{hirschvogel2022preprint,
