+++ docs/source/documentation.rst
@@ -77,12 +77,45 @@ Preface
   in the folder ``demos`` (with detailed setup descriptions) or amogst
   the test cases in the folder ``tests``.
 
+Installation
+============
+
+| In order to use Ambit, you need to install FEniCSx
+  (https://github.com/FEniCS/dolfinx#installation) (latest
+  Ambit-compatible dolfinx development version dates to 19 Aug 2023).
+| Ambit can then be installed using pip, either the current release
+
+::
+
+   python3 -m pip install ambit-fe
+
+or latest development version:
+
+::
+
+   python3 -m pip install git+https://github.com/marchirschvogel/ambit.git
+
+Alternatively, you can pull a pre-built Docker image with FEniCSx and
+Ambit installed:
+
+::
+
+   docker pull ghcr.io/marchirschvogel/ambit:latest
+
+If a Docker image for development is desired, the following image
+contains all dependencies needed to install and run Ambit (including the
+dolfinx mixed branch):
+
+::
+
+   docker pull ghcr.io/marchirschvogel/ambit:devenv
+
 .. _sec:solid:
 
 Solid mechanics
 ===============
 
-| – Example: ``demos/solid``
+| – Example: Sec. `7.1 <#subsec:demos:solid>`__ and ``demos/solid``
 | – Problem type: ``solid``
 | – Solid mechanics are formulated in a Total Lagrangian frame
 
@@ -263,7 +296,7 @@ Fluid mechanics
 Eulerian reference frame
 ------------------------
 
-| – Example: ``demos/fluid``
+| – Example: Sec. `7.2 <#subsec:demos:fluid>`__ and ``demos/fluid``
 | – Problem type: ``fluid``
 | – Incompressible Navier-Stokes equations in Eulerian reference frame
 
@@ -548,7 +581,7 @@ Stabilization
 ~~~~~~~~~~~~~
 
 | ALE forms of stabilization introduced in sec.
-  `3.1.3 <#subsubsec:stab>`__
+  `4.1.3 <#subsubsec:stab>`__
 | ``supg_pspg``:
 | – Velocity residual operator
   (`[eq:res_v_fluid_ale] <#eq:res_v_fluid_ale>`__) is augmented with the
@@ -615,7 +648,7 @@ elements (without stabilization)
 0D flow: Lumped parameter models
 ================================
 
-| – Example: ``demos/flow0d``
+| – Example: Sec. `7.3 <#subsec:demos:flow0d>`__ ``demos/flow0d``
 | – Problem type: ``flow0d``
 
 2-element Windkessel
@@ -912,7 +945,8 @@ Multi-physics coupling
 Solid + 0D flow
 ---------------
 
-| – Example: ``demos/solid_flow0d``
+| – Example: Sec. `7.4 <#subsec:demos:solid_flow0d>`__ and
+  ``demos/solid_flow0d``
 | – Problem type: ``solid_flow0d``
 | – (`[eq:res_u_solid] <#eq:res_u_solid>`__) or
   (`[eq:res_u_solid_incomp] <#eq:res_u_solid_incomp>`__) augmented by
@@ -967,7 +1001,8 @@ incompressible solid:
 Fluid + 0D flow
 ---------------
 
-| – Example: ``demos/fluid_flow0d``
+| – Example: Sec. `7.5 <#subsec:demos:fluid_flow0d>`__
+  ``demos/fluid_flow0d``
 | – Problem type: ``fluid_flow0d``
 | – (`[eq:res_v_fluid] <#eq:res_v_fluid>`__) augmented by following
   term:
@@ -1056,12 +1091,12 @@ Fluid-Solid Interaction (FSI) + 0D flow
 Demos
 =====
 
-.. _subsec:demos_solid:
[…]
