+++ docs/source/ambit_doc.tex
@@ -38,6 +38,7 @@
 \usepackage{algpseudocode}
 \usepackage{tikz}
 \usepackage{pifont}
+%\usepackage[hidelinks]{hyperref}
 
 %%%%%%%%%%%%%%%%%%%%%%%%%
 \usepackage{xcolor}
@@ -166,9 +167,33 @@ Examples of input files for the respective problem types can be found in the fol
 %\caption{Code design}\label{fig:code_design}
 %\end{figure}
 
+\section{Installation}
+
+In order to use Ambit, you need to install FEniCSx (https://github.com/FEniCS/dolfinx#installation) (latest Ambit-compatible dolfinx development version dates to 19 Aug 2023).\\
+
+Ambit can then be installed using pip, either the current release
+\begin{verbatim}
+python3 -m pip install ambit-fe
+\end{verbatim}
+
+or latest development version:
+\begin{verbatim}
+python3 -m pip install git+https://github.com/marchirschvogel/ambit.git
+\end{verbatim}
+
+Alternatively, you can pull a pre-built Docker image with FEniCSx and Ambit installed:
+\begin{verbatim}
+docker pull ghcr.io/marchirschvogel/ambit:latest
+\end{verbatim}
+
+If a Docker image for development is desired, the following image contains all dependencies needed to install and run Ambit (including the dolfinx mixed branch):
+\begin{verbatim}
+docker pull ghcr.io/marchirschvogel/ambit:devenv
+\end{verbatim}
+
 \section{Solid mechanics}\label{sec:solid}
 
--- Example: \verb.demos/solid.\\
+-- Example: Sec. \ref{subsec:demos:solid} and \verb"demos/solid"\\
 
 -- Problem type: \verb.solid.\\
 
@@ -293,7 +318,7 @@ r_p(\bs{u};\delta p) &:= \delta \mathcal{W}_{\mathrm{pres}}(\bs{u};\delta p) = 0
 
 \subsection{Eulerian reference frame}\label{subsec:fluid_eulerian}
 
--- Example: \verb.demos/fluid.\\
+-- Example: Sec. \ref{subsec:demos:fluid} and \verb"demos/fluid"\\
 
 -- Problem type: \verb.fluid.\\
 
@@ -526,7 +551,7 @@ r_p \leftarrow r_p &+ \frac{1}{\rho}\int\limits_{\Om_0} J\,d_1\,((\bs{\nabla}_{0
 
 \section{0D flow: Lumped parameter models}\label{sec:flow0d}
 
--- Example: \verb.demos/flow0d.\\
+-- Example: Sec. \ref{subsec:demos:flow0d} \verb"demos/flow0d"\\
 
 -- Problem type: \verb.flow0d.\\
 
@@ -774,7 +799,7 @@ Q_{\mathrm{v}}^{r} := -\frac{\mathrm{d}V_{\mathrm{v}}^{r}}{\mathrm{d}t} \nonumbe
 
 \subsection{Solid + 0D flow}\label{subsec:solid_flow0d}
 
--- Example: \verb.demos/solid_flow0d.\\
+-- Example: Sec. \ref{subsec:demos:solid_flow0d} and \verb"demos/solid_flow0d"\\
 
 -- Problem type: \verb.solid_flow0d.\\
 
@@ -811,7 +836,7 @@ r_{\mathit{\Lambda}}(\mathit{\Lambda},\bs{u};\delta\mathit{\Lambda}):= \left(\in
 
 \subsection{Fluid + 0D flow}\label{subsec:fluid_flow0d}
 
--- Example: \verb.demos/fluid_flow0d.\\
[…]
