+++ docs/source/generated/ambit_fe.rst
@@ -1,0 +1,23 @@
+﻿ambit\_fe
+=========
+
+.. automodule:: ambit_fe
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
