+++ docs/source/ambit_doc.tex
@@ -169,7 +169,7 @@ Examples of input files for the respective problem types can be found in the fol
 
 \section{Installation}
 
-In order to use Ambit, you need to install FEniCSx (https://github.com/FEniCS/dolfinx#installation) (latest Ambit-compatible dolfinx development version dates to 19 Aug 2023).\\
+In order to use Ambit, you need to install FEniCSx (https://github.com/FEniCS/dolfinx\#installation) (latest Ambit-compatible dolfinx development version dates to 19 Aug 2023).\\
 
 Ambit can then be installed using pip, either the current release
 \begin{verbatim}
@@ -1101,7 +1101,7 @@ This example shows how to set up 2D fluid flow in a channel around a rigid obsta
 
 \begin{figure}
 \centering
-\includegraphics[width=0.85\textwidth]{fig/channel_setup.png}
+\includegraphics[width=0.9\textwidth]{fig/channel_setup.png}
 \caption{Channel flow, problem setup.}
 \label{fig:channel_setup}
 \end{figure}
@@ -1120,7 +1120,7 @@ Fig. \ref{fig:channel_results} shows the velocity magnitude (top) as well as the
 
 \begin{figure}[ht]
 \centering
-\includegraphics[width=0.85\textwidth]{fig/channel_results.png}
+\includegraphics[width=0.9\textwidth]{fig/channel_results.png}
 \caption{Velocity magnitude (top part) and pressure (bottom part) at end of simulation.}
 \label{fig:channel_results}
 \end{figure}
@@ -1136,7 +1136,7 @@ a specified value, here \verb"`eps_periodic'" in the \verb"TIME_PARAMS" dictiona
 
 \begin{figure}
 \centering
-\includegraphics[width=0.7\textwidth]{fig/syspul_setup.png}
+\includegraphics[width=0.65\textwidth]{fig/syspul_setup.png}
 \caption{0D heart, systemic and pulmonary circulation, problem setup.}
 \label{fig:syspul_setup}
 \end{figure}
@@ -1160,7 +1160,7 @@ The solution is depicted in fig. \ref{fig:syspul_results}, showing the time cour
 
 \begin{figure}
 \centering
-\includegraphics[width=0.9\textwidth]{fig/syspul_results.png}
+\includegraphics[width=1.0\textwidth]{fig/syspul_results.png}
 \caption{A. Left heart and systemic pressures over time. B. Right heart and pulmonary pressures over time. C. Left and right ventricular and atrial volumes over time. D. Left and right ventricular pressure-volume relationships of periodic (5th) cycle.}
 \label{fig:syspul_results}
 \end{figure}
@@ -1185,7 +1185,7 @@ increase.
 
 \begin{figure}
 \centering
-\includegraphics[width=0.7\textwidth]{fig/heart_syspul_setup.png}
+\includegraphics[width=0.65\textwidth]{fig/heart_syspul_setup.png}
 \caption{Generic 3D ventricular heart model coupled to a closed-loop systemic and pulmonary circulation model.}
 \label{fig:heart_syspul_setup}
 \end{figure}
@@ -1218,7 +1218,7 @@ in ventricular volume?
 
 \begin{figure}
 \centering
-\includegraphics[width=0.9\textwidth]{fig/heart_syspul_results.png}
+\includegraphics[width=1.0\textwidth]{fig/heart_syspul_results.png}
 \caption{A. Left heart and systemic pressures over time. B. Left and right ventricular and atrial volumes over time. C. Left and right ventricular pressure-volume relationships. D. Snapshot of heart deformation at end-systole, color indicates displacement magnitude.}
 \label{fig:heart_syspul_results}
 \end{figure}
