+++ docs/source/generated/ambit_fe.rst
@@ -1,23 +1,0 @@
-﻿ambit\_fe
-=========
-
-.. automodule:: ambit_fe
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   
-
-
-
