+++ docs/requirements.txt
@@ -1,4 +1,4 @@
 myst_parser
 sphinx-rtd-theme
 sphinxcontrib-bibtex
-ambit_fe
+ambit_fe --no-deps
