### 2023-11-20 b2239af
Update requirements for API doc
