+++ docs/source/documentation.rst
@@ -76,6 +76,12 @@ Preface
 | Examples of input files for the respective problem types can be found
   in the folder ``demos`` (with detailed setup descriptions) or amogst
   the test cases in the folder ``tests``.
+| This documentation is structured as follows. In sec.
+  `2 <#sec:inst>`__, instructions on how to install and use Ambit are
+  given. The relevant supported physics models are described in sec.
+  `3 <#sec:phys>`__. Demos are presented in sec. `4 <#sec:demos>`__.
+
+.. _sec:inst:
 
 Installation
 ============
@@ -110,20 +116,25 @@ dolfinx mixed branch):
 
    docker pull ghcr.io/marchirschvogel/ambit:devenv
 
-.. _sec:solid:
+.. _sec:phys:
+
+Physics Models
+==============
+
+.. _subsec:solid:
 
 Solid mechanics
-===============
+---------------
 
-| – Example: Sec. `7.1 <#subsec:demos:solid>`__ and ``demos/solid``
+| – Example: Sec. `4.1 <#subsec:demos:solid>`__ and ``demos/solid``
 | – Problem type: ``solid``
 | – Solid mechanics are formulated in a Total Lagrangian frame
 
 Strong form
------------
+~~~~~~~~~~~
 
 Displacement-based
-~~~~~~~~~~~~~~~~~~
+^^^^^^^^^^^^^^^^^^
 
 – Primary variable: displacement :math:`\boldsymbol{u}`
 
@@ -137,7 +148,7 @@ Displacement-based
    \boldsymbol{v}(\boldsymbol{x}_{0},0) &= \hat{\boldsymbol{v}}_{0}(\boldsymbol{x}_{0}) &&\text{in} \; \mathit{\Omega}_{0}, \label{eq:ini_v}\end{aligned}
 
 Incompressible mechanics
-~~~~~~~~~~~~~~~~~~~~~~~~
+^^^^^^^^^^^^^^^^^^^^^^^^
 
 – Primary variables: displacement :math:`\boldsymbol{u}` and pressure
 :math:`p`
@@ -158,12 +169,12 @@ with velocity and acceleration
 respectively
 
 Weak form
----------
+~~~~~~~~~
 
 .. _displacement-based-1:
 
 Displacement-based
-~~~~~~~~~~~~~~~~~~
+^^^^^^^^^^^^^^^^^^
 
 | – Primary variable: displacement :math:`\boldsymbol{u}`
 | – Principal of Virtual Work:
@@ -246,7 +257,7 @@ Displacement-based
    \left. \boldsymbol{\mathsf{K}}_{uu} \right|_{n+1}^{k} \Delta\boldsymbol{\mathsf{u}}_{n+1}^{k+1}=-\left. \boldsymbol{\mathsf{r}}_{u} \right|_{n+1}^{k} \label{eq:lin_sys_solid}\end{aligned}
 
 Incompressible mechanics: 2-field displacement and pressure variables
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
[…]
