+++ docs/source/ambit_doc.tex
@@ -159,7 +159,9 @@ Avenues for future functionality include cardiac electrophysiology, scalar trans
 
 In the following, a brief description of the supported problem types is given, including the strong and weak form of the underlying equations as well as the discrete assembled systems that are solved.\\
 
-Examples of input files for the respective problem types can be found in the folder \verb.demos. (with detailed setup descriptions) or amogst the test cases in the folder \verb.tests..
+Examples of input files for the respective problem types can be found in the folder \verb"demos" (with detailed setup descriptions) or amogst the test cases in the folder \verb"tests".\\
+
+This documentation is structured as follows. In sec. \ref{sec:inst}, instructions on how to install and use Ambit are given. The relevant supported physics models are described in sec. \ref{sec:phys}. Demos are presented in sec. \ref{sec:demos}.
 
 %\begin{figure}[!htp]
 %\centering
@@ -167,7 +169,7 @@ Examples of input files for the respective problem types can be found in the fol
 %\caption{Code design}\label{fig:code_design}
 %\end{figure}
 
-\section{Installation}
+\section{Installation}\label{sec:inst}
 
 In order to use Ambit, you need to install FEniCSx (https://github.com/FEniCS/dolfinx\#installation) (latest Ambit-compatible dolfinx development version dates to 19 Aug 2023).\\
 
@@ -191,7 +193,9 @@ If a Docker image for development is desired, the following image contains all d
 docker pull ghcr.io/marchirschvogel/ambit:devenv
 \end{verbatim}
 
-\section{Solid mechanics}\label{sec:solid}
+\section{Physics Models}\label{sec:phys}
+
+\subsection{Solid mechanics}\label{subsec:solid}
 
 -- Example: Sec. \ref{subsec:demos:solid} and \verb"demos/solid"\\
 
@@ -199,9 +203,9 @@ docker pull ghcr.io/marchirschvogel/ambit:devenv
 
 -- Solid mechanics are formulated in a Total Lagrangian frame
 
-\subsection{Strong form}
+\subsubsection{Strong form}
 
-\subsubsection{Displacement-based}
+\paragraph{Displacement-based}
 -- Primary variable: displacement $\bs{u}$
 \begin{align}
 \bs{\nabla}_{0} \cdot \bs{P}(\bs{u},\bs{v}(\bs{u})) + \hat{\bs{b}}_{0} &= \rho_{0} \bs{a}(\bs{u}) &&\text{in} \; \mathit{\Omega}_{0} \times [0, T], \label{eq:divP} \\
[…]
