### 2023-11-20 marchirschvogel/ambit#5 @marchirschvogel
Thanks, I've made use of the sphinx-apidoc and added docstrings to the main function the user calls. Strangely, the docstrings are only shown in my local sphinx html build, but not on Readthedocs. I'll check what the issue could be.

Also, I've moved the demos description to the documentation.
