### 2023-11-20 marchirschvogel/ambit#7 @finsberg
[JOSS REVIEW] State of the field
The [joss paper](https://github.com/marchirschvogel/ambit/blob/master/joss/paper.md) is missing a section about the *State of the  field* where you compare your software to other commonly-used packages.

This is related to https://github.com/openjournals/joss-reviews/issues/5744
