### 2023-11-21 f1c9587
Doc and joss paper update
