+++ joss/paper.bib
@@ -1,3 +1,14 @@
+@article{africa2022-lifex,
+author={P. C. Africa},
+title={{lifex}: A flexible, high performance library for the numerical solution of complex finite element problems},
+journal={SoftwareX},
+year={2022},
+pages={101252},
+volume={},
+number={},
+doi={10.1016/j.softx.2022.101252}
+}
+
 @article{arthurs2016,
 author={C. J. Arthurs and K. D. Lau and K. N. Asrress and S. R. Redwood and C. A. Figueroa},
 title={A mathematical model of coronary blood flow control: simulation of patient-specific three-dimensional hemodynamics during exercise},
@@ -7,6 +18,34 @@ pages={H1242--H1258},
 volume={310},
 number={9},
 doi={10.1152/ajpheart.00517.2015}
+}
+
+@article{arthurs2021-crimson,
+author={C. J. Arthurs and R. Khlebnikov and A. Melville and M. Mar{\v{c}}an and A. Gomez and D. Dillon-Murphy and F. Cuomo and M. Silva Vieira 1 and
+J. Schollenberger and S. R. Lynch and C. Tossas-Betancourt and K. Iyer and S. Hopper and E. Livingston and P. Youssefi and A. Noorani 1 and
+S. Ben Ahmed and F. J. H. Nauta and T. M. J. van Bakel and Y. Ahmed and P. A. J. van Bakel and J. Mynard and P. Di Achille and H. Gharahi and 
+K. D. Lau and V. Filonova and M. Aguirre and N. Nama and N. Xiao and S. Baek and K. Garikipati and O. Sahni and D. Nordsletten and C. A. Figueroa},
+title={{CRIMSON}: An open-source software framework for cardiovascular integrated modelling and simulation},
+journal={{PLoS} Comput Biol},
+year={2021},
+pages={e1008881},
+volume={17},
+number={5},
+doi={10.1371/journal.pcbi.1008881}
+}
+
+@techreport{balay2022-petsc,
+author={S. Balay and S. Abhyankar and M. F. Adams and S. Benson and J. Brown
+    and P. Brune and K. Buschelman and E. Constantinescu and L. Dalcin and A. Dener
+    and V. Eijkhout and W. D. Gropp and V. Hapla and T. Isaac and P. Jolivet
+    and D. Karpeev and D. Kaushik and M. G. Knepley and F. Kong and S. Kruger
+    and D. A. May and L. C. McInnes and R. T. Mills and L. Mitchell and T. Munson
+    and J. E. Roman and K. Rupp and P. Sanan and J. Sarich and B. F. Smith
+    and S. Zampini and H. Zhang and H. Zhang and J. Zhang},
+title={{PETSc/TAO} Users Manual},
+institution={Argonne National Laboratory},
+number={ANL-21/39 - Revision 3.17},
+year={2022},
 }
 
 @article{bonini2022,
@@ -40,6 +79,28 @@ pages={625--662},
 volume={98},
 number={9},
 doi={10.1002/nme.4668}
+}
+
+@article{finsberg2019-pulse,
+author={H. N. T. Finsberg},
+title={{pulse}: A python package based on {FEniCS} for solving problems in cardiac mechanics},
+journal={The Journal of Open Source Software},
+year={2019},
+pages={1539},
+volume={4},
+number={41},
+doi={10.21105/joss.01539}
+}
+
+@article{finsberg2023-simcardems,
+author={H. N. T. Finsberg and I. G. M. van Herck and C. Daversin-Catty and H. Arevalo and S. Wall},
+title={{simcardems}: A {FEniCS}-based cardiac electro-mechanics solver},
[…]
