+++ joss/paper.md
@@ -24,7 +24,7 @@ bibliography: paper.bib
 
 # Summary
 
-Ambit is an open-source software tool written in Python for parallel multi-physics simulations focusing on -- but not limited to -- cardiac mechanics. Amongst others, it contains re-implementations and generalizations of methods developed by the author for his PhD thesis [@hirschvogel2018]. Ambit makes use of the open-source finite element library [FEniCS/dolfinx](https://fenicsproject.org) [@logg2012] along with the linear algebra package [PETSc](https://petsc.org) [@petsc-user-ref]. It is constantly updated to ensure compatibility with a recent dolfinx development version, hence guaranteeing a state-of-the-art finite element and linear algebra backend.
+Ambit is an open-source software tool written in Python for parallel multi-physics simulations focusing on -- but not limited to -- cardiac mechanics. Amongst others, it contains re-implementations and generalizations of methods developed by the author for his PhD thesis [@hirschvogel2019disspub]. Ambit makes use of the open-source finite element library [FEniCS/dolfinx](https://fenicsproject.org) [@logg2012] along with the linear algebra package [PETSc](https://petsc.org) [@balay2022-petsc]. It is constantly updated to ensure compatibility with a recent dolfinx development version, hence guaranteeing a state-of-the-art finite element and linear algebra backend.
 
 Ambit is designed such that the user only needs to provide input files that define parameters through Python dictionaries, hence no programming or in-depth knowledge of any library-specific syntax is required.
 
@@ -46,7 +46,9 @@ Avenues for future functionality include cardiac electrophysiology, scalar trans
 
 Cardiovascular disease entities are the most prevalent ones in the industrialized world [@dimmeler2011; @luepker2011] and a leading cause of death worldwise. Therefore, models that promote a better understanding of cardiac diseases and their progression represent a valuable tool to guide or assist therapy planning, support device dimensioning and design [@hirschvogel2019], or help predict intervention planning [@bonini2022; @taylor2013].
 
-Ambit represents a complete open-source code for simulating cardiac mechanics, encompassing advanced structural mechanics of the myocardium, ventricular fluid dynamics, reduced-dimensional blood flow and multi-physics coupling. Therefore, a wide range of mechanical problems can be simulated, and the code structure allows easy and straightforward extensibility (e.g. implementations of new constitutive models). Due to its simple design in terms of clearly organized input files, Ambit is easy to use and hence represents a valuable tool for novices or advanced researchers who want to address cardiovascular mechanics problems.
+Software packages that are tailored towards cardiac modeling have been provided to the open source community. Amogst them are the cardiovascular FSI solver svFSI [@zhu2022-svfsi] along with SimVascular [@updegrove2017-simvascular], provding a full medical image-to-model pipeline, as well as FEBio [@maas2012-febio], focusing on advanced structural mechanics of soft tissue. FEniCS-based open-source solvers are pulse [@finsberg2019-pulse] for cardiac solid mechanics and cbcbeat [@rognes2017-cbcbeat] for cardiac electrophysiology, both fused to a combined toolkit for cardiac electro-mechanics named simcardems [@finsberg2023-simcardems]. Another framework for simulating cardiac electrophysiology is openCARP [@plank2021-opencarp], and CRIMSON [@arthurs2021-crimson] provides a modeling suit for 3D and reduced-dimensional hemodynamics in arteries. A general purpose library that provides the building blocks for cardiac modeling is lifex [@africa2022-lifex].
+
[…]
