+++ joss/paper.md
@@ -44,11 +44,11 @@ Avenues for future functionality include cardiac electrophysiology, scalar trans
 
 # Statement of need
 
-Cardiovascular disease entities are the most prevalent ones in the industrialized world [@dimmeler2011; @luepker2011] and a leading cause of death worldwise. Therefore, models that promote a better understanding of cardiac diseases and their progression represent a valuable tool to guide or assist therapy planning, support device dimensioning and design [@hirschvogel2019], or help predict intervention planning [@bonini2022; @taylor2013].
+Cardiovascular disease entities are the most prevalent ones in the industrialized world [@dimmeler2011; @luepker2011] and a leading cause of death worldwide. Therefore, models that promote a better understanding of cardiac diseases and their progression represent a valuable tool to guide or assist therapy planning, support device dimensioning and design [@hirschvogel2019], or help predict intervention planning [@bonini2022; @taylor2013].
 
-Software packages that are tailored towards cardiac modeling have been provided to the open source community. Amogst them are the cardiovascular FSI solver svFSI [@zhu2022-svfsi] along with SimVascular [@updegrove2017-simvascular], provding a full medical image-to-model pipeline, as well as FEBio [@maas2012-febio], focusing on advanced structural mechanics of soft tissue. FEniCS-based open-source solvers are pulse [@finsberg2019-pulse] for cardiac solid mechanics and cbcbeat [@rognes2017-cbcbeat] for cardiac electrophysiology, both fused to a combined toolkit for cardiac electro-mechanics named simcardems [@finsberg2023-simcardems]. Another framework for simulating cardiac electrophysiology is openCARP [@plank2021-opencarp], and CRIMSON [@arthurs2021-crimson] provides a modeling suit for 3D and reduced-dimensional hemodynamics in arteries. A general purpose library that provides the building blocks for cardiac modeling is lifex [@africa2022-lifex].
+Software packages that are tailored towards cardiac modeling have been provided to the open source community. Amongst them are the cardiovascular FSI solver svFSI [@zhu2022-svfsi] along with SimVascular [@updegrove2017-simvascular], providing a full medical image-to-model pipeline, as well as FEBio [@maas2012-febio], focusing on advanced structural mechanics of soft tissue. FEniCS-based open-source solvers are pulse [@finsberg2019-pulse] for cardiac solid mechanics and cbcbeat [@rognes2017-cbcbeat] for cardiac electrophysiology, both fused to a combined toolkit for cardiac electro-mechanics named simcardems [@finsberg2023-simcardems]. Another framework for simulating cardiac electrophysiology is openCARP [@plank2021-opencarp], and CRIMSON [@arthurs2021-crimson] provides a modeling suit for 3D and reduced-dimensional hemodynamics in arteries. A general purpose library that provides the building blocks for cardiac modeling is lifex [@africa2022-lifex].
 
-Ambit represents a complete open-source code for simulating cardiac mechanics, encompassing advanced structural mechanics of the myocardium, ventricular fluid dynamics, reduced-dimensional blood flow and multi-physics coupling. Therefore, a wide range of mechanical problems can be simulated, and the code structure allows easy and straightforward extensibility (e.g. implementations of new constitutive models) without the need for low-level library-specific syntax or advanced programming. Due to its simple design in terms of clearly organized input files, Ambit is easy to use and hence represents a valuable tool for novices or advanced researchers who want to address cardiovascular mechanics problems.
[…]
