### 2023-11-21 marchirschvogel/ambit#7 @marchirschvogel
Thanks, I've updated the paper.md file accordingly, and included a paragraph under the "Statement of need" section on other open source solvers tailored towards cardiac mechanics. Please let me know if you feel that I've forgotten to mention any other important package(s).
