+++ joss/paper.bib
@@ -247,20 +247,6 @@ number={1},
 doi={10.1115/1.4005694}
 }
 
-@techreport{petsc2022,
-author={S. Balay and S. Abhyankar and M. F. Adams and S. Benson and J. Brown
-    and P. Brune and K. Buschelman and E. Constantinescu and L. Dalcin and A. Dener
-    and V. Eijkhout and W. D. Gropp and V. Hapla and T. Isaac and P. Jolivet
-    and D. Karpeev and D. Kaushik and M. G. Knepley and F. Kong and S. Kruger
-    and D. A. May and L. C. McInnes and R. T. Mills and L. Mitchell and T. Munson
-    and J. E. Roman and K. Rupp and P. Sanan and J. Sarich and B. F. Smith
-    and S. Zampini and H. Zhang and H. Zhang and J. Zhang},
-title={{PETSc/TAO} Users Manual},
-institution={Argonne National Laboratory},
-number={ANL-21/39 - Revision 3.17},
-year={2022},
-}
-
 @article{plank2021-opencarp,
 author={G. Plank and A. Loewe and A. Neic and C. Augustin and Y.-L. Huang and M. A. F. Gsell and E. Karabelas and M. Nothstein and A. J. Prassl and J. S{\'a}nchez and G. Seemann and E. J. Vigmond},
 title={The {openCARP} Simulation Environment for Cardiac Electrophysiology},
