### 2023-11-21 #5744 @finsberg
@Kevin-Mattheus-Moerman I am done with my review of ambit, and can recommend it for publication in JOSS.
