+++ docs/source/ambit_doc.tex
@@ -559,27 +559,31 @@ r_p \leftarrow r_p &+ \frac{1}{\rho}\int\limits_{\Om_0} J\,d_1\,((\bs{\nabla}_{0
 
 -- Problem type: \verb"flow0d"\\
 
+-- 0D model concentrated elements are resistances ($R$), impedances ($Z$, technically are resistances), compliances ($C$), inertances ($L$ or $I$), and elastances ($E$)\\
+
+-- 0D variables are pressures ($p$), fluxes ($q$ or $Q$), or volumes ($V$)
+
 \subsubsection{2-element Windkessel}
 
-- Model type : \verb"2elwindkessel"\\
+-- Model type : \verb"2elwindkessel"\\
 
 \subsubsection{4-element Windkessel (inertance parallel to impedance)}
 
-- Model type : \verb"4elwindkesselLpZ"\\
+-- Model type : \verb"4elwindkesselLpZ"\\
 
 \subsubsection{4-element Windkessel (inertance serial to impedance)}
 
-- Model type : \verb"4elwindkesselLsZ"\\
+-- Model type : \verb"4elwindkesselLsZ"\\
 
 \subsubsection{In-outflow CRL link}
 
-- Model type : \verb"CRLinoutlink"\\
+-- Model type : \verb"CRLinoutlink"\\
 
 \subsubsection{Systemic and pulmonary circulation}
 
-- Model type : \verb"syspul"\\
+-- Model type : \verb"syspul"\\
 
-- Allows to link in a coronary flow model
+-- Allows to link in a coronary flow model
 
 \begin{align}
 &\text{left heart and systemic circulation} && \nonumber\\
@@ -712,7 +716,7 @@ q(p-p_{\mathrm{open}}) = \begin{cases} c A_{\mathrm{o}} \sqrt{p-p_{\mathrm{open}
 
 \subsubsection{Systemic and pulmonary circulation, including capillary flow}
 
-- Model type : \verb"syspulcap"\\
+-- Model type : \verb"syspulcap", cf. \cite{hirschvogel2019disspub}, p. 51ff.\\
 
 % left heart and systemic circulation:
 \begin{align}
@@ -757,7 +761,9 @@ Q_{\mathrm{v}}^{r} := -\frac{\mathrm{d}V_{\mathrm{v}}^{r}}{\mathrm{d}t}\nonumber
 
 \subsubsection{Systemic and pulmonary circulation, including capillary and coronary flow}
 
-- Model type : \verb"syspulcapcor"\\
+-- Model type : \verb"syspulcapcor"\\
+
+-- Variant of \verb"syspulcap", with coronaries branching off directly after aortic valve
 
 % left heart and systemic circulation:
 \begin{align}
@@ -803,11 +809,11 @@ Q_{\mathrm{at}}^{r} := -\frac{\mathrm{d}V_{\mathrm{at}}^{r}}{\mathrm{d}t}, \qqua
 Q_{\mathrm{v}}^{r} := -\frac{\mathrm{d}V_{\mathrm{v}}^{r}}{\mathrm{d}t} \nonumber
 \end{align}
 
[…]
