+++ docs/source/documentation.rst
@@ -661,32 +661,38 @@ elements (without stabilization)
 
 | – Example: Sec. `4.3 <#subsec:demos:flow0d>`__ and ``demos/flow0d``
 | – Problem type: ``flow0d``
+| – 0D model concentrated elements are resistances (:math:`R`),
+  impedances (:math:`Z`, technically are resistances), compliances
+  (:math:`C`), inertances (:math:`L` or :math:`I`), and elastances
+  (:math:`E`)
+| – 0D variables are pressures (:math:`p`), fluxes (:math:`q` or
+  :math:`Q`), or volumes (:math:`V`)
 
 2-element Windkessel
 ~~~~~~~~~~~~~~~~~~~~
 
-| - Model type : ``2elwindkessel``
+| – Model type : ``2elwindkessel``
 
 4-element Windkessel (inertance parallel to impedance)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-| - Model type : ``4elwindkesselLpZ``
+| – Model type : ``4elwindkesselLpZ``
 
 4-element Windkessel (inertance serial to impedance)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-| - Model type : ``4elwindkesselLsZ``
+| – Model type : ``4elwindkesselLsZ``
 
 In-outflow CRL link
 ~~~~~~~~~~~~~~~~~~~
 
-| - Model type : ``CRLinoutlink``
+| – Model type : ``CRLinoutlink``
 
 Systemic and pulmonary circulation
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-| - Model type : ``syspul``
-| - Allows to link in a coronary flow model
+| – Model type : ``syspul``
+| – Allows to link in a coronary flow model
 
 .. math::
 
@@ -859,7 +865,8 @@ with
 Systemic and pulmonary circulation, including capillary flow
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-| - Model type : ``syspulcap``
+| – Model type : ``syspulcap``, cf.
+  :cite:p:`hirschvogel2019disspub`, p. 51ff.
 | 
 
   .. math::
@@ -907,30 +914,31 @@ with:
 Systemic and pulmonary circulation, including capillary and coronary flow
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-| - Model type : ``syspulcapcor``
-| 
+| – Model type : ``syspulcapcor``
+| – Variant of ``syspulcap``, with coronaries branching off directly
+  after aortic valve
 
-  .. math::
+.. math::
 
-     \begin{aligned}
[…]
