### 2023-12-08 fd2ced6
Add option to compute and output strain energy for solid and reduced solid; minor fixes
