+++ src/ambit_fe/fluid/fluid_main.py
@@ -72,7 +72,7 @@ class FluidmechanicsProblem(problem_base):
 
         # collect domain data
         self.rho = []
-        for n, N in enumerate(self.domain_ids):
+        for n, M in enumerate(self.domain_ids):
             # data for inertial forces: density
             self.rho.append(self.constitutive_models['MAT'+str(n+1)]['inertia']['rho'])
 
@@ -393,33 +393,33 @@ class FluidmechanicsProblem(problem_base):
         self.deltaW_int, self.deltaW_int_old = ufl.as_ufl(0), ufl.as_ufl(0)
         self.deltaW_p,   self.deltaW_p_old   = [], []
 
-        for n, N in enumerate(self.domain_ids):
+        for n, M in enumerate(self.domain_ids):
 
             if self.num_dupl==1: j=0
             else: j=n
 
             # kinetic virtual power
             if self.fluid_governing_type=='navierstokes_transient':
-                self.deltaW_kin     += self.vf.deltaW_kin_navierstokes_transient(self.acc, self.v, self.rho[n], self.dx(N), w=self.alevar['w'], F=self.alevar['Fale'])
-                self.deltaW_kin_old += self.vf.deltaW_kin_navierstokes_transient(self.a_old, self.v_old, self.rho[n], self.dx(N), w=self.alevar['w_old'], F=self.alevar['Fale_old'])
+                self.deltaW_kin     += self.vf.deltaW_kin_navierstokes_transient(self.acc, self.v, self.rho[n], self.dx(M), w=self.alevar['w'], F=self.alevar['Fale'])
+                self.deltaW_kin_old += self.vf.deltaW_kin_navierstokes_transient(self.a_old, self.v_old, self.rho[n], self.dx(M), w=self.alevar['w_old'], F=self.alevar['Fale_old'])
             elif self.fluid_governing_type=='navierstokes_steady':
-                self.deltaW_kin     += self.vf.deltaW_kin_navierstokes_steady(self.v, self.rho[n], self.dx(N), w=self.alevar['w'], F=self.alevar['Fale'])
-                self.deltaW_kin_old += self.vf.deltaW_kin_navierstokes_steady(self.v_old, self.rho[n], self.dx(N), w=self.alevar['w_old'], F=self.alevar['Fale_old'])
+                self.deltaW_kin     += self.vf.deltaW_kin_navierstokes_steady(self.v, self.rho[n], self.dx(M), w=self.alevar['w'], F=self.alevar['Fale'])
+                self.deltaW_kin_old += self.vf.deltaW_kin_navierstokes_steady(self.v_old, self.rho[n], self.dx(M), w=self.alevar['w_old'], F=self.alevar['Fale_old'])
             elif self.fluid_governing_type=='stokes_transient':
-                self.deltaW_kin     += self.vf.deltaW_kin_stokes_transient(self.acc, self.v, self.rho[n], self.dx(N), w=self.alevar['w'], F=self.alevar['Fale'])
-                self.deltaW_kin_old += self.vf.deltaW_kin_stokes_transient(self.a_old, self.v_old, self.rho[n], self.dx(N), w=self.alevar['w_old'], F=self.alevar['Fale_old'])
+                self.deltaW_kin     += self.vf.deltaW_kin_stokes_transient(self.acc, self.v, self.rho[n], self.dx(M), w=self.alevar['w'], F=self.alevar['Fale'])
+                self.deltaW_kin_old += self.vf.deltaW_kin_stokes_transient(self.a_old, self.v_old, self.rho[n], self.dx(M), w=self.alevar['w_old'], F=self.alevar['Fale_old'])
             elif self.fluid_governing_type=='stokes_steady':
[…]
