### 2023-12-09 04bb33f
Enable solid mechanics in 2 dimensions
