### 2023-12-09 cb4844e
Flexibilize membrane material choice; add option for immersed (internal) membrane
