### 2023-12-10 04d2259
Fix: Use potential to output strain energy
