### 2023-12-11 52f168c
Add option to compute and output internal power
