+++ src/ambit_fe/fluid/fluid_main.py
@@ -481,8 +481,8 @@ class FluidmechanicsProblem(problem_base):
                     self.io.readfunction(h0_func, self.bc_dict['membrane'][nm]['params']['h0']['field'])
                     self.wallfields.append(h0_func)
 
-            w_membrane, self.idmem, self.bstress, self.bstrainenergy = self.bc.membranesurf_bcs(self.bc_dict['membrane'], self.ufluid, self.v, self.acc, self.bmeasures, ivar=self.internalvars, wallfields=self.wallfields)
-            w_membrane_old, _, _, _                                  = self.bc.membranesurf_bcs(self.bc_dict['membrane'], self.uf_old, self.v_old, self.a_old, self.bmeasures, ivar=self.internalvars_old, wallfields=self.wallfields)
+            w_membrane, self.idmem, self.bstress, self.bstrainenergy, self.bintpower = self.bc.membranesurf_bcs(self.bc_dict['membrane'], self.ufluid, self.v, self.acc, self.bmeasures, ivar=self.internalvars, wallfields=self.wallfields)
+            w_membrane_old, _, _, _, _                                               = self.bc.membranesurf_bcs(self.bc_dict['membrane'], self.uf_old, self.v_old, self.a_old, self.bmeasures, ivar=self.internalvars_old, wallfields=self.wallfields)
 
         w_neumann_prestr, self.deltaW_prestr_kin = ufl.as_ufl(0), ufl.as_ufl(0)
         if self.prestress_initial or self.prestress_initial_only:
@@ -644,10 +644,32 @@ class FluidmechanicsProblem(problem_base):
         self.tau_a.interpolate(tau_a_proj)
 
 
-    # computes the total strain energy of a membrane (reduced) solid model
-    def compute_strain_energy_membrane(self, N, t):
+    # computes the fluid's total internal power
+    def compute_power(self, N, t):
 
-        se_mem_all = ufl.as_ufl(0)
+        ip_all = ufl.as_ufl(0)
+        for n, M in enumerate(self.domain_ids):
+            if self.num_dupl==1: j=0
+            else: j=n
+            ip_all += ufl.inner(self.ma[n].sigma(self.v, self.p_[j], F=self.alevar['Fale']), self.ki.gamma(self.v, F=self.alevar['Fale'])) * self.dx(M)
+
+        ip = fem.assemble_scalar(fem.form(ip_all))
+        ip = self.comm.allgather(ip)
+        internal_power = sum(ip)
+
+        if self.comm.rank == 0:
+            if self.io.write_results_every > 0 and N % self.io.write_results_every == 0:
+                if np.isclose(t,self.dt): mode='wt'
+                else: mode='a'
+                fp = open(self.io.output_path+'/results_'+self.simname+'_internalpower.txt', mode)
+                fp.write('%.16E %.16E\n' % (t,internal_power))
+                fp.close()
+
+
+    # computes the total strain energy and internal power of a membrane (reduced) solid model
+    def compute_strain_energy_power_membrane(self, N, t):
+
+        se_mem_all, ip_mem_all = ufl.as_ufl(0), ufl.as_ufl(0)
         for nm in range(len(self.bc_dict['membrane'])):
 
             try: internal = self.bc_dict['membrane'][nm]['internal']
@@ -657,21 +679,31 @@ class FluidmechanicsProblem(problem_base):
                 try: fcts = self.bc_dict['membrane'][nm]['facet_side']
                 except: fcts = '+'
                 se_mem_all += (self.bstrainenergy[nm])(fcts) * self.dS(self.idmem[nm])
+                ip_mem_all += (self.bintpower[nm])(fcts) * self.dS(self.idmem[nm])
             else:
[…]
