+++ src/ambit_fe/boundaryconditions.py
@@ -373,7 +373,7 @@ class boundary_cond():
     # set membrane surface BCs
     def membranesurf_bcs(self, bcdict, u, v, a, ds_, ivar=None, wallfields=[]):
 
-        w, idmem, bstress, bstrainenergy = ufl.as_ufl(0), [], [], []
+        w, idmem, bstress, bstrainenergy, bintpower = ufl.as_ufl(0), [], [], [], []
 
         mi=0
         for m in bcdict:
@@ -406,13 +406,14 @@ class boundary_cond():
                 idmem.append(m['id'][i])
 
                 w += self.vf.deltaW_ext_membrane(self.ki.F(u), self.ki.Fdot(v), a, m['params'], ds_[dind](m['id'][i]), ivar=ivar, fibfnc=self.ff, wallfield=wallfield, fcts=fcts)
-                bstr, bse = self.vf.deltaW_ext_membrane(self.ki.F(u), self.ki.Fdot(v), a, m['params'], ds_[dind](m['id'][i]), ivar=ivar, fibfnc=self.ff, se=True, wallfield=wallfield, fcts=fcts)
+                bstr, bse, bip = self.vf.deltaW_ext_membrane(self.ki.F(u), self.ki.Fdot(v), a, m['params'], ds_[dind](m['id'][i]), ivar=ivar, fibfnc=self.ff, wallfield=wallfield, fcts=fcts, returnquantity='stress_energy_power')
                 bstress.append(bstr)
                 bstrainenergy.append(bse)
+                bintpower.append(bip)
 
             mi+=1
 
-        return w, idmem, bstress, bstrainenergy
+        return w, idmem, bstress, bstrainenergy, bintpower
 
 
     # set body forces (technically, no "boundary" conditions, since acting on a volume element... but implemented here for convenience)
