+++ tests/test_fluid_taylorhood_cylinder.py
@@ -25,7 +25,7 @@ def test_main():
                            'mesh_boundary'         : basepath+'/input/cylinder-quad_boundary.xdmf',
                            'write_results_every'   : 1,
                            'output_path'           : basepath+'/tmp/',
-                           'results_to_write'      : ['velocity','pressure','acceleration','cauchystress','fluiddisplacement'],
+                           'results_to_write'      : ['velocity','pressure','acceleration','cauchystress','fluiddisplacement','internalpower'],
                            'simname'               : 'fluid_taylorhood_cylinder'}
 
     SOLVER_PARAMS_FLUID = {'solve_type'            : 'direct',

[…]
