+++ src/ambit_fe/solid/solid_main.py
@@ -1009,6 +1009,8 @@ class SolidmechanicsProblem(problem_base):
                 file = list(m.values())[0].replace('*',str(N))
                 func = list(m.keys())[0]
                 self.io.readfunction(func, file)
+                sc = m['scale']
+                if sc > 1.0: func.vector.scale(sc)
 
 
     def evaluate_post_solve(self, t, N):
