+++ src/ambit_fe/ale/ale_main.py
@@ -284,6 +284,8 @@ class AleProblem(problem_base):
                 file = list(m.values())[0].replace('*',str(N))
                 func = list(m.keys())[0]
                 self.io.readfunction(func, file)
+                sc = m['scale']
+                if sc > 1.0: func.vector.scale(sc)
 
 
     def evaluate_post_solve(self, t, N):
