+++ src/ambit_fe/fluid/fluid_main.py
@@ -1044,6 +1044,8 @@ class FluidmechanicsProblem(problem_base):
                 file = list(m.values())[0].replace('*',str(N))
                 func = list(m.keys())[0]
                 self.io.readfunction(func, file)
+                sc = m['scale']
+                if sc > 1.0: func.vector.scale(sc)
 
 
     def evaluate_post_solve(self, t, N):
