+++ src/ambit_fe/ioroutines.py
@@ -23,7 +23,7 @@ from .mathutils import spectral_decomposition_3x3
 
 class IO:
 
-    def __init__(self, io_params, entity_maps, comm):
+    def __init__(self, io_params, fem_params, entity_maps, comm):
 
         ioparams.check_params_io(io_params)
 
@@ -42,6 +42,8 @@ class IO:
         except: self.mesh_edge = None
         try: self.mesh_point = io_params['mesh_point']
         except: self.mesh_point = None
+
+        self.quad_degree = fem_params['quad_degree']
 
         try: self.fiber_data = io_params['fiber_data']
         except: self.fiber_data = []
@@ -186,6 +188,22 @@ class IO:
 
         else:
             raise AttributeError("Your mesh seems to be 1D! Not supported!")
+
+
+    # create domain and boundary integration measures
+    def create_integration_measures(self, msh):
+
+        # create domain and boundary integration measures
+        if self.mt_d_master is not None:
+            self.dx = ufl.Measure("dx", domain=msh, subdomain_data=self.mt_d_master, metadata={'quadrature_degree': self.quad_degree})
+        else:
+            self.dx = ufl.Measure("dx", domain=msh, metadata={'quadrature_degree': self.quad_degree})
+
+        self.ds = ufl.Measure("ds", domain=msh, subdomain_data=self.mt_b1_master, metadata={'quadrature_degree': self.quad_degree})
+        self.de = ufl.Measure("ds", domain=msh, subdomain_data=self.mt_b2_master, metadata={'quadrature_degree': self.quad_degree})
+        self.dS = ufl.Measure("dS", domain=msh, subdomain_data=self.mt_b1_master, metadata={'quadrature_degree': self.quad_degree})
+        # self.de = self.io.de
+        self.bmeasures = [self.ds,self.de,self.dS]
 
 
     # some mesh data that we might wanna use in some problems...
@@ -393,12 +411,12 @@ class IO_solid(IO):
                         u_out.interpolate(pb.u)
                         self.resultsfiles[res].write_function(u_out, indicator)
                     elif res=='velocity': # passed in v is not a function but form, so we have to project
-                        self.v_proj = project(pb.vel, pb.V_u, pb.dx, domids=pb.domain_ids, nm="Velocity", comm=self.comm) # class variable for testing
+                        self.v_proj = project(pb.vel, pb.V_u, self.dx, domids=pb.domain_ids, nm="Velocity", comm=self.comm) # class variable for testing
                         v_out = fem.Function(self.V_out_vector, name=self.v_proj.name)
                         v_out.interpolate(self.v_proj)
                         self.resultsfiles[res].write_function(v_out, indicator)
                     elif res=='acceleration': # passed in a is not a function but form, so we have to project
-                        self.a_proj = project(pb.acc, pb.V_u, pb.dx, domids=pb.domain_ids, nm="Acceleration", comm=self.comm) # class variable for testing
+                        self.a_proj = project(pb.acc, pb.V_u, self.dx, domids=pb.domain_ids, nm="Acceleration", comm=self.comm) # class variable for testing
                         a_out = fem.Function(self.V_out_vector, name=self.a_proj.name)
                         a_out.interpolate(self.a_proj)
                         self.resultsfiles[res].write_function(a_out, indicator)
@@ -410,7 +428,7 @@ class IO_solid(IO):
                         stressfuncs=[]
                         for n in range(pb.num_domains):
                             stressfuncs.append(pb.ma[n].sigma(pb.u,pb.p,pb.vel,ivar=pb.internalvars))
[…]
