+++ src/ambit_fe/coupling/fsi_main.py
@@ -15,6 +15,7 @@ from petsc4py import PETSc
 
 from ..solver import solver_nonlin
 from .. import ioparams, utilities
+from .. import boundaryconditions
 
 from ..solid.solid_main import SolidmechanicsProblem, SolidmechanicsSolverPrestr
 from .fluid_ale_main import FluidmechanicsAleProblem
@@ -27,20 +28,19 @@ FSI problem class
 
 class FSIProblem(problem_base):
 
-    def __init__(self, io_params, time_params_solid, time_params_fluid, fem_params_solid, fem_params_fluid, fem_params_ale, constitutive_models_solid, constitutive_models_fluid_ale, bc_dict_solid, bc_dict_fluid_ale, time_curves, coupling_params, io, ios, iof, mor_params={}, comm=None):
+    def __init__(self, io_params, time_params_solid, time_params_fluid, fem_params_solid, fem_params_fluid, fem_params_ale, constitutive_models_solid, constitutive_models_fluid_ale, bc_dict_solid, bc_dict_fluid_ale, bc_dict_lm, time_curves, coupling_params, io, ios, iof, mor_params={}, comm=None):
         super().__init__(io_params, time_params_solid, comm=comm)
 
         self.problem_physics = 'fsi'
 
         self.coupling_params = coupling_params
-        # self.coupling_surface = self.coupling_params['coupling_fluid_ale']['surface_ids']
 
         self.io = io
         self.ios, self.iof = ios, iof
 
-        # assert that we do not have conflicting timings
-        time_params_fluid['maxtime'] = time_params_solid['maxtime']
-        time_params_fluid['numstep'] = time_params_solid['numstep']
+        # assert that we do not have conflicting timings - TODO: Find better solution by moving these to global control parameters...
+        assert(time_params_fluid['maxtime'] == time_params_solid['maxtime'])
+        assert(time_params_fluid['numstep'] == time_params_solid['numstep'])
 
         # initialize problem instances (also sets the variational forms for the solid and fluid problem)
         self.pbs  = SolidmechanicsProblem(io_params, time_params_solid, fem_params_solid, constitutive_models_solid, bc_dict_solid, time_curves, ios, mor_params=mor_params, comm=self.comm)
@@ -60,6 +60,19 @@ class FSIProblem(problem_base):
 
         try: self.fsi_governing_type = self.coupling_params['fsi_governing_type']
         except: self.fsi_governing_type = 'solid_governed'
+
+        P_lm = ufl.VectorElement("CG", self.io.msh_emap_lm[0].ufl_cell(), self.pbs.order_disp)
+        self.V_lm = fem.FunctionSpace(self.io.msh_emap_lm[0], P_lm)
+
+        # Lagrange multiplier
+        self.LM = fem.Function(self.V_lm)
+        self.LM_old = fem.Function(self.V_lm)
+
+        self.dLM = ufl.TrialFunction(self.V_lm)    # incremental LM
+        self.var_LM = ufl.TestFunction(self.V_lm)  # LM test function
+
+        self.bclm = boundaryconditions.boundary_cond(self.io, dim=self.io.msh_emap_lm[0].topology.dim)
+        #self.bclm.dirichlet_bcs(bc_dict_lm['dirichlet'], self.V_lm)
 
         self.set_variational_forms()
 
[…]
