+++ src/ambit_fe/solver/projection.py
@@ -12,7 +12,7 @@ import ufl
 from petsc4py import PETSc
 
 
-def project(v, V, dx_, domids=np.arange(1,2), bcs=[], nm=None, comm=None, entity_maps=None):
+def project(v, V, dx_, domids=np.arange(1,2), bcs=[], nm=None, comm=None, entity_maps={}):
 
     w = ufl.TestFunction(V)
     Pv = ufl.TrialFunction(V)
@@ -38,7 +38,7 @@ def project(v, V, dx_, domids=np.arange(1,2), bcs=[], nm=None, comm=None, entity
     # solve linear system for projection
     function = fem.Function(V, name=nm)
 
-    if entity_maps is not None:
+    if bool(entity_maps):
         a_form, L_form = fem.form(a, entity_maps=entity_maps), fem.form(L, entity_maps=entity_maps)
     else:
         a_form, L_form = fem.form(a), fem.form(L)
