+++ joss/paper.bib
@@ -247,6 +247,17 @@ number={1},
 doi={10.1115/1.4005694}
 }
 
+@article{nordsletten2011,
+author={D. A. Nordsletten and M. McCormick and P. J. Kilner and P. Hunter and D. Kay and N. P. Smith},
+title={Fluid-solid coupling for the investigation of diastolic and systolic human left ventricular function},
+journal={International Journal for Numerical Methods in Biomedical Engineering},
+year={2011},
+pages={1017--1039},
+volume={27},
+number={7},
+doi={10.1002/cnm.1405}
+}
+
 @article{plank2021-opencarp,
 author={G. Plank and A. Loewe and A. Neic and C. Augustin and Y.-L. Huang and M. A. F. Gsell and E. Karabelas and M. Nothstein and A. J. Prassl and J. S{\'a}nchez and G. Seemann and E. J. Vigmond},
 title={The {openCARP} Simulation Environment for Cardiac Electrophysiology},
