+++ joss/paper.md
@@ -34,7 +34,7 @@ Furthermore, fluid dynamics in terms of incompressible Navier-Stokes/Stokes equa
 
 A variety of reduced 0D lumped models targeted at blood circulation modeling are implemented, including 3- and 4-element Windkessel models [@westerhof2009] as well as closed-loop full circulation [@hirschvogel2017] and coronary flow models [@arthurs2016].
 
-Monolithic multi-physics coupling of solid, fluid, and ALE-fluid with 0D lumped models is implemented such that cardiovascular simulations with realistic boundary conditions can be performed. Monolithic fluid-solid interaction (FSI) will be provided in the near future once mixed domain functionality is fully supported in the finite element backend dolfinx.
+Monolithic fluid-solid interaction (FSI) [@nordsletten2011] along with coupling of 3D and 0D models (solid or fluid with 0D lumped circulation systems) is implemented such that cardiovascular simulations with realistic boundary conditions can be performed.
 
 Implementations for a recently proposed novel physics- and projection-based model reduction for FSI, denoted as fluid-reduced-solid interaction (FrSI) [@hirschvogel2022preprint], are provided, along with POD-based Galerkin model reduction techniques [@farhat2014] using full or boundary subspaces.
 
@@ -52,7 +52,7 @@ Ambit represents a complete open-source code for simulating cardiac mechanics, e
 
 # Basic code structure
 
-Figure \autoref{fig:codedesign} represents a basic sketch of the main building blocks of Ambit. Depending on the physics of interest, the respective problem class is instantiated along with all the necessary input parameters, including boundary conditions (Dirichlet, Neumann, Robin), load curves, specification of coupling interfaces, etc. Single-physics problems like nonlinear elastodynamics (problem type `solid`) or fluid mechanics (problem type `fluid`) as well as 0D blood flow (problem type `flow0d`) can be solved as standalone problems. Additionally, 3D-0D coupling for 0D flow to 3D solid or fluid domains is supported (problem types `solid_flow0d` and `fluid_flow0d`), as well as fluid mechanics in ALE description (problem type `fluid_ale`), plus coupling to 0D models (problem type `fluid_ale_flow0d`). The implementations for FSI (problem type `fsi`) and FSI + 0D flow (problem type `fsi_flow0d`) are about to be finalized but currently are pending due to missing support in the finite element backend.
+Figure \autoref{fig:codedesign} represents a basic sketch of the main building blocks of Ambit. Depending on the physics of interest, the respective problem class is instantiated along with all the necessary input parameters, including boundary conditions (Dirichlet, Neumann, Robin), load curves, specification of coupling interfaces, etc. Single-physics problems like nonlinear elastodynamics (problem type `solid`) or fluid mechanics (problem type `fluid`) as well as 0D blood flow (problem type `flow0d`) can be solved as standalone problems. Additionally, FSI (problem type `fsi`) and 3D-0D coupling for 0D flow to 3D solid or fluid domains is supported (problem types `solid_flow0d` and `fluid_flow0d`), as well as fluid mechanics in ALE description (problem type `fluid_ale`), plus coupling to 0D models (problem types `fluid_ale_flow0d` and `fsi_flow0d`).
 
 The (coupled) problem object then is passed to a solver class, which calls the main routine to solve the nonlinear problem. This routine implements a time stepping scheme and a monolithic Newton solver which solves the (coupled multi-physics or single-field) problem and updates all variables simultaneously.
 
