+++ src/ambit_fe/coupling/fluid_flow0d_main.py
@@ -532,6 +532,12 @@ class FluidmechanicsFlow0DProblem(problem_base):
         self.pbf.write_output_ini()
 
 
+    def write_output_pre(self):
+
+        self.pbf.write_output_pre()
+        self.pb0.write_output_pre()
+
+
     def get_time_offset(self):
 
         return (self.pb0.ti.cycle[0]-1) * self.pb0.cardvasc0D.T_cycl * self.noperiodicref # zero if T_cycl variable is not specified
