+++ src/ambit_fe/coupling/fsi_flow0d_main.py
@@ -32,7 +32,7 @@ from ..meshutils import gather_surface_dof_indices
 
 class FSIFlow0DProblem(FSIProblem,problem_base):
 
-    def __init__(self, io_params, time_params_solid, time_params_fluid, time_params_flow0d, fem_params_solid, fem_params_fluid, fem_params_ale, constitutive_models_solid, constitutive_models_fluid_ale, model_params_flow0d, bc_dict_solid, bc_dict_fluid_ale, bc_dict_lm, time_curves, coupling_params_fluid_ale, coupling_params_fluid_flow0d, io, ios, iof, mor_params={}, comm=None, comm_sq=None):
+    def __init__(self, io_params, time_params_solid, time_params_fluid, time_params_flow0d, fem_params_solid, fem_params_fluid, fem_params_ale, constitutive_models_solid, constitutive_models_fluid_ale, model_params_flow0d, bc_dict_solid, bc_dict_fluid_ale, time_curves, coupling_params_fluid_ale, coupling_params_fluid_flow0d, io, ios, iof, mor_params={}, comm=None, comm_sq=None):
         problem_base.__init__(self, io_params, time_params_solid, comm=comm, comm_sq=comm_sq)
 
         ioparams.check_params_coupling_fluid_ale(coupling_params_fluid_ale)
@@ -60,6 +60,9 @@ class FSIFlow0DProblem(FSIProblem,problem_base):
 
         try: self.fsi_governing_type = self.coupling_params['fsi_governing_type']
         except: self.fsi_governing_type = 'solid_governed'
+
+        try: self.zero_lm_boundary = self.coupling_params['zero_lm_boundary']
+        except: self.zero_lm_boundary = False
 
         self.have_dbc_fluid_ale, self.have_weak_dirichlet_fluid_ale, self.have_dbc_ale_fluid, self.have_robin_ale_fluid = False, False, False, False
 
@@ -100,8 +103,11 @@ class FSIFlow0DProblem(FSIProblem,problem_base):
         self.var_LM = ufl.TestFunction(self.V_lm)  # LM test function
 
         self.bclm = boundaryconditions.boundary_cond(self.io, dim=self.io.msh_emap_lm[0].topology.dim)
-        # TODO: Application of DBCs to LM space not yet working, but might be needed when fluid and solid share mutual DBC nodes!
-        #self.bclm.dirichlet_bcs(bc_dict_lm['dirichlet'], self.V_lm)
+        # set the whole boundary of the LM subspace to zero (beneficial when we have solid and fluid with overlapping DBCs)
+        if self.zero_lm_boundary:
+            self.io.msh_emap_lm[0].topology.create_connectivity(self.io.msh_emap_lm[0].topology.dim-1, self.io.msh_emap_lm[0].topology.dim)
+            boundary_facets_lm = mesh.exterior_facet_indices(self.io.msh_emap_lm[0].topology)
+            self.bclm.dbcs.append( fem.dirichletbc(self.LM, boundary_facets_lm) )
 
         self.set_variational_forms()
 
@@ -157,6 +163,7 @@ class FSIFlow0DProblem(FSIProblem,problem_base):
 
         # even though this is zero, we still want to explicitly form and create the matrix for DBC application
         self.jac_ll = fem.form(self.weakform_lin_ll, entity_maps=self.io.entity_maps)
+        self.jac_ll_dummy = fem.form(self.from_ll_diag_dummy, entity_maps=self.io.entity_maps)
 
         te = time.time() - ts
         utilities.print_status("t = %.4f s" % (te), self.comm)
[…]
