+++ docs/source/ref.bib
@@ -156,6 +156,17 @@ doi={10.1007/978-3-642-23099-8},
 isbn={978-3-642-23098-1}
 }
 
+@article{nordsletten2011,
+author={D. A. Nordsletten and M. McCormick and P. J. Kilner and P. Hunter and D. Kay and N. P. Smith},
+title={Fluid-solid coupling for the investigation of diastolic and systolic human left ventricular function},
+journal={International Journal for Numerical Methods in Biomedical Engineering},
+year={2011},
+pages={1017--1039},
+volume={27},
+number={7},
+doi={10.1002/cnm.1405}
+}
+
 @article{schein2021,
 author={A. Schein and M. W. Gee},
 title={Greedy maximin distance sampling based model order reduction of prestressed and parametrized abdominal aortic aneurysms},

[…]
