+++ docs/source/ref.bib
@@ -189,6 +189,18 @@ number={3--4},
 doi={10.1016/S0045-7825(00)00211-5}
 }
 
+@incollection{turek2006,
+author={S. Turek and J. Hron},
+title={Proposal for Numerical Benchmarking of Fluid-Structure Interaction between an Elastic Object and Laminar Incompressible Flow}, 
+editor={H.-J.Bungartz and M. Sch{\"a}fer}, 
+booktitle={Lecture Notes in Computational Science and Engineering}, 
+year={2006},
+volume={53},
+pages={371--385},
+publisher={Springer Berlin Heidelberg},
+doi={10.1007/3-540-34596-5\_15}
+}
+
 @article{westerhof2009,
 author={N. Westerhof and J.-W. Lankhaar and B. E. Westerhof},
 title={The arterial {Windkessel}},
