+++ docs/source/ambit_doc.tex
@@ -1359,6 +1359,13 @@ Figure shows the velocity streamlines and magnitude at the end of the simulation
 
 \subsubsection*{Channel flow around elastic flag}
 
+Incompressible fluid flow in a 2D channel around an (incompressible) elastic flag is studied. The setup corresponds to the well-known Turek benchmark \cite{turek2006}, here with a sligh geometric modification (flag thickess is half of the original one). A prescribed inflow velocity is used, here without the parabolic profile proposed in the benchmark, but a constant one over the surface.\\
+
+Study the setup shown in fig. ... and the comments in the input file \verb"fsi_channel_flag.py". Run the simulation, either in one of the provided Docker containers or using your own FEniCSx/Ambit installation, using the command
+\begin{verbatim}
+mpiexec -n 1 python3 fsi_channel_flag.py
+\end{verbatim}
+
 \bibliography{ref}
 
 \end{document}
