+++ demos/fsi/fsi_channel_flag.py
@@ -28,7 +28,7 @@ def main():
                             'domain_ids_solid'      : [1], 
                             'domain_ids_fluid'      : [2],
                             'surface_ids_interface' : [1],
-                            'simname'               : 'tmp'}
+                            'simname'               : 'fsi_channel_flag'}
 
     """
     Parameters for the linear and nonlinear solution schemes
