### 2024-01-11 #5744 @marchirschvogel
Thank you all again for your reviews and the feedback. FYI, I've just done an update to enable monolithic FSI and added a demo and documentation. Accordingly, I've updated the paper.md file such that this additional functionality is reflected (which previously was stated as an outlook).
@Kevin-Mattheus-Moerman Is the review process finished or are there any further actions to be taken on my side? Thanks.
