### 2024-01-11 #5744 @Kevin-Mattheus-Moerman
@marchirschvogel indeed this is now finished as all reviewers have recommended this work for acceptance in JOSS. I will now help to process the final steps towards acceptance in JOSS.

I'd like to apologies for the delay on my side (due to the holidays and the start of the teaching semester).
