+++ src/ambit_fe/solid/solid_main.py
@@ -1059,7 +1059,8 @@ class SolidmechanicsProblem(problem_base):
 
 
     def destroy(self):
-        pass
+
+        self.io.close_output_files(self)
 
 
 
