### 2024-01-12 7bbfbb6
Change error handling for 3D-0D
