+++ docs/source/ambit_doc.tex
@@ -129,7 +129,7 @@
 
 \begin{document}
 
-\title{Ambit -- A FEniCS-based cardiovascular physics solver}
+\title{Ambit -- A FEniCS-based cardiovascular multi-physics solver}
 \author{Dr.-Ing. Marc Hirschvogel}
 
 \maketitle
