+++ pyproject.toml
@@ -3,8 +3,8 @@ requires = ["setuptools>=61.0.0", "wheel"]
 
 [project]
 name = "ambit-fe"
-version = "1.2.1"
-description = "A FEniCS-based cardiovascular physics solver"
+version = "1.2.2"
+description = "A FEniCS-based cardiovascular multi-physics solver"
 authors = [{name = "Marc Hirschvogel", email = "marc.hirschvogel@deepambit.com"}]
 license = {file = "LICENSE"}
 readme = "README.md"
