### 2024-01-12 0473572
More updates to error handling and output writing
