+++ src/ambit_fe/coupling/fsi_flow0d_main.py
@@ -19,11 +19,8 @@ from .. import utilities
 from .. import boundaryconditions
 from ..mpiroutines import allgather_vec
 
-from .fluid_ale_main import FluidmechanicsAleProblem
-from .fluid_flow0d_main import FluidmechanicsFlow0DProblem
-from ..fluid.fluid_main import FluidmechanicsSolverPrestr
-from ..solid.solid_main import SolidmechanicsProblem
 from .fsi_main import FSIProblem
+from ..solid.solid_main import SolidmechanicsProblem
 from .fluid_ale_flow0d_main import FluidmechanicsAleFlow0DProblem
 from ..ale.ale_main import AleProblem
 from ..base import problem_base, solver_base
@@ -45,18 +42,6 @@ class FSIFlow0DProblem(FSIProblem,problem_base):
         # assert that we do not have conflicting timings - TODO: Find better solution by moving these to global control parameters...
         assert(time_params_fluid['maxtime'] == time_params_solid['maxtime'])
         assert(time_params_fluid['numstep'] == time_params_solid['numstep'])
-
-        try: self.coupling_fluid_ale = coupling_params_fluid_ale['coupling_fluid_ale']
-        except: self.coupling_fluid_ale = {}
-
-        try: self.coupling_ale_fluid = coupling_params_fluid_ale['coupling_ale_fluid']
-        except: self.coupling_ale_fluid = {}
-
-        try: self.fluid_on_deformed = coupling_params_fluid_ale['fluid_on_deformed']
-        except: self.fluid_on_deformed = 'consistent'
-
-        try: self.coupling_strategy = coupling_params_fluid_ale['coupling_strategy']
-        except: self.coupling_strategy = 'monolithic'
 
         try: self.fsi_governing_type = self.coupling_params['fsi_governing_type']
         except: self.fsi_governing_type = 'solid_governed'
@@ -104,7 +89,7 @@ class FSIFlow0DProblem(FSIProblem,problem_base):
 
         self.bclm = boundaryconditions.boundary_cond(self.io, dim=self.io.msh_emap_lm[0].topology.dim)
         # set the whole boundary of the LM subspace to zero (beneficial when we have solid and fluid with overlapping DBCs)
-        if self.zero_lm_boundary:
+        if self.zero_lm_boundary: # TODO: Seems to not work properly!
             self.io.msh_emap_lm[0].topology.create_connectivity(self.io.msh_emap_lm[0].topology.dim-1, self.io.msh_emap_lm[0].topology.dim)
             boundary_facets_lm = mesh.exterior_facet_indices(self.io.msh_emap_lm[0].topology)
             self.bclm.dbcs.append( fem.dirichletbc(self.LM, boundary_facets_lm) )
@@ -292,52 +277,6 @@ class FSIFlow0DProblem(FSIProblem,problem_base):
         self.K_list[5+off][5+off] = self.pbfa0.K_list[3][3]
 
 
-    def get_index_sets(self, isoptions={}):
-
-        if self.rom is not None: # currently, ROM can only be on (subset of) first variable
-            vvec_or0 = self.rom.V.getOwnershipRangeColumn()[0]
-            vvec_ls = self.rom.V.getLocalSize()[1]
-        else:
-            vvec_or0 = self.pbf.v.vector.getOwnershipRange()[0]
-            vvec_ls = self.pbf.v.vector.getLocalSize()
-
-        offset_v = vvec_or0 + self.pbf.p.vector.getOwnershipRange()[0] + self.pbf0.lm.getOwnershipRange()[0] + self.pba.d.vector.getOwnershipRange()[0]
-        iset_v = PETSc.IS().createStride(vvec_ls, first=offset_v, step=1, comm=self.comm)
-
-        if isoptions['rom_to_new']:
[…]
