### 2024-01-12 #5744 @marchirschvogel
@Kevin-Mattheus-Moerman Thanks.
- I've double-checked and modified affiliations: I'm still affiliated with King's College London but will start a new position at Politecnico di Milano next week, so I've added both.
(I've also modified the paper title by adding a "multi-" infront of "physics")
- I've made a release, https://pypi.org/project/ambit-fe, with the version number being 1.2.2
- I've archived the current release version on Zenodo, cf. https://zenodo.org/records/10498075
- Double-checked that archive and JOSS paper author data match
- Double-checked that archive and JOSS paper license match
