### 2024-01-12 #5744 @Kevin-Mattheus-Moerman
@marchirschvogel sorry one more pedantic sounding thing. The version tag you used on your repo, `v1.2.2` should be matched for the archive which currently says `1.2.2`. We can solve this 1) by editing the archive version to also have the `v` to match (but perhaps then you'd also feel the need to edit that Python package listing?). Or 2) by creating a tagged release on your repository without the `v` (then I'll match the JOSS paper one here).
