### 2024-01-13 6c5873f
Some doc updates and fix for mass solve in FSI; added FSI-0D test
