+++ src/ambit_fe/postprocess/flow0d_plot.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-# Copyright (c) 2019-2023, Dr.-Ing. Marc Hirschvogel
+# Copyright (c) 2019-2024, Dr.-Ing. Marc Hirschvogel
 # All rights reserved.
 
 # This source code is licensed under the BSD-style license found in the
