+++ tests/runtests.py
@@ -87,6 +87,10 @@ if category=='fsi' or category=='all':
         errs['test_fsi_p1p1_stab_artseg 2'] = subprocess.call(['mpiexec', '-n', '2', 'python3', 'test_fsi_p1p1_stab_artseg.py'])
         errs['test_fsi_p1p1_stab_artseg 2 restart'] = subprocess.call(['mpiexec', '-n', '2', 'python3', 'test_fsi_p1p1_stab_artseg.py', str(4)])
 
+if category=='fsi_flow0d' or category=='all':
+    if branch=='mixed':
+        errs['test_fsi_flow0d_p1p1_stab_artseg 2'] = subprocess.call(['mpiexec', '-n', '2', 'python3', 'test_fsi_flow0d_p1p1_stab_artseg.py'])
+
 if category=='fluid_flow0d' or category=='all':
     errs['test_fluid_flow0d_monolagr_taylorhood_cylinder 1'] = subprocess.call(['mpiexec', '-n', '1', 'python3', 'test_fluid_flow0d_monolagr_taylorhood_cylinder.py'])
     errs['test_fluid_flow0d_monolagr_taylorhood_cylinder 3'] = subprocess.call(['mpiexec', '-n', '3', 'python3', 'test_fluid_flow0d_monolagr_taylorhood_cylinder.py'])
