+++ tests/test_fsi_flow0d_p1p1_stab_artseg.py
@@ -1,0 +1,161 @@
+#!/usr/bin/env python3
+
+"""
+FSI of arterial segment connected to resistive element: stabilized equal-order formulation, compressible solid
+"""
+
+import ambit_fe
+
+import sys, traceback
+import numpy as np
+from pathlib import Path
+import pytest
+
+
+@pytest.mark.fsi_flow0d
+@pytest.mark.fluid_solid_flow0d
+def test_main():
+
+    basepath = str(Path(__file__).parent.absolute())
+    
+    # reads in restart step from the command line
+    try: restart_step = int(sys.argv[1])
+    except: restart_step = 0
+
+    IO_PARAMS            = {'problem_type'          : 'fsi_flow0d',
+                            'USE_MIXED_DOLFINX_BRANCH' : True,
+                            'write_results_every'   : 1,
+                            'indicate_results_by'   : 'step',
+                            'write_restart_every'   : 1,
+                            'restart_step'          : restart_step,
+                            'restart_io_type'       : 'petscvector', # petscvector, rawtxt
+                            'output_path'           : basepath+'/tmp/',
+                            'mesh_domain'           : basepath+'/input/artseg-fsi-tet-lin_domain.xdmf',
+                            'mesh_boundary'         : basepath+'/input/artseg-fsi-tet-lin_boundary.xdmf',
+                            'results_to_write'      : [['displacement','velocity'], [['fluiddisplacement','velocity','pressure'],['aledisplacement','alevelocity']]],
+                            'domain_ids_solid'      : [1], 
+                            'domain_ids_fluid'      : [2],
+                            'surface_ids_interface' : [1],
+                            'simname'               : 'fsi_flow0d_p1p1_stab_artseg'}
+
+    SOLVER_PARAMS        = {'solve_type'            : 'direct',
+                            'direct_solver'         : 'mumps',
+                            'tol_res'               : [1e-8,1e-8,1e-8,1e-8,1e-8,1e-6],
+                            'tol_inc'               : [1e-0,1e-0,1e-0,1e-0,1e-0,1e-0],
+                            'subsolver_params'      : {'tol_res' : 1.0e-8, 'tol_inc' : 1.0e-8}}
+
+    TIME_PARAMS_SOLID    = {'maxtime'               : 3.0,
+                            'numstep'               : 150,
+                            'numstep_stop'          : 5,
+                            'timint'                : 'ost',
+                            'theta_ost'             : 1.0}
+
+    TIME_PARAMS_FLUID    = {'maxtime'               : 3.0,
+                            'numstep'               : 150,
+                            'numstep_stop'          : 5,
+                            'timint'                : 'ost',
+                            'theta_ost'             : 1.0}
+
+    TIME_PARAMS_FLOW0D   = {'timint'                : 'ost',
+                            'theta_ost'             : 1.0,
+                            'initial_conditions'    : {'Q_0' : 0.0, 'p_0' : 0.0}}
+
+    MODEL_PARAMS_FLOW0D  = {'modeltype'             : '2elwindkessel',
+                            'parameters'            : {'C' : 0., 'R' : 1e6, 'p_ref' : 0.}} # resistive blockage
+
[…]
