### 2024-01-13 #5744 @marchirschvogel
@Kevin-Mattheus-Moerman I've did one final version update with minor fixes. So now v1.2.3 is the final version, with the new Zenodo archive https://zenodo.org/records/10501275 and DOI 10.5281/zenodo.10501274.
I've now added the "v" infront of both the Github release and Zenodo archive version, since it seems to be common practice on Github.
