+++ joss/paper.bib
@@ -48,6 +48,17 @@ number={ANL-21/39 - Revision 3.17},
 year={2022},
 }
 
+@article{bergersen2020-turtlefsi,
+author={Aslak W. Bergersen and A. Slyngstad and S. Gjertsen and A. Souche and K. Valen-Sendstad},
+title={{turtleFSI}: A Robust and Monolithic FEniCS-based Fluid-Structure Interaction Solver},
+journal={The Journal of Open Source Software},
+year={2020},
+pages={2089},
+volume={5},
+number={50},
+doi={10.21105/joss.02089}
+}
+
 @article{bonini2022,
 author={M. Bonini and M. Hirschvogel and Y. Ahmed and H. Xu and A. Young and P. C. Tang and D. Nordsletten},
 title={Hemodynamic Modeling for Mitral Regurgitation},
