+++ joss/paper.md
@@ -27,10 +27,7 @@ bibliography: paper.bib
 
 # Summary
 
-Ambit is an open-source multi-physics finite element solver written in Python, supporting solid and fluid mechanics, fluid-structure interaction (FSI), and lumped-parameter models. It is tailored towards solving problems in cardiac mechanics, but may also be used for more general nonlinear finite element analysis. Ambit makes use of the open-source finite element library [FEniCS/dolfinx](https://fenicsproject.org) [@logg2012-fenics] along with the linear algebra package [PETSc](https://petsc.org) [@balay2022-petsc], hence guaranteeing a state-of-the-art finite element and linear algebra backend.
-The code encompasses re-implementations and generalizations of methods developed by the author for his PhD thesis [@hirschvogel2019disspub] and beyond. It is constantly updated to ensure compatibility with a recent dolfinx development version.
-
-Ambit is designed such that the user only needs to provide input files that define parameters through Python dictionaries, hence no programming or in-depth knowledge of any library-specific syntax is required.
+Ambit is an open-source multi-physics finite element solver written in Python, supporting solid and fluid mechanics, fluid-structure interaction (FSI), and lumped-parameter models. It is tailored towards solving problems in cardiac mechanics, but may also be used for more general nonlinear finite element analysis. The code encompasses re-implementations and generalizations of methods developed by the author for his PhD thesis [@hirschvogel2019disspub] and beyond. Ambit makes use of the open-source finite element library [FEniCS/dolfinx](https://fenicsproject.org) [@logg2012-fenics] along with the linear algebra package [PETSc](https://petsc.org) [@balay2022-petsc], hence guaranteeing a state-of-the-art finite element and linear algebra backend. It is constantly updated to ensure compatibility with a recent dolfinx development version. I/O routines are designed such that the user only needs to provide input files that define parameters through Python dictionaries, hence no programming or in-depth knowledge of any library-specific syntax is required.
 
 Ambit provides general nonlinear (compressible or incompressible) finite strain solid dynamics [@holzapfel2000], implementing a range of hyperelastic, viscous, and active material models. Specifically, the well-known anisotropic Holzapfel-Ogden [@holzapfel2009] and Guccione models [@guccione1995] for structural description of the myocardium are provided, along with a bunch of other models. It further implements strain- and stress-mediated volumetric growth models [@goektepe2010] that allow to model (maladaptive) ventricular shape and size changes. Inverse mechanics approaches to imprint loads into a reference state are implemented using the so-called prestressing method [@gee2010] in displacement formulation [@schein2021].
 
@@ -50,17 +47,17 @@ Avenues for future functionality include cardiac electrophysiology, scalar trans
 
 Cardiovascular disease entities are the most prevalent ones in the industrialized world [@dimmeler2011; @luepker2011] and a leading cause of death worldwide. Therefore, models that promote a better understanding of cardiac diseases and their progression represent a valuable tool to guide or assist therapy planning, support device dimensioning and design [@hirschvogel2019], or help predict intervention planning [@bonini2022; @taylor2013].
 
[…]
