+++ demos/fsi/README.md
@@ -1,0 +1,3 @@
+# README #
+
+Refer to example description at https://ambit.readthedocs.io/en/latest/documentation.html#fsi
