### 2024-01-16 #5744 @marchirschvogel
@Kevin-Mattheus-Moerman Awesome, thanks! Just a tiny thing: I wonder if the double-dash "--" in the title shown on the JOSS link can be replaced by a single dash like it appears in the PDF?
