### 2024-01-16 #5744 @Kevin-Mattheus-Moerman
@marchirschvogel we may be able to update that yeah. I suppose the large dash you used got seen as a double dash some somehow. It looks like our website doesn't support that larger dash type. Would you like to edit your paper title to just have a normal dash i.e. - instead of – ? Let me know if you do, then I'll re-accept to update the title and the meta-data.

@openjournals/dev What would you recommend in this case? The paper title and meta-data say:
`Ambit – A FEniCS-based cardiovascular multi-physics`
But the website renders it as:
`Ambit -- A FEniCS-based cardiovascular multi-physics solver`
Is this a general problem? Editing the title as I suggest :point_up: should not be a problem, correct?
