### 2024-01-16 #5744 @kyleniemeyer
@xuanxu @openjournals/dev is the above error due to using an en-dash in the paper title? That should be permitted, right?
