### 2024-01-19 #5744 @arfon
> @editorialbot generate pdf

At this point you must do a 'reaccept' command if you want to update the paper.
